Brown-Resnick-Specific {RandomFields}	R Documentation

Simulation methods for Brown-Resnick processes

Description

These models define particular ways to simulate Brown-Resnick processes.

Usage

RPbrmixed(phi, tcf, xi, mu, s, meshsize, vertnumber, optim_mixed,
          optim_mixed_tol,lambda, areamat, variobound) 

RPbrorig(phi, tcf, xi, mu, s)

RPbrshifted(phi, tcf, xi, mu, s)

RPloggaussnormed(variogram, prob, optimize_p, nth, burn.in, rejection)

Arguments

phi,variogram
object of class RMmodel; specifies the covariance model to be simulated.

tcf
the extremal correlation function; either phi or tcf must be given.

xi, mu, s
the shape parameter, the location parameter and the scale parameter, respectively, of the generalized extreme value distribution. See Details.

lambda
positive constant factor in the intensity of the Poisson point process used in the M3 representation, cf. Thm. 6 and Remark 7 in Oesting et. al (2012); can be estimated by setting optim_mixed if unknown. Default value is 1.

areamat
vector of values in [0,1]. The value of the kth component represents the portion of processes whose maximum is located at a distance d with k-1 <= d < k from the origin taken into account for the simulation of the shape function in the M3 representation. areamat can be used for isotropic models only; can be optimized by setting optim_mixed if unknown. Default value is 1.

meshsize, vertnumber, optim_mixed, optim_mixed_tol, variobound
further arguments for simulation via the mixed moving maxima (M3) representation; see RFoptions.

prob
to do

optimize_p
to do

nth
to do

burn.in
to do

rejection
to do

Details

The argument xi is always a number, i.e. ξ is constant in space. In contrast, μ and s might be constant numerical values or given an RMmodel, in particular by an RMtrend model.

The functions RPbrorig, RPbrshifted and RPbrmixed simulate a Brown-Resnick process, which is defined by

Z(x) = max_{i=1, 2, ...} X_i * exp(W_i(x) - gamma),

where the X_i are the points of a Poisson point process on the positive real half-axis with intensity 1/x^2 dx, W_i ~ Y are iid centered Gaussian processes with stationary increments and variogram gamma given by model. The functions correspond to the following ways of simulation:

RPbrorig

simulation using the original definition (method 0 in Oesting et al., 2012)

RPbrshifted

simulation using a random shift (similar to method 1 and 2)

RPbrmixed

simulation using M3 representation (method 4)

Value

The functions return an object of class RMmodel.

Note

Advanced options for RPbroriginal and RPbrshifted are maxpoints and max_gauss, see RFoptions.

Author(s)

Marco Oesting, oesting@mathematik.uni-siegen.de, https://www.uni-siegen.de/fb6/src/scheffler/mitarbeiter/oesting; Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Oesting, M., Kabluchko, Z. and Schlather M. (2012) Simulation of Brown-Resnick Processes, Extremes, 15, 89-107.

See Also

RPbrownresnick, RMmodel, RPgauss, maxstable, maxstableAdvanced.

Examples

#
RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## currently does not work

[Package RandomFields version 3.3.6 Index]