Hierarchical Modelling {RandomFields}	R Documentation

Bayesian Spatial Modelling

Description

RandomFields provides Bayesian modelling to some extend: (i) simulation of hierarchical models at arbitrary depth; (ii) estimation of the parameters of a hierarchical model of depth 1 by means of maximizing the likelihood.

Details

A Bayesian approach can be taken for scalar, real valued model parameters, e.g. the shape parameter nu in the RMmatern model. A random parameter can be passed through a distribution of an existing family, e.g. (dnorm, pnorm, qnorm, rnorm) or self-defined. It is passed without the leading letter d, p, q, r, but as a function call e.g norm(). This function call may contain arguments that must be named, e.g. norm(mean=3, sd=5).

Usage:

exp() denotes the exponential distribution family with rate 1,

exp(3) is just the scalar e^3 and

exp(rate=3) is the exponential distribution family with rate 3.

The family can be passed in three ways:

implicitly, e.g. RMwhittle(nu=exp()) or

explicitly through RRdistr, e.g. RMwhittle(nu=RRdistr(exp())).

by use of RRmodels of the package.

The first is more convenient, the second more flexible and slightly safer.

Note

While simulating any depth of hierarchical modelling is possible, estimation is currently restricted to one level of hierarchy.

The effect of the distribution family varies between the different processes:

in max-stable fields and RPpoisson, a new realization of the prior distribution(s) is drawn for each shape function

in all other cases: a realization of the prior(s) is only drawn once. This effects, in particular, Gaussian fields with argument n>1, where all realizations are based on the same realization out of the prior distribution(s).

Note that checking the validity of the arguments is rather limited for such complicated models, in general.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmodelsAdvanced. For hierarchical modelling see RR.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
## See 'RRmodels' for hierarchical models

## the following model defines the argument nu of the Whittle-Matern
## model to be an exponential random variable with rate 5.
model <- ~ 1 + RMwhittle(scale=NA, var=NA, nu=exp(rate=5)) + RMnugget(var=NA)

data(soil)
fit <- RFfit(model, x=soil$x, y=soil$y, data=soil$moisture, modus="careless")
print(fit)

[Package RandomFields version 3.3.6 Index]