Changings {RandomFields}	R Documentation

Documentation of some further changings

Description

Version 3.3

RFempiricalvariogram, RFempiricalcovariance and RFempiricalmadogram became obsolete. Use RFvariogram, RFcov, RFmadogram instead.

RFoptions(grDefault=FALSE) returns to the old style of graphical device handling. Otherwise there is no handling.

C code is started to be parallelized.

Some new Multivariate RMmodels

New way of passing models, see RFformula, which allows connections (formulae) between parameters, e.g. one parameter value might be twice as large as another parameter value. Also dummy variables can be RMdeclared.

Options getting obsolete (Version 3 and older)

oldstyle is becoming warn_oldstyle

newstyle is becoming warn_newstyle

newAniso is becoming warn_newAniso

ambiguous is becoming warn_ambiguous

normal_mode is becoming warn_normal_mode

colour_palette is becoming warn_colour_palette

Changings in option names

several changes in RFoptions()$graphics in version 3.1.11

pdfnumber became in version 3.0.42 filenumber

pdfonefile became in version 3.0.42 onefile

pdffile became in version 3.0.42 file

tbmdim became in version 3.0.41 reduceddim

coord_units became in version 3.0.39 coordunits

new_coord_units became in version 3.0.39 new_coordunits

variab_units became in version 3.0.39 varunits

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

MajorRevisions, RandomFields.

Examples

## no examples given

[Package RandomFields version 3.3.6 Index]