GetPracticalRange {RandomFields}	R Documentation

Determination of the practical range

Description

the function returns the practical range of a covariance model, i.e. the distance for which the model with standard parameters (variance=1, nugget=0, scale=1) values 0.05

Usage

GetPracticalRange(model, param, dim=1)

Arguments

model
covariance model in the new list not; the model must be primitive, i.e. may not call other models. See Covariance and CovarianceFct for details.

param
optional vector of parameters for the model, see CovarianceFct

dim
space-time dimension; except for some few anisotropic models, this parameter does not have an effect

Details

In case the practical range has to be determined numerically, only a rough approximation is provided.

The function can only be applied to isotropic models

Value

real number (the practical range)

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de

References

Goovaerts, P. (1997) Geostatistics for Natural Resources Evaluation. New York: Oxford University Press.

See Also

CovarianceFct, RandomFields, RFparameters,

Examples

GetPracticalRange(list("exponential"))
GetPracticalRange(list("whittle", nu=2))
try(GetPracticalRange(list("bessel", nu=-0.5))) # leads to an error

[Package RandomFields version 2.0.66 Index]