GetRegisterInfo {RandomFields}	R Documentation

Internal information

Description

The function returns internal information about the simulation of a random field and the calculation of covariance functions

Usage

GetRegisterInfo(register, ignore.active = FALSE, max.elements=10^6,
                meth=NULL, modelname=NULL)
GetModelInfo(register, modelreg, level=3, gatter=FALSE)
GetModel(register, modelreg, modus=0)

Arguments

register
-2, -1, 0:9; (-1 and -2 only work with GetModelInfo) place where intermediate calculations are stored; the numbers 0:9 are aliases for 10 internal registers, see also GaussRF

modelreg
value in 0..4 with the following meaning:

0 : register for the functions Covariance, CovarianceFct, ect

1 : register for simulation

2 : register for covariance calculation within Kriging

3 : register for MLE (likelihood)

4 : register for MLE (bounds)

Positive values refer to the registers, see GaussRF.

ignore.active
logical. If FALSE and the register has non-active flag (because of an error or it is deleted) then a list is returned that contains only the element active=FALSE. Otherwise, a more complete list is returned – only for internal debugging purposes. Warning: ignore.active=TRUE may cause failure of R.

max.elements
integer; since GetRegisterInfo might be a dump of simulation that needs a huge amount of memory and since all entries are copied, the maximal amount of available memory might be easily exceeded. Therefore, only the size of the critical parts are returned and not the vector or matrix itself, if the number of elements exceeds max.elements.

meth
vector of strings. If meth is given GetRegisterInfo returns not the whole register info but only part method part.

modelname
string. If modelname is given then GetRegisterInfo returns the first appearance of the covariance model with name modelname. If meth is given then the model within the method is returned.

level
integer; level of details

modus
integer; see details

gatter
boolean; only for very advanced interests.

Details

GetRegisterInfo(register, ignore.active=TRUE) is useful for debugging and specialists' need to control the algorithm, see the examples in RFparameters and GaussRF.

If RFparameters()$Storage=FALSE then values of the internal registers are not kept if GaussRF or DoSimulateRF has been called. Hence GetRegisterInfo cannot provide any information.

GetModelInfo returns the complete information on the internal model structure. It allows for register=-1 returning the model structure for the last use of CovarianceFct, sophisticated models or Variogram or similar commands. register=-2 is for internal use only.

GetModel returns only the parts of the internal model structure that have been defined by the user. The modus is ignored if PracticalRange=FALSE in RFparameters. In case PracticalRange=TRUE the modus has the following effects

0the model is returned without the explicite scale transformation

1the model is returned the way it is stored, including the scale transformation

2the scale transformation of the the practical range is included, but a simplified model will be returned

Value

List of internal information is returned.

Note

Put Storing=TRUE, see RFparameters if you like to have more internal information in case of an expected failure of an initialisation of a random field simulation.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de

See Also

GaussRF

Examples

GaussRF(1:4, grid=TRUE, model="exp", param=c(1,2,3,4), Storing=TRUE)
Print(GetRegisterInfo(0))

# finally, a register that could not be successfully initialised
try(GaussRF(runif(4), grid=FALSE, model="exp", param=c(1,2,3,4),
            me="ci", Storing=TRUE))
if (interactive()) Print(GetRegisterInfo(0, TRUE))

[Package RandomFields version 2.0.71 Index]