Locator {RandomFields}	R Documentation

Graphical Input

Description

Reads the position of the graphics cursor when the (first) mouse button is pressed, or replays from storage.

Usage

Locator(n, type="n", info=NULL, ...)

Arguments

n
the maximum number of points to locate; the value of n is ignored if

type
One of '"n"', '"p"', '"l"' or '"o"'. If '"p"' or '"o"' the points are plotted; if '"l"' or '"o"' they are joined by lines.

info
arbitrary object; tracing information that is useful in case the user likes to edit the stored sequence of inputs; the value of info is considered only if useraction has been called by action="start.register" or action="continue.register"

...
additional graphics parameters

Details

The behaviour of Locator depends on the the value of action set by useraction, see there for more information.

See also locator for further information on the parameters.

Value

A list containing 'x' and 'y' components which are the coordinates of the identified points in the user coordinate system, i.e., the one specified by 'par("usr")'.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de

See Also

getactions, Readline, useraction, userinput

Examples

## see useraction

[Package RandomFields version 2.0.71 Index]