Major Revisions {RandomFields}	R Documentation

Documentation of major changings

Description

This man page documents some major changings in RandomFields.

Changes done in 3.1.0 (Summer 2015)

full (univariate) trend modelling

error in particular in RFfit corrected

RFfit runs much faster now

effects of modus operandi changed for estimating

Corrections done in 3.0.56 (Jan 2015)

log Gauss field corrected (has been a log log Gauss field)

RMconstant is now called RMfixcov

Corrections done in 3.0.55 (Jan 2015)

Conditional simulation: several severe typos corrected.

Major Revision: changings from Version 2 to Version 3 (Jan 2014)

S4 objects

RandomFields is now based on S4 objects using the package sp. The functions accept both sp objects and simple objects as used in version 2. See also above.

Documentation

each model has now its own man page;

classes of models and functions are bundled in several pages: Covariance models start with RM, distribution families with RR, processes with RP, user functions with RF

the man pages of several functions are split into two parts:

(i) a beginners man page which includes a link to

(ii) man pages for advanced users

Interfaces

The interfaces become simpler, at the same time more powerful than the functions in version 2. E.g., RFsimulate can perform unconditional simulation, conditional simulation and random imputing.

Only those arguments are kept in the functions that are considered as being absolutely necessary. All the other arguments can be included as options.

RFgui is an instructive interface based on tcl/tk, replacing the former ShowModels

Inference for Gaussian random fields

RFfit has undergone a major revision. E.g.:

(i) estimation of random effect models with spatial covariance structure

(ii) automatic estimation of 10 and more arguments in multivariate and/or space-time models

RFvariogram is now based on an fft algorithm if the data are on a grid, even allowing for missing values.

RFratiotest has been added.

Processes

Modelling of maxstable processes has been enhanced, including

(i) the simulation of Brown-Resnick processes

(ii) initial support of tail correlation functions;

Further processes chi2 processes, compound Poisson processes, binary processes added.

Models

the formula notation for linear models may now be defined

Novel, user friendly definition of the covariance models

Multivariate and vector-valued random fields are now fully included

The user may now define his own functions, to some extend.

The trend allows for much more flexibility

Distributions may now be included which will be extended to Bayesian modelling in future.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## S4 vs S3
x <- seq(0, 10, 0.1)
model <- RMexp()
plot(RFsimulate(model, x)) ## S4
plot(RFsimulate(model, x, spConform=FALSE)) ## no class

[Package RandomFields version 3.3.6 Index]