PrintModelList {RandomFields}	R Documentation

Information about the implemented covariance models

Description

PrintModelList prints the list of currently implemented models including the corresponding simulation methods

GetModelList returns a matrix of currently implemented models and their simulation methods

GetModelNames returns a list of currently implemented models

Usage

PrintModelList(operators=FALSE, internal=FALSE)

GetModelList(abbr=TRUE, internal=FALSE)

GetModelNames(stationary = c("any", "stationary", "variogram", "irf",
                             "auxmatrix", "auxvector", "gencovariance",
                             "nonstationary", "genvariogram", "prev.model"),
              isotropy = c("any", "isotropic", "spaceisotropic",
                           "zerospaceiso", "anisotropic", "prev.model"),
              multivariate = -9999:9999,
              operator, normalmix, finiterange, methods, internal=FALSE,
              dim = 1)

Arguments

operators
logical. Flag whether operators should be also considered.

internal
logical. Flag whether internal models should be also considered. In case of PrintModelList and internal=2, variants of internal models are also printed.

abbr
logical or numerical. If TRUE the names for the methods are abbreviated. If numerical, abbr gives the number of letters.

stationary
Selection criterion for the models

isotropy
Selection criterion for the models

multivariate
the dimension of the multivariate model

operator
logical. If missing any covariance model is returned; otherwise, the (non)primitive models are returned if FALSE (TRUE); see Covariance for details.

normalmix
logical. If missing any covariance model is returned; otherwise all normal scale mixture models are returned if TRUE

finiterange
logical. If missing any covariance model is returned; otherwise all models with finite range are returned if TRUE; note that some option are excluding, e.g. no normal scale model can have a finite range.

methods
If missing any covariance model is returned; otherwise the models with respective attribute are returned; see PrintMethodList for a complete list of methods

dim
dimension in which the model should be valid; there are some models where the validity in a given dimension depends on the some parameter; those models are returned partially only.

Details

See CovarianceFct and sophisticated models for a description of the models and their use

Value

PrintModelList prints a table of the currently implemented covariance functions and the matching methods. PrintModelList returns NULL.

GetModelNames returns a list of implemented models

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de

See Also

sophisticated, CovarianceFct, GetModel

Examples

PrintModelList()
GetModelList(abbr=TRUE)
GetModelNames("stationary", "spaceisotropic")

[Package RandomFields version 2.0.66 Index]