Constants {RandomFields}	R Documentation

Constants used in RandomFields (RC constants)

Description

Several constants are provided that might make the use of some functions easier, e.g. RFgetModelNames.

Value

RC_TYPE_NAMES = c( "tail correlation", "positive definite", "variogram", "negative definite", "point-shape function", "shape function", "trend", "distribution or shape", "of manifold type", "process", "method for Gauss process", "normed process (non-negative values with maximum value being 0 or 1)", "method for Brown-Resnick process", "Smith", "Schlather", "Poisson", "PoissonGauss", "distribution family", "interface", "mathematical operator", "other type")

RC_DOMAIN_NAMES = c("single variable", "kernel", "framework dependent", "submodel dependent", "parameter dependent", "<keep copy>", "mismatch")

RC_ISO_NAMES = c("isotropic", "space-isotropic", "vector-isotropic", "symmetric", "cartesian system", "gnomonic projection", "orthographic projection", "spherical isotropic", "spherical symmetric", "spherical system", "earth isotropic", "earth symmetric", "earth system", "cylinder system", "non-dimension-reducing", "framework dependent", "submodel dependent", "parameter dependent", "<internal keep copy>", "<mismatch>")

RC_MONOTONE_NAMES = c( "not set", "mismatch in monotonicity", "submodel dependent monotonicity", "previous model dependent monotonicity", "parameter dependent monotonicity", "not monotone", "monotone", "Gneiting-Schaback class", "normal mixture", "completely monotone", "Bernstein")

RC_ISOTROPIC gives the numerical code for option "isotropic"

RC_DOUBLEISOTROPIC gives the numerical code for option "space-isotropic"

RC_CARTESIAN_COORD gives the numerical code for option "cartesian system"

RC_GNOMONIC_PROJ gives the numerical code for the gnomonic projection, see also zenit in RFoptions.

RC_ORTHOGRAPHIC_PROJ gives the numerical code for the orthographic projection, see also zenit in RFoptions.

RC_EARTH_COORDS gives the numerical code for option "earth coordinates"

RC_EARTH_ISOTROPIC gives the numerical code for option "earth isotropic"

RC_SPHERICAL_COORDS gives the numerical code for option "earth coordinates"

RC_OPTIMISER_NAMES and RC_NLOPTR_NAMES give the names for the options optimiser and algorithm, respectively, RFfitoptimiser.

RC_LIKELIHOOD_NAMES = c("auto", "full", "composite", "tesselation") gives the names of the ML variants: (i) internal choice according to the number of data, (ii) full likelihood, (iii) (pairwise) composite likelihood, and (iv) composite likelihood based on a tessellation of the space.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RF, RM, RP, RR, R., RFgetModelNames, RMmodelgenerator-class, RMtrafo.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

RC_ISO_NAMES
RC_ISO_NAMES[RC_ISOTROPIC:RC_CARTESIAN_COORD + 1]
## Not run: 
RFgetModelNames(isotropy=RC_ISO_NAMES[RC_ISOTROPIC:RC_CARTESIAN_COORD +
1])

## End(Not run)

[Package RandomFields version 3.3.6 Index]