RFfunction {RandomFields}	R Documentation

Evaluation operators (RF commands)

Description

Here, all the RF_name_ commands are listed.

Functionals of RMmodels

The user's RMmodel is supplemented internally by operators that are tacitely assumed, e.g. RPgauss.

Further completions of the user's model determine what should be done with the model, e.g. calculation of the covariance (RFcov). The following list gives those RFfunctions that have an internal representation as completion to the user's model.

RFcalc 	 performs some simple calculations based on R.models

RFcov 	 assigns to a covariance model the covariance values at given locations

RFcovmatrix 	 assigns to a covariance model the matrix of covariance values at given locations

RFdistr 	 generic function assigning to a distribution family various values of the distribution

RFfctn 	 assigns to a model the value of the function at given locations. In case of a covariance model RFfctn is identical to RFcov.

RFlikelihood 	 assigns to a model and a dataset the (log)likelihood value.

RFlinearpart 	 assigns to a model and a set of coordinates the linear part of the model, i.e. the deterministic trend and the design matrix.

RFpseudovariogram 	 assigns to a model the values of the pseudo variogram at given locations

RFsimulate 	 assigns to a model a realisation of the corresponding random field

RFvariogram 	 assigns to a model the values of the (cross-)variogram at given locations

Estimation and Inference

RFcrossvalidate 	 cross validation for Gaussian fields

RFvariogram 	 empirical variogram

RFfit 	 (maximum likelihood) fitting of the parameters

RFinterpolate 	 'kriging' and 'imputing'

RFratiotest 	 likelihood ratio test for Gaussian fields

Graphics for Gaussian fields

RFgui 	 educational tool for

* manual selection of a covariance model

* manual fitting to the empirical variogram

RFfractaldim 	 determination of the fractal dimension

RFhurst 	 determination of the Hurst effect (long range dependence)

Coordinate transformations

RFearth2cartesian 	 transformation of earth coordinates to cartesian coordinates

RFearth2dist 	 transformation of earth coordinates to Euclidean distances

Information from and to RandomFields

RFgetMethodNames 	 currently implemented list of simulation methods

RFgetModel 	 returns the model used in a RFfunction, with some more details

RFgetModelInfo 	 similar to RFgetModel, but with detailed information on the implementation

RFgetModelNames 	 lists the implemented models

RFoptions 	 options of package RandomFields

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RC, RM, RP, RR, R., RMmodelgenerator

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

z <- RFsimulate(model=RMexp(), 1:10)
RFgetModel(RFsimulate, show.call = TRUE)  # user's definition
RFgetModel(RFsimulate, show.call = FALSE) # main internal part

[Package RandomFields version 3.3.6 Index]