RFempVariog-class {RandomFields}	R Documentation

Class RFempVariog

Description

Class for RandomFields' representation of empirical variograms

Usage

RFplotEmpVariogram(x, model = NULL, nmax.phi = NA, nmax.theta = NA,
                    nmax.T = NA,
                     plot.nbin = TRUE, plot.sd=FALSE, method = "ml",
                     variogram=TRUE,
                               boundaries = TRUE,
                               ...)
## S4 method for signature 'RFempVariog,missing'
plot(x, y, ...)
## S4 method for signature 'RFempVariog'
persp(x, ...)

Arguments

x
object of class RFempVariog

y
unused

model
object of class RMmodel, RFformula or formula; best is to consider the examples below, first.
The argument params is a list that specifies free parameters in a formula description, see RMformula.. Or a list of such models. Tit gives the covariance or variogram models that are to be plotted into the same plot as the empirical variogram (and the fitted models)

nmax.phi
even integer; only for class(x)=="RFempVariog"; the number of bins of angle phi that are to be plotted

.

nmax.theta
integer; only for class(x)=="RFempVariog"; the number of bins of angle theta that are to be plotted

nmax.T
integer; only for class(x)=="RFempVariog"; the maximal number of different time bins that are to be plotted

plot.nbin
logical; only for class(x)=="RFempVariog"; indicates whether the number of pairs per bin are to be plotted

plot.sd
logical; only for class(x)=="RFempVariog"; indicates whether the calculated standard deviation (x@sd) is to be plotted (in form of arrows of length +-1*sd)

method
character. Currently restricted to "ml" for maximum-likelihood method.

variogram
logical; This argument should currently not be set by the user. If TRUE then the empirical variogram is plotted, else an estimate for the covariance function.

boundaries
logical; only for class(x)=="RFempVariog" and the anisotropic case where model is given. As the empirical variogram is calculated on a sector of angles, no exact variogram curve corresponds to the mean values in this sector. If boundaries=TRUE the values of the variogram on the sector boundaries are plotted. If FALSE some kind of mean model values are plotted. Neither the boundaries may contain the values of empirical variogram nor does the mean values need to be close the empirical variogram.

...
arguments to be passed to methods; mainly graphical arguments.

Slots

centers:

the bin centres of the spatial distances

empirical:

value of the empirical variogram

var:

the empirical (overall) variance in the data

sd:

standard deviation of the variogram cloud within each bin

n.bin:

number of bins

phi.centers:

centres of the bins with respect to the (first) angle (for anisotropic empirical variograms only)

theta.centers:

centres of the bins with respect to the second angle (for anisotropic empirical variograms in 3D only)

T:

the bin centres of the time axis

vdim:

the multivariate dimension

coordunits:

string giving the units of the coordinates, see also option coordunits of RFoptions.

varunits:

string giving the units of the variables, see also option varunits of RFoptions.

call:

language object; the function call by which the object was generated

method:

integer; variogram (0), covariance (2), madogram (4)

Methods

plot

signature(x = "RFempVariog"): gives a plot of the empirical variogram, for more details see plot-method.

plot

signature(x = "RFempVariog", y = "missing")

Gives nice plots of the empirical variogram; handles binning in up to three space-dimensions and a time-dimension, where the empirical variogram is plotted along lines which are directed according to the angle-centers given in x@phi.centers and x@theta.centers; arbitrary theoretical model curves can be added to the plot by using the argument model. If no bins are given, i.e. (x@bin=NULL), image-plots are generated.

as

signature(x = "RFempVariog"): converts into other formats, only implemented for target class list.

show

signature(x = "RFfit"): returns the structure of x

persp

signature(obj = "RFempVariog"): generates nice persp plots

print

signature(x = "RFfit"): identical with show-method

summary

provides a summary

Details

print returns also an invisible list that is convenient to access.

Author(s)

Alexander Malinowski, Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RFvariogram, plot-method

Examples

#  see 'RFvariogram'

[Package RandomFields version 3.3.6 Index]