RFfit-class {RandomFields}	R Documentation

Class RFfit

Description

Class for RandomFields' representation of model estimation results

Usage

## S4 method for signature 'RFfit'
residuals(object, ..., method="ml", full=FALSE)
## S4 method for signature 'RFfit'
summary(object, ..., method="ml")
## S4 method for signature 'RFfit,missing'
plot(x, y, ...) 

## S3 method for class 'RFfit'
contour(x, ...) 
## S3 method for class 'RFempVariog'
contour(x, ...)

RFhessian(model)

Arguments

object
see the generic function;

...

plot: arguments to be passed to methods; mainly graphical arguments, or further models in case of class CLASS_CLIST, see Details.

summary: see the generic function

contour : see RFplotEmpVariogram

method
character; only for class(x)=="RFfit"; a vector of slot names for which the fitted covariance or variogram model is to be plotted; should be a subset of slotNames(x) for which the corresponding slots are of class CLASS_FIT; by default, the maximum likelihood fit ("ml") will be plotted

full
logical. if TRUE submodels are reported as well (if available).

x
object of class RFsp or RFempVariog or RFfit or RMmodel; in the latter case, x can be any sophisticated model but it must be either stationary or a variogram model

y
unused

model
class(x)=="RF_fit" or class(x)=="RFfit", obtained from RFfit

Details

for the definition of plot see RFplotEmpVariogram.

Creating Objects

Objects are created by the function RFfit

Slots

autostart:

RMmodelFit; contains the estimation results for the method 'autostart' including a likelihood value, a constant trend and the residuals

boxcox:

logical; whether the parameter of a Box Cox tranformation has been estimated

coordunits:

string giving the units of the coordinates, see also option coordunits of RFoptions.

deleted:

integer vector. Positions of the parameters that have been deleted to get the set of variables, used in the optimization.

ev:

list; list of objects of class RFempVariog, contains the empirical variogram estimates of the data

fixed:

list of two vectors. The fist gives the position where the parameters are set to zero. The second gives the position where the parameters are set to one.

internal1:

RMmodelFit; analog to slot 'autostart'

internal2:

RMmodelFit; analog to slot 'autostart'

internal3:

RMmodelFit; analog to slot 'autostart'

lowerbounds:

RMmodel; covariance model in which each parameter value gives the lower bound for the respective parameter

ml:

RMmodelFit; analog to slot 'autostart'

modelinfo:

table with information on the parameters: name, boundaries, type of parameter

n.covariates:

number of covariates

n.param:

number of parameters (given by the user)

n.variab:

number of variables (used internally); n.variab is always less than or equal to n.param

number.of.data:

the number of data values passed to RFfit that are not NA or NaN

number.of.parameters:

total number of parameters of the model that had to be estimated including variances, scales, co-variables, etc.

p.proj:

vector of integers. The original position of those parameters that are used in the submodel

plain:

RMmodelFit; analog to slot 'autostart'

report:

If not empty, it indicates that this model should be reported and gives a standard name of the model.

Various functions, e.g. print.RMmodelFit, use this information if their argument full equals TRUE.

self:

RMmodelFit; analog to slot 'autostart'

sd.inv:

RMmodelFit; analog to slot 'autostart'

sqrt.nr:

RMmodelFit; analog to slot 'autostart'

submodels:

list. Sequence (in some cases even nested sequence) of models that is used to determine an initial value in

table:

matrix; summary of estimation results of different methods

transform:

function;

true.tsdim:

time space dimension of the (original!) data, even for submodels that consider parts of separable models.

true.vdim:

multivariability of the (original!) data, even for submodels that consider independent models for the multivariate components.

upperbounds:

RMmodel; see slot 'lowerbounds'

users.guess:

RMmodelFit; analog to slot 'autostart'

ml:

RMmodelFit; analog to slot 'autostart'; with maximum likelihood method

v.proj:

vector of integers. The components selected in one of the submodels

varunits:

string giving the units of the variables, see also option varunits of RFoptions.

x.proj:

logical or integer. If logical, it means that no separable model is considered there. If integer, then it gives the considered directions of a separable model.

Z:

standardized list of information on the data

Methods

plot

signature(x = "RFfit"): gives a plot of the empirical variogram together with the fitted model, for more details see plot-method.

show

signature(x = "RFfit"): returns the structure of x

persp

signature(obj = "RFfit"): generates persp plots

print

signature(x = "RFfit"): identical with show-method, additional argument is max.level

[

signature(x = "RFfit"): enables accessing the slots via the "["-operator, e.g. x["ml"]

as

signature(x = "RFfit"): converts into other formats, only implemented for target class RFempVariog

anova

performs a likelihood ratio test base on a chisq approximation

summary

provides a summary

logLik

provides an object of class "logLik"

AIC,BIC

provides the AIC and BIC information, respectively

signature(x = "RFfit", y = "missing")

Combines the plot of the empirical variogram with the estimated covariance or variogram model (theoretical) curves; further models can be added via the argument model.

Further 'methods'

AICc.RFfit(object, ..., method="ml", full=FALSE)

AICc.RF_fit(object, ..., method="ml", full=TRUE)

Author(s)

Alexander Malinowski; Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

AICc:

Hurvich, C.M. and Tsai, C.-L. (1989) Regression and Time Series Model Selection in Small Samples Biometrika, 76, 297-307.

See Also

RFfit, RFvariogram, RMmodel-class, RMmodelFit-class, plot-method.

Examples

# see RFfit

[Package RandomFields version 3.3.6 Index]