RFgetMethodNames {RandomFields}	R Documentation

Simulation Techniques

Description

RFgetMethodNames prints and returns a list of currently implemented methods for simulating Gaussian random fields and max stable random fields

Usage

RFgetMethodNames()

Details

By default, RFsimulate automatically chooses an appropriate method for simulation. The method can also be set explicitly by the user via RFoptions, in particular by passing gauss.method=_a valid method string_ as an additional argument to RFsimulate or by globally changing the options via RFoptions(gauss.method=_a valid method string_). The following methods are available:

(random spatial) Averages
– details soon

Boolean functions.
See marked point processes.

circulant embedding.
Introduced by Dietrich & Newsam (1993) and Wood and Chan (1994).

Circulant embedding is a fast simulation method based on Fourier transformations. It is garantueed to be an exact method for covariance functions with finite support, e.g. the spherical model.

See also cutoff embedding and intrinsic embedding for variants of the method.

cutoff embedding.
Modified circulant embedding method so that exact simulation is guaranteed for further covariance models, e.g. the whittle matern model. In fact, the circulant embedding is called with the cutoff hypermodel, see RMmodel, and A=B there. cutoff embedding halfens the maximum number of elements models used to define the covariance function of interest (from 10 to 5).

Here, multiplicative models are not allowed (yet).

direct matrix decomposition.
This method is based on the well-known method for simulating any multivariate Gaussian distribution, using the square root of the covariance matrix. The method is pretty slow and limited to about 8000 points, i.e. a 20x20x20 grid in three dimensions. This implementation can use the Cholesky decomposition and the singular value decomposition. It allows for arbitrary points and arbitrary grids.

hyperplane method.
The method is based on a tessellation of the space by hyperplanes. Each cell takes a spatially constant value of an i.i.d. random variables. The superposition of several such random fields yields approximatively a Gaussian random field.

intrinsic embedding.
Modified circulant embedding so that exact simulation is guaranteed for further variogram models, e.g. the fractal brownian one. Note that the simulated random field is always non-stationary. In fact, the circulant embedding is called with the Stein hypermodel, see RMmodel, and A=B there.

Here, multiplicative models are not allowed (yet).

Marked point processes.
Some methods are based on marked point process P = ([x_1,m_1], [x_2,m_2], ...) where the marks m_i are deterministic or i.i.d. random functions on R^d.

add.MPP (Random coins).
Here the functions are elements of the intersection (L1 cap L2) of the Hilbert spaces L1 and L2. A random field Z is obtained by adding the marks:

Z(.) = sum_i m_i( . - x_i)

In this package, only stationary Poisson point fields are allowed as underlying unmarked point processes. Thus, if the marks m_i are all indicator functions, we obtain a Poisson random field. If the intensity of the Poisson process is high we obtain an approximative Gaussian random field by the central limit theorem - this is the add.mpp method.

max.MPP (Boolean functions).
If the random functions are multiplied by suitable, independent random values, and then the maximum is taken, a max-stable random field with unit Frechet margins is obtained - this is the max.mpp method.

nugget.
The method allows for generating a random field of independent Gaussian random variables. This method is called automatically if the nugget effect is positive except the method "circulant embedding" or "direct" has been explicitly chosen.

The method has been extended to zonal anisotropies, see also argument nugget.tol in RFoptions.

particular method
– details missing –

Random coins.
See marked point processes.

sequential This method is programmed for spatio-temporal models where the field is modelled sequentially in the time direction conditioned on the previous k instances. For k=5 the method has its limits for about 1000 spatial points. It is an approximative method. The larger k the better. It also works for certain grids where the last dimension should contain the highest number of grid points.

spectral TBM (Spectral turning bands).
The principle of spectral TBM does not differ from the other turning bands methods. However, line simulations are performed by a spectral technique (Mantoglou and Wilson, 1982).

The standard method allows for the simulation of 2-dimensional random fields defined on arbitrary points or arbitrary grids. Here, a realisation is given as the cosine with random amplitude and random phase.

TBM2, TBM3 (Turning bands methods; turning layers).
It is generally difficult to use the turning bands method (TBM2) directly in the 2-dimensional space. Instead, 2-dimensional random fields are frequently obtained by simulating a 3-dimensional random field (using TBM3) and taking a 2-dimensional cross-section. TBM3 allows for multiplicative models; in case of anisotropy the anisotropy matrices must be multiples of the first matrix or the anisotropy matrix consists of a time component only (i.e. all components are zero except the very last one).
TBM2 and TBM3 allow for arbitrary points, and arbitrary grids (arbitrary number of points in each direction, arbitrary grid length for each direction).

Note: Both the precision and the simulation time depend heavily on TBM*.linesimustep and TBM*.linesimufactor that can be set by RFoptions. For covariance models with larger values of the scale parameter, TBM*.linesimufactor=2 is too small.

The turning layers are used for the simulations with time component. Here, if the model is a multiplicative covariance function then the product may contain matrices with pure time component. All the other matrices must be equal up to a factor and the temporal part of the anisotropy matrix (right column) may contain only zeros, except the very last entry.

Value

an invisible string vector of the Gaussian methods.

Automatic selection algorithm

— details coming soon —

Note

Most methods possess additional arguments, see RFoptions() that control the precision of the result. The default arguments are chosen such that the simulations are fine for many models and their parameters. The example in RFvariogram() shows a way of checking the precision.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. and Schlather, M. (2004) Statistical modeling with covariance functions. In preparation.

Lantuejoul, Ch. (2002) Geostatistical simulation. New York: Springer.

Schlather, M. (1999) An introduction to positive definite functions and to unconditional simulation of random fields. Technical report ST 99-10, Dept. of Maths and Statistics, Lancaster University.

Original work:

Circulant embedding:

Chan, G. and Wood, A.T.A. (1997) An algorithm for simulating stationary Gaussian random fields. J. R. Stat. Soc., Ser. C 46, 171-181.

Dietrich, C.R. and Newsam, G.N. (1993) A fast and exact method for multidimensional Gaussian stochastic simulations. Water Resour. Res. 29, 2861-2869.

Dietrich, C.R. and Newsam, G.N. (1996) A fast and exact method for multidimensional Gaussian stochastic simulations: Extensions to realizations conditioned on direct and indirect measurement Water Resour. Res. 32, 1643-1652.

Wood, A.T.A. and Chan, G. (1994) Simulation of stationary Gaussian processes in [0,1]^d J. Comput. Graph. Stat. 3, 409-432.

The code used in RandomFields is based on Dietrich and Newsam (1996).

Intrinsic embedding and Cutoff embedding:

Stein, M.L. (2002) Fast and exact simulation of fractional Brownian surfaces. J. Comput. Graph. Statist. 11, 587–599.

Gneiting, T., Sevcikova, H., Percival, D.B., Schlather, M. and Jiang, Y. (2005) Fast and Exact Simulation of Large Gaussian Lattice Systems in R^2: Exploring the Limits J. Comput. Graph. Statist. Submitted.

Markov Gaussian Random Field:

Rue, H. (2001) Fast sampling of Gaussian Markov random fields. J. R. Statist. Soc., Ser. B, 63 (2), 325-338.

Rue, H., Held, L. (2005) Gaussian Markov Random Fields: Theory and Applications. Monographs on Statistics and Applied Probability, no 104, Chapman \& Hall.

Turning bands method (TBM), turning layers:

Dietrich, C.R. (1995) A simple and efficient space domain implementation of the turning bands method. Water Resour. Res. 31, 147-156.

Mantoglou, A. and Wilson, J.L. (1982) The turning bands method for simulation of random fields using line generation by a spectral method. Water. Resour. Res. 18, 1379-1394.

Matheron, G. (1973) The intrinsic random functions and their applications. Adv. Appl. Probab. 5, 439-468.

Schlather, M. (2004) Turning layers: A space-time extension of turning bands. Submitted

Random coins:

Matheron, G. (1967) Elements pour une Theorie des Milieux Poreux. Paris: Masson.

See Also

RMmodel, RFsimulate, RandomFields.

Examples

RFgetMethodNames()

[Package RandomFields version 3.3.6 Index]