RFgetModel {RandomFields}	R Documentation

Internally stored model

Description

The function returns the stored model.

Usage

RFgetModel(register, explicite.natscale, show.call=FALSE,
           origin="original")

Arguments

register
0,...,21 or an evaluating function, e.g. RFsimulate. Place where intermediate calculations are stored. See also section Registers in RFoptions.

explicite.natscale
logical. Advanced option. If missing, then the model is returned as stored. If FALSE then any RMnatsc is ignored. If TRUE then any RMnatsc is tried to be combined with leading RMS, or returned as such.

show.call
logical or character. If FALSE then the model is shown as interpreted. If TRUE then the user's input including the calling function is returned. See example below.

If show.call is a character it behaves as which.submodels.

origin
character; one of "original", "MLE conform", "all". This argument determines the parameters that are returned.

Details

Whereas RFgetModel returns a model that can be re-used by the user, RFgetModelInfo can return detailed information.

Value

The stored model is returned in list format.

Note

Put Storing=TRUE, see RFoptions, if you like to have (more) internal information in case of failure of an initialization of a random field simulation.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RFgetModelInfo, RFsimulate.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
model <- RMexp(scale=4, var=2) + RMnugget(var=3) + RMtrend(mean=1)
z <- RFsimulate(model, 1:4)
RFgetModel(show.call=FALSE)
RFgetModel(show.call=TRUE)

[Package RandomFields version 3.3.6 Index]