RFgetModelInfo {RandomFields}	R Documentation

Information on RMmodels

Description

The function returns information about an RMmodel, either internal information when used in simulations, for instance, or general information

Usage

RFgetModelInfo(...) 

RFgetModelInfo_register(register, level = 1, spConform =
                 RFoptions()$general$spConform, which.submodels =
                 c("user", "internal", "call+user", "call+internal",
                 "user.but.once", "internal.but.once",
                 "user.but.once+jump", "internal.but.once+jump", "all"),
                 modelname = NULL, origin = "original")

RFgetModelInfo_model(model, params, dim = 1, Time = FALSE,
                     kernel = FALSE, exclude_trend = TRUE, ...)

Arguments

...
See the argument of RFgetModelInfo_register and RFgetModelInfo_model; RFgetModelInfo is an abbreviation for the other two functions.

register
0,...,21 or an evaluating function, e.g. RFsimulate. Place where intermediate calculations are stored. See also section Registers in RFoptions.

level
integer [0...5]; level of details, i.e. the higher the number the more details are given.

spConform
see RFoptions

which.submodels
Internally, the sub-models are represented in two different ways: ‘internal’ and ‘user’. The latter is very close to the model defined by the user.

Most models have a leading internal model. The values "call+user" and "call+internal" also return this leading model if existent.

The values "user.but.once", "internal.but.once" "user.but.once" returns the user path of the internal model following the leading model. "internal.but.once" would return the internal path of the user model following the leading model, but this path should never exist. So as all the other options if a certain direction does not exist, the alternative path is taken.

The values "user.but.once+jump", "internal.but.once+jump" same as "user.but.once" and "internal.but.once", except that the first submodel below the leading model is not given.

The value "all" returns the whole tree of models (very advanced).

modelname
string. If modelname is given then it returns the first appearance of the covariance model with name modelname. If meth is given then the model within the method is returned.

model,params
object of class RMmodel, RFformula or formula; best is to consider the examples below, first.
The argument params is a list that specifies free parameters in a formula description, see RMformula. Here, NAs should be placed where information on the parameters is desired..

dim
positive integer. Spatial dimension.

Time
logical. Should time be considered, too?

kernel
logical. Should the model be considered as a kernel?

exclude_trend
logical. Currently, only TRUE is available.

origin
character; one of "original", "MLE conform", "all". This argument determines the parameters that are returned.

Details

RFgetModelInfo branches either into RFgetModelInfo_register or RFgetModelInfo_model, depending on the type of the first argument. The latter two are usually not called by the user.

RFgetModelInfo has three standard usages:

RFgetModelInfo() returns internal information on the last call of an RF function.

RFgetModelInfo(RFfunction) returns internal information on the last call of RFfunction.

RFgetModelInfo(RMmodel) returns general information on RMmodel

Whereas RFgetModelInfo() can return detailed internal information, RFgetModel returns a model that can be re-used by the user.

Value

If RFgetModelInfo(model) is called a list is returned with the following elements:

trans.inv : logical. Whether the model is translation invariant (stationary)

isotropic : logical. Whether the model is rotation invariant (stationary)

NAs : in case of an additive model it gives the number of NAs in each submodel

minmax : a data frame containing information on all arguments set to NAs

pmin, pmax : lower and upper endpoint of the parameter values usually found in practice

type : integer; recognized particularities of a parameter; an explanation of the values is given after the table, if printed.

NAN : the number of NANs found

min, max : mathematically valid lower and upper endpoints of the parameter values

omin, omax : logical. If FALSE the respective mathematical endpoint is included

col, row : the dimension of the parameter. If the parameter is a scalar then col = row = 1. If it is a vector then col = 1.

bayes : currently not used (always FALSE)

Else a list of internal structure is returned.

Note

Put Storing=TRUE, see RFoptions if you like to have more internal information in case of failure of an initialisation of a random field simulation.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

commandRFgetModel, RFsimulate

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMexp(scale=4, var=2) + RMnugget(var=3) + RMtrend(mean=1)
z <- RFsimulate(model, 1:4, storing=TRUE)
RFgetModelInfo()

model <-  RMwhittle(scale=NA, var=NA, nu=NA) + RMnugget(var=NA)
RFgetModelInfo(model)

[Package RandomFields version 3.3.6 Index]