RFgetModelNames {RandomFields}	R Documentation

Names of implemented covariance and variogram models

Description

Displays the names of covariance and variogram models (see RMmodel) and returns them as a list. The user may specify and group the models according to the following properties:

type of function ("positive definite", "variogram", etc.)

whether the function depends on two arguments ("kernel") or on one argument only ("single variable")

types of isotropy

whether the model is an operator

whether the model is a normal scale mixture

whether the model has a finite range covariance

validity in certain dimensions of the coordinate space

maximal possible dimension of the coordinate space

uni- or multivariety

See Details for an explanation and RMmodelgenerator for possible states (values) of these properties.

Usage

RFgetModelNames(type = RC_TYPE_NAMES, domain = RC_DOMAIN_NAMES,
                isotropy = RC_ISO_NAMES, operator = c(TRUE, FALSE),
                monotone = RC_MONOTONE_NAMES,
                implied_monotonicities = length(monotone) == 1,
                finiterange = c(TRUE, FALSE, NA),
                valid.in.dim = c(1, Inf), 
                vdim = c(1, 5),
                group.by,
                exact.match = !missing(group.by),
                simpleArguments = FALSE,
                internal, newnames)

Arguments

type, domain, isotropy, operator, monotone, finiterange, vdim
see constants for the definition of RC_TYPE_NAMES, RC_DOMAIN_NAMES, etc. See also RMmodelgenerator.

implied_monotonicities
logical. If TRUE then all the models with a stronger monotonocity than the required one are also shown.

valid.in.dim
an optional integer indicating the dimension of the space where the model is valid

group.by
an optional character string or NULL; must be one of 'type', 'domain', 'isotropy', 'operator', 'monotone', 'finiterange','maxdim','vdim'. If group.by is not given, the result is grouped by 'type' if more than one type is given.

exact.match
logical. If not TRUE, then all categories that are subclasses or might match are show as well.

simpleArguments
logical. If TRUE, only models are considered whose arguments are all integer or real valued.

internal, newnames
both logical; internal might be also integer valued. If any of them are given, RFgetModelNames behaves very differently. See the Notes below.

Details

The plain call RFgetModelNames() simply gives back a vector of the names of all implemented covariance and variogram models and operators, i.e. members of the class RMmodelgenerator.

The following arguments can be specified. In general, only exact matches are returned. One exception exists: If the length of type equals 1 and if group.by is not given, then types included in type are also returned. E.g. if type="variogram" and group.by is not given then only models are returned that are negative definite. However, also positive definite functions and tail correlaton functions are returned if "type" is included in group.by.

type

specifies the class of functions; for the meaning of the possible values see RMmodelgenerator

stationarity

specifies the type of stationarity; for the meaning of the possible values see RMmodelgenerator

isotropy

specifies the type of isotropy; for the meaning of the possible values see RMmodelgenerator

operator

indicates whether the model is an operator, i.e. it requires at least one submodel, e.g. + or RMdelay are operators; see RMmodelgenerator

monotone

indicates what kind of monotonicity is known, e.g., whether the model is a normal scale mixture, the latter including RMexp or RMcauchy; see RMmodelgenerator

finiterange

indicates whether the covariance of the model has finite range, e.g. RMcircular or RMnugget have covariances with finite range; see RMmodelgenerator. NA is used if the finiteness depends on the submodel.

valid.in.dim

If valid.in.dim=n is passed, all models which are valid in dimension n are displayed. Otherwise valid.in.dim should be a bivariate vector giving the range of requested dimensions.

maxdim

if a positive integer, it specifies the maximal possible dimension of the coordinate space; note that a model which is valid in dimension n is also valid in dimension n-1; maxdim=-1 means that the maximal possible dimension depends on the parameters of the RMmodel object; maxdim=-2 means that the maximal possible dimension is adopted from the called submodels; see also RMmodelgenerator

vdim

if a positive integer, vdim specifies, whether the model is vdim-variate; vdim=-1 means that being multivariate in a certain dimension depends on the parameters of the RMmodel object; vdim=-2 means that being multivariate in a certain dimension is adopted from the called submodels; see also RMmodelgenerator

If vdim is bivariate then a range is given.

group.by

If group.by="propertyname" is passed, the displayed models are grouped according to propertyname.

All arguments allow also for vectors of values. In case of valid.in.dim the smallest value is taken. The interpretation is canonical.

Note that the arguments stationarity, isotropy, operator, monotone, finiterange, maxdim, vdim are also slots (attributes) of the SP4-class RMmodelgenerator.

Value

Either a vector of model names if the argument group.by is not used; or a list of vectors of model names if the argument group.by is used (with list elements specified by the categories of the grouping argument).

In case internal or newnames is given, RFgetModelNames prints a table of the currently implemented covariance functions and the matching methods. RFgetModelNames returns NULL.

Note

In case internal or newnames is given, only the values of internal, newnames and operator are considered. All the other arguments are ignored and RFgetModelNames prints a table of the currently implemented covariance functions and the matching methods:

internal:
if TRUE also RMmodels are listed that are internal, hence invisible to the user. Default: FALSE.

newnames:
The model names of version 2 of RandomFields and earlier can still be used in the model definitions. Namely when the list notation is chosen; see Advanced RMmodels for the latter. If internal or newnames is given, then these old names are shown; if newnames=TRUE then also the usual names are shown. Default: FALSE.

In fact, both internal and public models can have different variants implemented. These variants are also shown if internal has a value greater than or equal to 2,

operator:
see above.

Here, also an indication is given, which method for simulating Gaussian random fields matches the model.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

constants, RMmodelgenerator, RMmodel, RandomFields, RC_DOMAIN_NAMES, RC_ISO_NAMES

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

# get list of names of all functions
RFgetModelNames()

# any kind of positive definite functions
RFgetModelNames(type="positive definite", exact.match=TRUE)
## Not run: RFgetModelNames(type="positive definite")

# get a list of names of all stationary models
RFgetModelNames(type="positive definite", domain="single variable",
                 exact.match=TRUE)
## Not run: RFgetModelNames(type="positive definite", domain="single variable")

# get a vector of all model names
RFgetModelNames(group.by=NULL)

[Package RandomFields version 3.3.6 Index]