RFgui {RandomFields}	R Documentation

Graphical User Interface For Fitting Covariance Models And Variograms

Description

This is a nice instructive graphical tool useful in particular for teaching classes

Usage

RFgui(data, x, y, same.algorithm = TRUE, ev, bin = NULL, xcov, ycov,
      sim_only1dim=FALSE, wait = 0, ...)

Arguments

data
See RFvariogram. If data is given, the empirical variogram is shown.

x
a sequence of the locations of the simulated process; if not given, x is determined by data and if data is not given by default values

y
a sequence of numbers if a simulation on R^d is performed. Default is y = x; see x for details.

same.algorithm
Force the picture being simulated with the same algorithm so that the pictures are always directly comparable. The disadvantage is that some models are simulated only (very) approximatively.

ev
instead of the data, the empirical variogram itself might be passed

bin
only considered if data is given. See RFvariogram for details.

xcov
sequence of the locations where the covariance function is plotted

ycov
Only for anisotropic models. sequence of the locations where the covariance function is also plotted

sim_only1dim
Logical. The argument determines whether a process should be simulated on the line or on the plane

wait
integer. See details.

...
further options and control arguments for the simulation that are passed to and processed by RFoptions.

Details

If wait is negative the xterm does not wait for the tkltk-window to be finished. Further the variable RFgui.model is created in the environment .GlobalEnv and contains the currently chosen variable in the gui. RFgui always returns NULL.

If wait is non-negative the xterm waits for the tkltk-window to be finished. RFgui returns invisibly the last chosen model (or NULL if no model has been chosen). RFgui idles a lot when wait=0. It idles less for higher values by sleeping about wait microseconds. Of course the handling in the tkltk window gets slower as well. Reasonable values for wait are within [0,1000].

same.alg = TRUE is equivalent to setting circulant.trials=1, circulant.simu_method = "RPcirculant", circulant.force=TRUE, circulant.mmin=-2.

Value

If wait < 0 the function returns NULL else it returns the last chosen RMmodel.

If wait < 0, a side effect of RFgui is the creation of the variable RFgui.model on .GlobalEnv.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

Author(s) of the code

Daphne Boecker; Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

soil for a further example

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
RFgui()

[Package RandomFields version 3.3.6 Index]