RFhurst {RandomFields}	R Documentation

Hurst coefficient

Description

The function estimates the Hurst coefficient of a process

Usage

RFhurst(x, y = NULL, z = NULL, data, sort = TRUE,
 block.sequ = unique(round(exp(seq(log(min(3000, dimen[1]/5)),
 log(dimen[1]),
 len = min(100, dimen[1]))))),
 fft.m = c(1, min(1000, (fft.len - 1)/10)),
 fft.max.length = Inf, method = c("dfa", "fft", "var"),
 mode = if (interactive ()) c("plot", "interactive") else "nographics", 
 pch = 16, cex = 0.2, cex.main = 0.85,
 printlevel = RFoptions()$basic$printlevel, height = 3.5,
 ...)

Arguments

x
vector of x coordinates, or object of class GridTopology or raster; for more options see RFsimulateAdvanced.

y,z
optional vectors of y (z) coordinates, which should not be given if x is a matrix.

data
the data

sort
logical. If TRUE then the coordinates are permuted such that the largest grid length is in x-direction; this is of interest for algorithms that slice higher dimensional fields into one-dimensional sections.

block.sequ
ascending sequences of block lengths for which the detrended fluctuation analysis and the variance method are performed.

fft.m
vector of 2 integers; lower and upper endpoint of indices for the frequency which are used in the calculation of the regression line for the periodogram near the origin.

fft.max.length
if the number of points in x-direction is larger than fft.max.length then the segments of length fft.max.length are considered, shifted by fft.max.length/2 (WOSA-estimator).

method
list of implemented methods to calculate the Hurst parameter; see Details

mode
character. A vector with components 'nographics', 'plot' or 'interactive':

'nographics'

no graphical output

'plot'

the regression line is plotted

'interactive'

the regression domain can be chosen interactively

Usually only one mode is given. Two modes may make sense in the combination c("plot", "interactive") in which case all the results are plotted first, and then the interactive mode is called. In the interactive mode, the regression domain is chosen by two mouse clicks with the left mouse; a right mouse click leaves the plot.

pch
vector or scalar; sign by which data are plotted.

cex
vector or scalar; size of pch.

cex.main
font size for title in regression plot; only used if mode includes 'plot' or 'interactive'

printlevel
integer. If printlevel is 0 or 1 nothing is printed. If printlevel=2 warnings and the regression results are given. If printlevel>2 tracing information is given.

height
height of the graphics window

...
graphical arguments

Details

The function is still in development. Several functionalities do not exist - see the code itself for the current stage.

The function calculates the Hurst coefficient by various methods:

detrended fluctuation analysis (dfa)

aggregated variation (var)

periodogram or WOSA estimator (fft)

Value

The function returns a list with elements dfa, varmeth, fft corresponding to the three methods given in the Details.

Each of the elements is itself a list that contains the following elements.

x
the x-coordinates used for the regression fit

y
the y-coordinates used for the regression fit

regr
the coefficients of the lm.

sm
smoothed curve through the (x,y) points

x.u
NULL or the restricted x-coordinates given by the user in the interactive plot

y.u
NULL or y-coordinates according to x.u

regr.u
NULL or the coefficients of lm for x.u and y.u

H
the Hurst coefficient

H.u
NULL or the Hurst coefficient corresponding to the user's regression line

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

detrended fluctuation analysis

Peng, C.K., Buldyrev, S.V., Havlin, S., Simons, M., Stanley, H.E. and Goldberger, A.L. (1994) Mosaic organization of DNA nucleotides Phys. Rev. E 49, 1685-1689

aggregated variation

Taqqu, M.S. and Teverovsky, V. (1998) On estimating the intensity of long range dependence in finite and infinite variance time series. In: Adler, R.J., Feldman, R.E., and Taqqu, M.S. A Practical Guide to Heavy Tails, Statistical Techniques an Applications. Boston: Birkhaeuser

Taqqu, M.S. and Teverovsky, V. and Willinger, W. (1995) Estimators for long-range dependence: an empirical study. Fractals 3, 785-798

periodogram

Percival, D.B. and Walden, A.T. (1993) Spectral Analysis for Physical Applications: Multitaper and Conventional Univariate Techniques, Cambridge: Cambridge University Press.

Welch, P.D. (1967) The use of Fast Fourier Transform for the estimation of power spectra: a method based on time averaging over short, modified periodograms IEEE Trans. Audio Electroacoustics 15, 70-73.

See Also

RMmodel, RFfractaldim

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

x <- runif(1000)
h <- RFhurst(1:length(x), data=x)

[Package RandomFields version 3.3.6 Index]