RFoldstyle {RandomFields}	R Documentation

RFoldstyle

Description

This function is written only for package writers who have based their code on RandomFields version 2.

It avoids warnings if the old style is used, and sets spConform = FALSE.

Usage

RFoldstyle(old=TRUE) 

Arguments

old
logical

Value

NULL

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

See ‘version2’ for details on the commands of version 2.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

GaussRF(x=1:10, model="exp", param=c(0,1,0,1), grid=TRUE)

RFoldstyle()
GaussRF(x=1:10, model="exp", param=c(0,1,0,1), grid=TRUE)

[Package RandomFields version 3.3.6 Index]