RFpar {RandomFields}	R Documentation

Graphical parameters for plots

Description

This function sets globally graphical parameters for plots of RMmodels, simulations and estimations.

Usage

RFpar(...)

Arguments

...
see par

Value

If RFpar is called without arguments, the current list is returned.

If RFpar is called with NULL only, the current list is deleted.

Otherwise the arguments are stored for global use in RandomFields.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

plot-method

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

RFpar(col="red")
plot(RMexp())

[Package RandomFields version 3.3.6 Index]