RFsp-class {RandomFields}	R Documentation

Class RFsp

Description

"RFsp" is a virtual class which contains the four classes RFspatialGridDataFrame (data on a full grid and space-time-dimension > 2), RFspatialPointsDataFrame (data not on a grid and space-time-dimension > 2), RFgridDataFrame (data on a full grid and space-time-dimension = 1), RFpointsDataFrame (data not on a grid and space-time-dimension = 1)

The first two subclasses are summarized in "RFspatialDataFrame" whilst the latter two are summarized in "RFdataFrame".

Objects from the Class

are never to be generated; only derived classes can be meaningful.

Methods

summary

signature(obj = "RFsp"): returns a summary of the object; uses or imitates summary method of class Spatial from the sp-package

dimensions

signature(obj = "RFsp"): retrieves the number of spatial or spatio-temporal dimensions spanned

RFspDataFrame2dataArray

signature(obj = "RFsp"): transforms RFsp objects to array

RFspDataFrame2conventional

signature(obj = "RFsp"): transforms RFsp objects to a list with additional information

[

signature(obj = "RFsp"): selects columns of the data-slot, while all other slots are kept unmodified

[<-

signature(obj = "RFsp"): replaces columns of the data-slot, while all other slots are kept unmodified

variance

signature(object = "RFsp"): returns the kriging variance if available

Warning

This class is not useful itself, but the above mentioned classes in this package derived from it.

Author(s)

Alexander Malinowski; Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RFspatialGridDataFrame, RFspatialPointsDataFrame, RFgridDataFrame, RFpointsDataFrame, sp2RF

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## Generating an object of class "RFspatialGridDataFrame"
## and one of class "RFspatialPointsDataFrame".

model <- RMcauchy(gamma=4, var=0.1, scale=2)
x <- seq(0,10,len=100)
r <- cbind(runif(100, min=1, max=9),runif(100, min=1, max=9))

z1 <- RFsimulate(model=model, x=x, y=x, n=4)
z2 <- RFsimulate(model=model, x=r, n=2)

## Applying available functions.

class(z1)
class(z2)
summary(z1)
summary(z2)
plot(z1)
plot(z2)

z4 <- RFspDataFrame2conventional(z2)
str(z2)
str(z4)

dta <- data.frame(coords=z4$x,data=z2@data[,1])
z3<-RFinterpolate(model=model, x=x, y=x, data=dta) 
plot(z3,z2)

## Illustrating the warning.

a1 <- new("RFpointsDataFrame")
str(a1)
try(a2 <- new("RFsp")) ## ERROR

[Package RandomFields version 3.3.6 Index]