RMmodels Overview {RandomFields}	R Documentation

Overview over classes of RMmodels

Description

Various classes of models RMxxx are implemented in RandomFields, that have their own man pages. Here, an overview over these man pages are given.

Man pages

Beginners should start with RMmodels, then go for RMmodelsAdvanced if more information is needed.

RMmodels 	 general introduction and a collection of simple models

RMmodelsAdvanced 	 includes more advanced stationary and isotropic models, variogram models, non-stationary models and trend models

Bayesian 	 hierarchical models

RMmodelsMultivariate 	 multivariate covariance models and multivariate trend models

RMmodelsNonstationary 	 non-stationary covariance models

RMmodelsSpaceTime 	 space-time covariance models

Spherical models 	 models based on the polar coordinate system, usually used in earth models

Tail correlation functions 	 models related to max-stable random fields

trend modelling 	 how to pass trend specifications

Mathematical functions 	 simple mathematical functions that are typically used to build non-stationary covariance models and arbitrary trends

RMmodelsAuxiliary 	 rather specialized models, most of them not having positive definiteness property, but used internally in certain simulation algorithms, for instance.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RC, RR, RF, R.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

RFgetModelNames(type="positive definite", domain="single variable",
                isotropy="isotropic", operator=!FALSE) ## RMmodel.Rd

[Package RandomFields version 3.3.6 Index]