RMS {RandomFields}	R Documentation

Scaling operator

Description

RMS is an operator that modifies the variance and the coordinates or distances of a submodel φ by

C(h) = v * φ(A*h/s).

Most users will never call RMS directly, see Details. However, the following describes the arguments var, scale, Aniso, proj that are common to nearly all models. See RMSadvanced for advanced use of these arguments.

Usage

RMS(phi, var, scale, Aniso, proj, anisoT)

Arguments

phi
submodel

var
is the optional variance parameter v.

scale
scaling parameter s which is positive.

Aniso
matrix or RMmodel. The optional anisotropy matrix A, multiplied from the right by a distance vector x, i.e. Ax.

proj
is the optional projection vector which defines a diagonal matrix of zeros and ones and proj gives the positions of the ones (integer values between 1 and the dimension of x). It also allows for the values 'space' and 'time' in case of space-time modelling.

anisoT
the transpose of the anisotropy matrix B, multiplied from the left by a distance vector x, i.e. x^\top B.

Details

The call in the usage section is equivalent to phi(..., var, scale, anisoT, Aniso, proj), where phi has to be replaced by a valid RMmodel.

Most users will never call RMS directly.

Value

RMS returns an object of class RMmodel.

Note

At most one of the arguments Aniso, anisoT and proj may be given at the same time.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMSadvanced, RMmodel,

RMprod for an alternative way to define an arbitrary, location dependent variance. There, the standard deviation is given so that RMprod might be used even in the multivariate case.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
model1 <- RMS(RMexp(), scale=2)
model2 <- RMexp(scale=2)
x <- seq(0, 10, 0.02)
print(all(RFcov(model1, x) == RFcov(model2, x))) # TRUE

[Package RandomFields version 3.3.6 Index]