RMSadvanced {RandomFields}	R Documentation

Scaling operator – comments for advanced applications

Description

Here advances uses are given for the arguments var, scale, Aniso, proj that are available to most of the models

Usage

RMS(phi, var, scale, Aniso, proj, anisoT)

Arguments

phi

submodel

var

Instead of a constant it can be also an arbitrary non-negative function, see R. and RMuser for defining arbitrary functions.

scale

instead of a positive constant it can be an arbitrary, positive deterministic function. In case of the latter, the scale should be given by one of the functions RMbubble or RMscale. In case none of them are given, RMscale is assumed with scale penality \|s(x) - s(y)\|^2 for the square of the norm.

The scale can be also a random variable in case of Bayesian modelling.

Aniso

matrix or RMmodel. Instead of a matrix, Aniso can be an arbitrary, vector-valued function .

proj

is the optional projection vector which defines a diagonal matrix of zeros and ones and proj gives the positions of the ones (integer values between 1 and the dimension of x). It also allows for the values 'space' and 'time' in case of space-time modelling.

anisoT

the transpose of the anisotropy matrix B, multiplied from the left by a distance vector x, i.e. x^\top B.

Details

See the reference for Gneitings nsst model used for modelling scales. See also the example below.

Value

RMSadvanced returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Bonat, W.H. , Ribeiro, P. Jr. and Schlather, M. (2019) Modelling non-stationarity in scale. In preparation.

See Also

RMS, RMblend for a different approach on modelling different scales

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

x <- seq(0,1, if (interactive()) 0.01 else 0.5)
d <- sqrt(rowSums(as.matrix(expand.grid(x-0.5, x-0.5))^2))
d <- matrix(d < 0.25, nc=length(x))
image(d)

scale <- RMcovariate(data=as.double(d) * 2 + 0.5, raw=TRUE)

S <- RMexp(scale = scale)
plot(zS <- RFsimulate(S, x, x))
CS <- RFcovmatrix(S, x, x)

[Package RandomFields version 3.3.6 Index]