RMaskey {RandomFields}	R Documentation

Askey model

Description

Askey's model

C(x)= (1-x)^α 1_{[0,1]}(x)

Usage

RMaskey(alpha, var, scale, Aniso, proj)
RMtent(var, scale, Aniso, proj)

Arguments

alpha
a numerical value in the interval [0,1]

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

This covariance function is valid for dimension d if α ≥ (d+1)/2. For α=1 we get the well-known triangle (or tent) model, which is only valid on the real line.

Value

RMaskey returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Covariance function

Askey, R. (1973) Radial characteristic functions. Technical report, Research Center, University of Wisconsin-Madison.

Golubov, B. I. (1981) On Abel-Poisson type and Riesz means, Anal. Math. 7, 161-184.

Applications as covariance function

Gneiting, T. (1999) Correlation functions for atmospheric data analysis. Quart. J. Roy. Meteor. Soc., 125:2449-2464.

Gneiting, T. (2002) Compactly supported correlation functions. J. Multivar. Anal., 83:493-508.

Wendland, H. (1994) Ein Beitrag zur Interpolation mit radialen Basisfunktionen. Diplomarbeit, Goettingen.

Wendland, H. Piecewise polynomial, positive definite and compactly supported radial functions of minimal degree. Adv. Comput. Math., 4:389-396, 1995.

Tail correlation function (for α ≥ [d / 2] + 1)

Strokorb, K., Ballani, F., and Schlather, M. (2014) Tail correlation functions of max-stable processes: Construction principles, recovery and diversity of some mixing max-stable processes with identical TCF. Extremes, Submitted.

See Also

RMmodel, RMbigneiting, RMgengneiting, RMgneiting, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMtent()
x <- seq(0, 10, 0.02) 
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]