RMball {RandomFields}	R Documentation

RMball

Description

RMball refers to the indicator function of a ball with radius 1.

Usage

RMball(var, scale, Aniso, proj)

Arguments

var, scale, Aniso, proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMpolygon, RMspheric, RFsimulate, RMmodel.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

x <- seq(0,10,len=100)
model <- RMball(var=2,scale=1.5)
plot(model)
z <- RFsimulate(RPpoisson(model),x=x,y=x,intensity=0.1)
plot(z)

[Package RandomFields version 3.3.6 Index]