RMbcw {RandomFields}	R Documentation

Model bridging stationary and intrinsically stationary processes

Description

RMbcw is a variogram model that bridges between some intrinsically stationary isotropic processes and some stationary ones. It reunifies the RMgenfbm ‘b’, RMgencauchy ‘c’ and RMdewijsian ‘w’.

The corresponding centered semi-variogram only depends on the distance r ≥ 0 between two points and is given by

γ(r)=[(r^{α}+1)^{β/α}-1] / (2^{β/α}-1)

where 0 < α ≤ 2 and β <= 2.

Usage

RMbcw(alpha, beta, c, var, scale, Aniso, proj)

Arguments

alpha
a numerical value; should be in the interval (0,2].

beta
a numerical value; should be in the interval (-infty,2].

c
only for experts. If given, a not necessarily positive definite function c-γ(r) is built.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above variogram remains unmodified.

Details

For β >0, β<0, β=0 we have the generalized fractal Brownian motion RMgenfbm, the generalized Cauchy model RMgencauchy, and the de Wisjian model RMdewijsian, respectively.

Hence its two arguments alpha and beta allow for modelling the smoothness and a wide range of tail behaviour, respectively.

Value

RMbcw returns an object of class RMmodel

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Schlather, M (2014) A parametric variogram model bridging between stationary and intrinsically stationary processes. arxiv 1412.1914.

See Also

RMlsfbm is equipped with Matheron's constant c for the fractional brownian motion, RMgenfbm, RMgencauchy, RMdewijsian, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMbcw(alpha=1, beta=0.5)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]