RMbicauchy {RandomFields}	R Documentation

Bivariate Cauchy Model

Description

RMbicauchy is a bivariate stationary isotropic covariance model whose corresponding covariance function only depends on the distance r ≥ 0 between two points.

For constraints on the constants see Details.

Usage

RMbicauchy(alpha, beta, s, rho, var, scale, Aniso, proj)

Arguments

alpha
[to be done]

beta
[to be done]

s
a vector of length 3 of numerical values; each entry positive; the vector (s_{11},s_{21},s_{22})

rho
[to be done]

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

Constraints on the constants: [to be done]

Value

RMbicauchy returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Moreva, O., Schlather, M. (2016) Modelling and simulation of bivariate Gaussian random fields. arXiv 1412.1914

See Also

RMcauchy, Multivariate RMmodels.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## todo

[Package RandomFields version 3.3.6 Index]