RMbigneiting {RandomFields}	R Documentation

Gneiting-Wendland Covariance Models

Description

RMbigneiting is a bivariate stationary isotropic covariance model family whose elements are specified by seven parameters.

Let

δ_{ij} = μ + γ_{ij} + 1.

Then,

C_{n}(h) = c_{ij} (C_{n, δ} (h / s_{ij}))_{i,j=1,2}

and C_{n, δ} is the generalized Gneiting model with parameters n and δ, see RMgengneiting, i.e.,

C_{κ=0, δ}(r) = (1 - r)^β 1_{[0,1]}(r), β=δ + 2κ + 1/2;

C_{κ=1, δ}(r) = (1+ β r)(1-r)^β 1_{[0,1]}(r), β = δ + 2κ + 1/2;

C(_{κ=2, δ}(r) = (1 + β r + (β^2-1) r^(2)/3)(1-r)^β 1_{[0,1]}(r), β = δ + 2κ + 1/2;

C_{κ=3, δ}(r) = (1 + β r + (2 β^2-3 )r^(2)/5+(β^2 - 4) β r^(3)/15)(1-r)^β 1_{[0,1]}(r), β=δ + 2κ + 1/2.

Usage

RMbigneiting(kappa, mu, s, sred12, gamma, cdiag, rhored, c, var, scale, Aniso, proj)

Arguments

kappa
argument that chooses between the four different covariance models and may take values 0,...,3. The model is k times differentiable.

mu
mu has to be greater than or equal to d/2 where d is the (arbitrary) dimension of the random field.

s
vector of two elements giving the scale of the models on the diagonal, i.e. the vector (s_{11}, s_{22}).

sred12
value in [-1,1]. The scale on the offdiagonals is given by s_{12} = s_{21} = sred12 * min{s_{11}, s_{22}}.

gamma
a vector of length 3 of numerical values; each entry is positive. The vector gamma equals (γ_{11},γ_{21},γ_{22}). Note that γ_{12} =γ_{21}.

cdiag
a vector of length 2 of numerical values; each entry positive; the vector (c_{11},c_{22}).

c
a vector of length 3 of numerical values; the vector (c_{11}, c_{21}, c_{22}). Note that c_{12}= c_{21}.

Either rhored and cdiag or c must be given.

rhored
value in [-1,1]. See also the Details for the corresponding value of c_{12}=c_{21}.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

A sufficient condition for the constant c_{ij} is

c_{ij} = ρ_r m (c_{11} c_{22})^{1/2}

where ρ_r in [-1,1].

The constant m in the formula above is obtained as follows:

m = min{1, m_{-1}, m_{+1}}

Let

a = 2 γ_{12} - γ_{11} -γ_{22}

b = -2 γ_{12} (s_{11} + s_{22}) + γ_{11} (s_{12} + s_{22}) + γ_{22} (s_{12} + s_{11})

e = 2 γ_{12} s_{11}s_{22} - γ_{11}s_{12}s_{22} - γ_{22}s_{12}s_{11}

d = b^2 - 4ae

t_j =(-b + j √ d) / (2 a)

If d ≥0 and t_j in (0, s_{12})^c then m_j=∞ else

m_j = \frac{(1 - t_j/s_{11})^{γ_{11}}(1 - t_j/s_{22})^{γ_{22}}}{(1 - t_j/s_{12})^{2 γ_{11}} }{ m_j = (1 - t_j/s_{11})^{γ_{11}} (1 - t_j/s_{22})^{γ_{22}} / (1 - t_j/s_{12})^{2 γ_{11}} }

In the function RMbigneiting, either c is passed, then the above condition is checked, or rhored is passed; then c_{12} is calculated by the above formula.

Value

RMbigneiting returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Bevilacqua, M., Daley, D.J., Porcu, E., Schlather, M. (2012) Classes of compactly supported correlation functions for multivariate random fields. Technical report.

RMbigeneiting is based on this original work. D.J. Daley, E. Porcu and M. Bevilacqua have published end of 2014 an article intentionally without clarifying the genuine authorship of RMbigneiting, in particular, neither referring to this original work nor to RandomFields, which has included RMbigneiting since version 3.0.5 (05 Dec 2013).

Gneiting, T. (1999) Correlation functions for atmospherical data analysis. Q. J. Roy. Meteor. Soc Part A 125, 2449-2464.

Wendland, H. (2005) Scattered Data Approximation. Cambridge Monogr. Appl. Comput. Math.

See Also

RMaskey, RMbiwm, RMgengneiting, RMgneiting, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMbigneiting(kappa=2, mu=0.5, gamma=c(0, 3, 6), rhored=1)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]