RMbiwm {RandomFields}	R Documentation

Full Bivariate Whittle Matern Model

Description

RMbiwm is a bivariate stationary isotropic covariance model whose corresponding covariance function only depends on the distance r ≥ 0 between two points and is given for i,j = 1,2 by

C_{ij}(r)=c_{ij} W_{ν_{ij}}(r/s_{ij}).

Here W_ν is the covariance of the RMwhittle model. For constraints on the constants see Details.

Usage

RMbiwm(nudiag, nured12, nu, s, cdiag, rhored, c, notinvnu, var,
 scale, Aniso, proj)

Arguments

nudiag
a vector of length 2 of numerical values; each entry positive; the vector (ν_{11},ν_{22})

nured12
a numerical value in the interval [1,∞); ν_{21} is calculated as 0.5 (ν_{11} + ν_{22})*ν_{red}.

nu
alternative to nudiag and nured12: a vector of length 3 of numerical values; each entry positive; the vector (ν_{11},ν_{21},ν_{22}). Either nured and nudiag, or nu must be given.

s
a vector of length 3 of numerical values; each entry positive; the vector (s_{11},s_{21},s_{22}).

cdiag
a vector of length 2 of numerical values; each entry positive; the vector (c_{11},c_{22}).

rhored
a numerical value; in the interval [-1,1]. See also the Details for the corresponding value of c_{12}=c_{21}.

c
a vector of length 3 of numerical values; the vector (c_{11},c_{21}, c_{22}). Either rhored and cdiag or c must be given.

notinvnu
logical or NULL. If not given (default) then the formula of the (RMwhittle) model applies. If logical then the formula for the RMmatern model applies. See there for details.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

Constraints on the constants: For the diagonal elements we have

ν_{ii}, s_{ii}, c_{ii} > 0.

For the offdiagonal elements we have

s_{12}=s_{21} > 0,

ν_{12} =ν_{21} = 0.5 (ν_{11} + ν_{22}) * ν_{red}

for some constant ν_{red} \in [1,∞) and

c_{12} =c_{21} = ρ_{red} √{f m c_{11} c_{22}}

for some constant ρ_{red} in [-1,1].

The constants f and m in the last equation are given as follows:

f = (Γ(ν_{11} + d/2) Γ(ν_{22} + d/2)) / (Γ(ν_{11}) Γ(ν_{22})) * (Γ(ν_{12}) / Γ(ν_{12}+d/2))^2 * ( s_{12}^{2*ν_{12}} / (s_{11}^{ν_{11}} s_{22}^{ν_{22}}) )^2

where Γ is the Gamma function and d is the dimension of the space. The constant m is the infimum of the function g on [0,∞) where

g(t) = (1/s_{12}^2 +t^2)^{2ν_{12} + d} (1/s_{11}^2 + t^2)^{-ν_{11}-d/2} (1/s_{22}^2 + t^2)^{-ν_{22}-d/2}

(cf. Gneiting, T., Kleiber, W., Schlather, M. (2010), Full Bivariate Matern Model (Section 2.2)).

Value

RMbiwm returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T., Kleiber, W., Schlather, M. (2010) Matern covariance functions for multivariate random fields JASA

See Also

RMparswm, RMwhittle, RMmodel, RFsimulate, RFfit, Multivariate RMmodels.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

x <- y <- seq(-10, 10, 0.2)
model <- RMbiwm(nudiag=c(0.3, 2), nured=1, rhored=1, cdiag=c(1, 1.5), 
                s=c(1, 1, 2))
plot(model)
plot(RFsimulate(model, x, y))

[Package RandomFields version 3.3.6 Index]