RMbr2bg {RandomFields}	R Documentation

Transformation from Brown-Resnick to Bernoulli

Description

This function can be used to model a max-stable process based on a binary field, with the same extremal correlation function as a Brown-Resnick process

C_{bg}(h) = \cos(π (2Φ(√{γ(h) / 2}) -1) )

Here, Φ is the standard normal distribution function, and γ is a semi-variogram with sill

4(erf^{-1}(1/2))^2 = 2 * { Φ^{-1}( 3 / 4 ) }^2 = 1.819746 / 2 = 0.9098728

Usage

RMbr2bg(phi, var, scale, Aniso, proj)

Arguments

phi
covariance function of class RMmodel.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

RMbr2bg
binary random field RPbernoulli simulated with RMbr2bg(RMmodel()) has a uncentered covariance function that equals

the tail correlation function of the max-stable process constructed with this binary random field

the tail correlation function of Brown-Resnick process with variogram RMmodel.

Note that the reference paper is based on the notion of the (genuine) variogram, whereas the package RandomFields is based on the notion of semi-variogram. So formulae differ by factor 2.

Value

object of class RMmodel

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Strokorb, K., Ballani, F., and Schlather, M. (2014) Tail correlation functions of max-stable processes: Construction principles, recovery and diversity of some mixing max-stable processes with identical TCF. Extremes, Submitted.

See Also

maxstableAdvanced, RMbr2eg, RMmodel, RMm2r, RPbernoulli, RPbrownresnick, RPschlather.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMexp(var=1.62 / 2) 
x <- seq(0, 10, 0.05)
z <- RFsimulate(RPschlather(RMbr2eg(model)), x, x)
plot(z)

[Package RandomFields version 3.3.6 Index]