RMbrownresnick {RandomFields}	R Documentation

Tail correlation function of the Brown-Resnick process

Description

RMbrownresnick defines the tail correlation function of the Brown-Resnick process.

C(h) = 2 - 2Φ(√{γ(h)} / 2)

where φ is the standard normal distribution function and γ is the semi-variogram.

Usage

RMbrownresnick(phi, var, scale, Aniso, proj)

Arguments

phi
variogram of class RMmodel.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

For a given RMmodel the function RMbrownresnick(RMmodel()) 'returns' the tail correlation function of a Brown-Resnick process with variogram RMmodel.

Value

object of class RMmodel

Note

In the paper Kabluchko et al. (2009) the variogram instead of the semi-variogram is considered, so the formulae differ slightly.

In Version 3.0.33 a typo has been corrected.

Here, a definition is used that is consistent with the rest of the package.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Kabluchko, Z., Schlather, M. & de Haan, L (2009) Stationary max-stable random fields associated to negative definite functions Ann. Probab. 37, 2042-2065.

Strokorb, K., Ballani, F., and Schlather, M. (2014) Tail correlation functions of max-stable processes: Construction principles, recovery and diversity of some mixing max-stable processes with identical TCF. Extremes, Submitted.

See Also

RFsimulate, RMm2r, RMm3b, RMmps, RMmodel.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

#plot covariance model of type RMbrownresnick
RMmodel <- RMfbm(alpha=1.5, scale=0.2)
plot(RMbrownresnick(RMmodel))

#simulate and plot corresponding Gaussian random field
x <- seq(-5, 5, 0.05)
z <- RFsimulate(RMbrownresnick(RMmodel), x=x, y=x)
plot(z)

[Package RandomFields version 3.3.6 Index]