RMcauchy {RandomFields}	R Documentation

Cauchy Family Covariance Model

Description

RMcauchy is a stationary isotropic covariance model belonging to the Cauchy family. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(r) = (1 + r^2)^(-γ)

where γ > 0. See also RMgencauchy.

Usage

RMcauchy(gamma, var, scale, Aniso, proj)

Arguments

gamma
a numerical value; should be positive to provide a valid covariance function for a random field of any dimension.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The paramater γ determines the asymptotic power law. The smaller γ, the longer the long-range dependence. The covariance function is very regular near the origin, because its Taylor expansion only contains even terms and reaches its sill slowly.

Each covariance function of the Cauchy Family is a normal scale mixture.

The generalized Cauchy Family (see RMgencauchy) includes this family for the choice α = 2 and β = 2 γ. The generalized Hyperbolic Family (see RMhyperbolic) includes this family for the choice ξ = 0 and γ = -ν/2; in this case scale=δ.

Value

RMcauchy returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. and Schlather, M. (2004) Stochastic models which separate fractal dimension and Hurst effect. SIAM review 46, 269–282.

Gelfand, A. E., Diggle, P., Fuentes, M. and Guttorp, P. (eds.) (2010) Handbook of Spatial Statistics. Boca Raton: Chapman & Hall/CRL.

See Also

RMcauchytbm, RMgencauchy, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMcauchy(gamma=1)
x <- seq(0, 10, 0.02)
plot(model, xlim=c(-3, 3))
plot(RFsimulate(model, x=x, n=4))

[Package RandomFields version 3.3.6 Index]