RMcauchytbm {RandomFields}	R Documentation

Modifications of the Cauchy Family Covariance Model

Description

RMcauchytbm() is a shortcut of RMtbm(RMgencauchy()) and is given here for downwards compatibility.

Usage

RMcauchytbm(alpha, beta, gamma, var, scale, Aniso, proj)

Arguments

alpha,beta
See RMgencauchy.

gamma
is the same as fulldim in RMtbm.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Value

RMcauchytbm returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. and Schlather, M. (2004) Stochastic models which separate fractal dimension and Hurst effect. SIAM review 46, 269–282.

See Also

RMcauchy, RMgencauchy, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMcauchytbm(alpha=1, beta=1, gamma=3)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]