RMchoquet {RandomFields}	R Documentation

Schoenberg's representation for the classes psi_d and psi_{∞} in d=2

Description

RMchoquet is an isotropic covariance model. The corresponding covariance function only depends on the angle 0 ≤ θ ≤ π between two points on the sphere and is given for d=2 by

ψ(θ) = ∑_{n=0}^{∞} b_{n,2}/(n+1)*P_n(cos(θ)),

where

∑_{n=0}^{∞} b_{n,d}=1

and P_n is the Legendre Polynomial of integer order n >= 0.

Usage

RMchoquet(b)

Arguments

b
a numerical vector of weights in (0,1), such that sum(b)=1.

Details

By the results (cf. Gneiting, T. (2013), p.1333) of Schoenberg and others like Menegatto, Chen, Sun, Oliveira and Peron, the class psi_d of all real valued funcions on [0,π], with ψ(0)=1 and such that the associated isotropic function

h(x,y)=ψ(theta) with cos(θ)=<x,y>

for x,y in {x in R^d: ||x|| = 1}

is (strictly) positive definite is represented by this covariance model. The model can be interpreted as Choquet representation in terms of extremal members, which are non-strictly positive definite.

Special cases are the multiquadric family (see RMmultiquad) and the model of the sine power function (see RMsinepower).

Value

RMchoquet returns an object of class RMmodel.

Author(s)

Christoph Berreth; Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. (2013) Strictly and non-strictly positive definite functions on spheres. Bernoulli, 19(4), 1327-1349.

Schoenberg, I.J. (1942) Positive definite functions on spheres. Duke Math.J.,9, 96-108.

Menegatto, V.A. (1994) Strictly positive definite kernels on the Hilbert sphere. Appl. Anal., 55, 91-101.

Chen, D., Menegatto, V.A., and Sun, X. (2003) A necessary and sufficient condition for strictly positive definite functions on spheres. Proc. Amer. Math. Soc.,131, 2733-2740.

Menegatto, V.A., Oliveira, C.P. and Peron, A.P. (2006) Strictly positive definite kernels on subsets of the complex plane. Comput. Math. Appl., 51, 1233-1250.

See Also

RMmodel, RFsimulate, RFfit, spherical models, RMmultiquad, RMsinepower

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## to do

[Package RandomFields version 3.3.6 Index]