RMconstant {RandomFields}	R Documentation

Covariance Matrix Constant in Space

Description

RMconstant defines a spatially constant covariance function.

Usage

RMconstant(M, var)

Arguments

M
a numerical matrix defining the user-defined covariance for a random field; the matrix should be positive definite, symmetric and its dimension should be equal to the length of observation or simulation vector.

var
variance

Value

RMconstant returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMfixcov, RMmodel.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMconstant(diag(2),var=3)
plot(model)
x <- seq(0,10,length=100)
z <- RFsimulate(model=model,x=x)

[Package RandomFields version 3.3.6 Index]