RMcovariate {RandomFields}	R Documentation

Model for covariates

Description

The model makes covariates available.

Usage

RMcovariate(formula=NULL, data, x, y=NULL, z=NULL, T=NULL, grid,
            raw, norm, addNA, factor)

Arguments

formula, data
formula and by which the data should be modelled, similar to lm.

If formula is not given, the the linear model is given by the data themselves.

x,y,z,T,grid
optional. The usual arguments as in RFsimulate to define the locations where the covariates are given.

raw
logical. If FALSE then the data are interpolated. This approach is always save, but might be slow. If TRUE then the data may be accessed when covariance matrices are calculated. No rescaling or anisotropy definition is allowed in combination with the model. The use is dangerous, but fast.
Default: FALSE.

norm
optional model that gives the norm between locations

addNA
If addNA is TRUE, then an additional (linear) factor is estimated in an estimation framework. This parameter must be set in particular when RMcovariate passes several covariates.

factor
real value. From user's point of view very much the same as setting the argument var

.

Details

The function interpolates (nearest neighbour) between the values.

Value

RMcovariate returns an object of class RMmodel.

Note

c, x also accept lists of data. However, its use is not in an advanced stage yet.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMfixcov, RMmodel, RMtrend

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

z <- 0.2 + (1:10)
RFfctn(RMcovariate(z), 1:10)
RFfctn(RMcovariate(data=z, x=1:10), c(2, 2.1, 2.5, 3))

[Package RandomFields version 3.3.6 Index]