RMcoxisham {RandomFields}	R Documentation

Cox Isham Covariance Model

Description

RMcoxisham is a stationary covariance model which depends on a univariate stationary isotropic covariance model C_0, which is a normal scale mixture.

The corresponding covariance function only depends on the difference (h,t) between two points in d+1-dimensional space and is given by

C(h,t)=|E + t^β D|^{-1/2} C_0([(h - t μ)^T (E + t^β D)^{-1} (h - t μ)]^{1/2})

Here μ is a vector in d-dimensional space; E is the d x d-identity matrix and D is a d x d-correlation matrix with |D| > 0. The parameter β is in (0,2]. Currently, the implementation is done only for d=2.

Usage

RMcoxisham(phi,mu,D,beta,var, scale, Aniso, proj)

Arguments

phi
a univariate stationary isotropic covariance model for random fields on d-dimensional space, which is moreover a normal scale mixture, that means an RMmodel whose monotone property equals 'normal mixture', see
RFgetModelNames(monotone="normal mixture")
and whose maxdim is at least 2.

mu
a vector in d-dimensional space

D
a d x d-correlation matrix with |D| > 0

beta
numeric in the interval (0,2]; default value is 2

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

This model stems from a rainfall model (cf. Cox, D.R., Isham, V.S. (1988)) and equals the following expectation

C(h,t)=\bold{E}_V C_0(h-Vt)

where the random wind speed vector V follows a d-variate normal distribution with expectation mu and covariance matrix D/2 (cf. Schlather, M. (2010), Example 9).

Value

RMcoxisham returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Cox, D.R., Isham, V.S. (1988) A simple spatial-temporal model of rainfall. Proc. R. Soc. Lond. A, 415, 317-328.

Schlather, M. (2010) On some covariance models based on normal scale mixtures. Bernoulli, 16, 780-797.

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMcoxisham(RMgauss(), mu=1, D=1)
x <- seq(0, 10, 0.3)
plot(model, dim=2)
plot(RFsimulate(model, x=x, y=x))

[Package RandomFields version 3.3.6 Index]