RMcurlfree {RandomFields}	R Documentation

Curlfree Covariance Model

Description

RMcurlfree is a multivariate covariance model which depends on a univariate stationary covariance model where the covariance function phi(h) is twice differentiable.

The corresponding matrix-valued covariance function C of the model only depends on the difference h between two points and it is given by the following components:

the potential

the vector field given by

C(h)=( - \nabla_h (\nabla_h)^T ) C_0(h)

the field of sinks and sources

Usage

RMcurlfree(phi, which, var, scale, Aniso, proj)

Arguments

phi
a univariate stationary covariance model (2- or 3-dimensional).

which
vector of integers. If not given all components are returned; otherwise the selected components are returned.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The model returns the potential field in the first component, the corresponding curlfree field and field of sources and sinks in the last component.

See also the models RMdivfree and RMvector.

Value

RMcurlfree returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Scheuerer, M. and Schlather, M. (2012) Covariance Models for Divergence-Free and Curl-Free Random Vector Fields. Stochastic Models 28:3.

See Also

RMderiv, RMdivfree, RMvector, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMcurlfree(RMgauss(), scale=4)
plot(model, dim=2)

x.seq <- y.seq <- seq(-10, 10, 0.2)
simulated <- RFsimulate(model=model, x=x.seq, y=y.seq)
plot(simulated, select.variables=list(1, c(1, 2:3), 4))

[Package RandomFields version 3.3.6 Index]