RMcutoff {RandomFields}	R Documentation

Gneiting's modification towards finite range

Description

RMcutoff is a functional on univariate stationary isotropic covariance functions phi.

The corresponding function C (which is not necessarily a covariance function, see details) only depends on the distance r between two points in d-dimensional space and is given by

C(r)=φ(r), 0≤ r ≤ d

C(r) = b_0 ((dR)^a - r^a)^{2 a}, d ≤ r ≤ dR

C(r) = 0, dR ≤ r

The parameters R and b_0 are chosen internally such that C is a smooth function.

Usage

RMcutoff(phi, diameter, a, var, scale, Aniso, proj)

Arguments

phi
a univariate stationary isotropic covariance model. See, for instance,

RFgetModelNames(type="positive definite", domain="single variable", isotropy="isotropic", vdim=1).

diameter
a numerical value; should be greater than 0; the diameter of the domain on which the simulation is done

a
a numerical value; should be greater than 0; has been shown to be optimal for a = 1/2 or a =1.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The algorithm that checks the given parameters knows only about some few necessary conditions. Hence it is not ensured that the cutoff-model is a valid covariance function for any choice of φ and the parameters.

For certain models phi, e.g. RMstable, RMwhittle and RMgencauchy, some sufficient conditions are known (cf. Gneiting et al. (2006)).

Value

RMcutoff returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T., Sevecikova, H, Percival, D.B., Schlather M., Jiang Y. (2006) Fast and Exact Simulation of Large Gaussian Lattice Systems in $R^2$: Exploring the Limits. J. Comput. Graph. Stat. 15, 483–501.

Stein, M.L. (2002) Fast and exact simulation of fractional Brownian surfaces. J. Comput. Graph. Statist. 11, 587–599

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMexp()
plot(model, model.cutoff=RMcutoff(model, diameter=1), xlim=c(0, 4))

model <- RMstable(alpha = 0.8)
plot(model, model.cutoff=RMcutoff(model, diameter=2), xlim=c(0, 5))
x <- y <- seq(0, 4, 0.05)
plot(RFsimulate(RMcutoff(model), x=x, y = y))

[Package RandomFields version 3.3.6 Index]