RMdagum {RandomFields}	R Documentation

Dagum Covariance Model Family

Description

RMdagum is a stationary isotropic covariance model. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(r) = 1-(1+r^(-β))^(-γ / β).

The parameters β and γ can be varied in the intervals (0,1] and (0,1), respectively.

Usage

RMdagum(beta, gamma, var, scale, Aniso, proj)

Arguments

beta
numeric in (0,1]

gamma
numeric in (0,1)

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

Like the generalized Cauchy model the Dagum family can be used to model fractal dimension and Hurst effect. For a comparison of these see Berg, C. and Mateau, J. and Porcu, E. (2008). This paper also establishes valid parameter choices for the Dagum family, but be careful because therein the model is parameterized differently.

Value

RMdagum returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Berg, C. and Mateau, J. and Porcu, E. (2008) The dagum family of isotropic correlation functions. Bernoulli 14(4), 1134–1149.

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMdagum(beta=0.5, gamma=0.5, scale=0.2)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]