RMdampedcos {RandomFields}	R Documentation

Exponentially Damped Cosine

Description

RMdampedcos is a stationary isotropic covariance model. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(r) = exp(-λ r) cos(r).

Usage

RMdampedcos(lambda, var, scale, Aniso, proj)

Arguments

lambda
numeric. The range depends on the dimension of the random field (see details).

var, scale, Aniso, proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The model is valid for any dimension d. However, depending on the dimension of the random field the following bound for the argument λ has to be respected:

λ ≥ 1/{tan(π/(2d))}.

This covariance models a hole effect (cf. Chiles, J.-P. and Delfiner, P. (1999), p. 92).

For λ = 0 we obtain the covariance function

C(r)=cos(r)

which is only valid for d=1 and corresponds to RMbessel for ν=-0.5, there.

Value

RMdampedcos returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Chiles, J.-P. and Delfiner, P. (1999) Geostatistics. Modeling Spatial Uncertainty. New York: Wiley.

Gelfand, A. E., Diggle, P., Fuentes, M. and Guttorp, P. (eds.) (2010) Handbook of Spatial Statistics. Boca Raton: Chapman & Hall/CRL.

See Also

RMbessel, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMdampedcos(lambda=0.3, scale=0.1)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]