RMdelay {RandomFields}	R Documentation

Bivariate Delay Effect

Description

RMdelay is a (m+1)-variate stationary covariance model. which depends on a univariate stationary covariance model C_0.

The corresponding covariance function only depends on the difference h between two points in d-dimensional space and is given by

C(h)=(C_0(h - s_i +s_j))_{i,j=0,…,m}

where h \in R^{d x m} and s_0=0

Usage

RMdelay(phi,s,var, scale, Aniso, proj)

Arguments

phi
a univariate stationary covariance model, that means an RMmodel whose vdim equals 1.

s
a d x m-dimensional shift matrix, where d is the dimension of the space, giving the components s=(s_1,..., s_m) where the s_i are vectors.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

Here, a multivariate random field is obtained from a single univariate random field by shifting it by a fixed value.

Value

RMdelay returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Schlather, M., Malinowski, A., Menck, P.J., Oesting, M. and Strokorb, K. (2015) Analysis, simulation and prediction of multivariate random fields with package RandomFields. Journal of Statistical Software, 63 (8), 1-25, url = ‘http://www.jstatsoft.org/v63/i08/’

Wackernagel, H. (2003) Multivariate Geostatistics. Berlin: Springer, 3nd edition.

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

x <- y <- seq(-10,10,0.2)
model <- RMdelay(RMstable(alpha=1.9, scale=2), s=c(4,4))
plot(model, dim=2, xlim=c(-6, 6), ylim=c(-6,6))

simu <- RFsimulate(model, x, y)
plot(simu, zlim="joint")

[Package RandomFields version 3.3.6 Index]