RMdewijsian {RandomFields}	R Documentation

Modified De Wijsian Variogram Model

Description

The modified RMdewijsian model is an intrinsically stationary isotropic variogram model. The corresponding centered semi-variogram only depends on the distance r ≥ 0 between two points and is given by

γ(r)=log(r^{α}+1)

where 0 < α ≤ 2.

Usage

RMdewijsian(alpha, var, scale, Aniso, proj)

Arguments

alpha
a numerical value; in the interval (0,2].

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above variogram remains unmodified.

Details

Originally, the logarithmic model γ(r) = \log(r) was named after de Wijs and reflects a principle of similarity (cf. Chiles, J.-P. and Delfiner, P. (1999), p. 90). But note that γ(r) = \log(r) is not a valid variogram (γ(0) does not vanish) and can only be understood as a characteristic of a generalized random field.

The modified RMdewijsian model γ(r) = \log(r^{α}+1) is a valid variogram model (cf. Wackernagel, H. (2003), p. 336).

Value

RMdewijsian returns an object of class RMmodel.

Note

Note that the (non-modified) de Wijsian model equals γ(r) = \log(r).

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Wackernagel, H. (2003) Multivariate Geostatistics. Berlin: Springer, 3nd edition.

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
model <- RMdewijsian(alpha=1)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]