RMeaxxa {RandomFields}	R Documentation

Special models for rotation like fields

Description

RMeaxxa and RMetaxxa define the auxiliary functions

C(h) = h^T A A^T h + diag(E)

and

C(h) = h^T A R R^R A^T h + diag(E),

respectively.

Usage

RMeaxxa(E, A) 
RMetaxxa(E, A, alpha)

Arguments

E
m-variate vector of positive values

A
m x k matrix

alpha
angle for the rotation matrix R

Details

RMeaxxa is defined in space and returns an m-variate model.

RMetaxxa is a space-time model with two spatial dimensions. The matrix R is a rotation matrix with angle β t where t is the time component.

Value

RMeaxxa and RMetaxxa return an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Schlather, M. (2010) On some covariance models based on normal scale mixtures. Bernoulli, 16, 780-797.

See Also

RMmodel, S10

Examples

# see S10

[Package RandomFields version 3.3.6 Index]