RMepscauchy {RandomFields}	R Documentation

Generalized Cauchy Family Covariance Model

Description

RMepscauchy is a stationary isotropic covariance model belonging to the generalized Cauchy family. In contrast to most other models it is not a correlation function. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(r) = (ε + r^α)^(-β/α)

where ε > 0, 0 < α ≤ 2 and β > 0. See also RMcauchy.

Usage

RMepscauchy(alpha, beta, eps, var, scale, Aniso, proj)

Arguments

alpha
a numerical value; should be in the interval (0,2] to provide a valid covariance function for a random field of any dimension.

beta
a numerical value; should be positive to provide a valid covariance function for a random field of any dimension.

eps
a positive value

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

This model has a smoothness parameter α and a parameter β which determines the asymptotic power law. More precisely, this model admits simulating random fields where fractal dimension D of the Gaussian sample and Hurst coefficient H can be chosen independently (compare also RMlgd): Here, we have

D = d + 1 - α/2, 0 < α ≤ 2

and

H = 1 - β/2, β > 0.

I. e. the smaller β, the longer the long-range dependence.

The covariance function is very regular near the origin, because its Taylor expansion only contains even terms and reaches its sill slowly.

Each covariance function of the Cauchy family is a normal scale mixture.

Note that the Cauchy Family (see RMcauchy) is included in this family for the choice α = 2 and β = 2 γ.

Value

RMepscauchy returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. and Schlather, M. (2004) Stochastic models which separate fractal dimension and Hurst effect. SIAM review 46, 269–282.

See Also

RMcauchy, RMcauchytbm, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMepscauchy(alpha=1.5, beta=1.5, scale=0.3, eps=0.5)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]