RMexp {RandomFields}	R Documentation

Exponential Covariance Model

Description

RMexp is a stationary isotropic covariance model whose corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(r) = exp(-r).

Usage

RMexp(var, scale, Aniso, proj)

Arguments

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

This model is a special case of the Whittle covariance model (see RMwhittle) if ν=0.5 and of the symmetric stable family (see RMstable) if ν=1. Moreover, it is the continuous-time analogue of the first order autoregressive time series covariance structure.

The exponential covariance function is a normal scale mixture.

Value

RMexp returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Covariance model

Gelfand, A. E., Diggle, P., Fuentes, M. and Guttorp, P. (eds.) (2010) Handbook of Spatial Statistics. Boca Raton: Chapman & Hall/CRL.

Tail correlation function

Strokorb, K., Ballani, F., and Schlather, M. (2014) Tail correlation functions of max-stable processes: Construction principles, recovery and diversity of some mixing max-stable processes with identical TCF. Extremes, Submitted.

See Also

RMwhittle, RMstable, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMexp()
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]