RMexponential {RandomFields}	R Documentation

Exponential operator

Description

RMexponential yields a covariance model from a given variogram or covariance model. The covariance C is given as

C(h) = (\exp(φ(h)) -∑_{k=0}^n φ^k(h)/k!) / (\exp(φ(0)) -∑_{k=0}^n φ^k(0)/k!)

if φ is a covariance model, and as

C(h) = \exp(-φ(h))

if φ is a variogram model.

Usage

RMexponential(phi, n, standardised, var, scale, Aniso, proj)

Arguments

phi
a valid RMmodel; either a variogram model or a covariance model

n
integer, see formula above. Default is -1; if the multivariate dimension of the submodel is greater than 1 then only the default value is valid.

standardised
logical. If TRUE then the above formula holds. If FALSE then only the nominator of the above formula is returned. Default value is TRUE.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

If γ is a variogram, then \exp(-γ) is a valid covariance.

Value

RMexponential returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

See, for instance,

Berg, C., Christensen, J. P. R., Ressel, P. (1984) Harmonic Analysis on Semigroups. Theory of Positive Definite and Related Functions. Springer, New York.

Sasvari, Z. (2013) Multivariate Characteristic and Correlation Functions. de Gruyter, Berlin.

Schlather, M. (2010) Some covariance models based on normal scale mixtures, Bernoulli 16, 780-797.

Schlather, M. (2012) Construction of covariance functions and unconditional simulation of random fields. In Porcu, E., Montero, J. M., Schlather, M. Advances and Challenges in Space-time Modelling of Natural Events, Springer, New York.

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
model <- RMexponential(RMfbm(alpha=1))  ## identical to RMexp()
plot(RMexp(), model=model, type=c("p", "l"), pch=20) 

[Package RandomFields version 3.3.6 Index]