RMfbm {RandomFields}	R Documentation

Variogram Model of Fractal Brownian Motion

Description

RMfbm is an intrinsically stationary isotropic variogram model. The corresponding centered semi-variogram only depends on the distance r ≥ 0 between two points and is given by

γ(r) = r^α

where 0 < α ≤ 2.
By now, the model is implemented for dimensions up to 3.
For a generalized model see also RMgenfbm.

Usage

RMfbm(alpha, var, scale, Aniso, proj)

Arguments

alpha
numeric in (0,2]; refers to the fractal dimension of the process

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above variogram remains unmodified.

Details

The variogram is unbounded and belongs to a non-stationary process with stationary increments. For α=1 and scale=2 we get a variogram corresponding to a standard Brownian Motion.

For 0 < α < 2 the quantity H=α/2 is called Hurst index and determines the fractal dimension D of the corresponding Gaussian sample paths

D = d + 1 - H

where d is the dimension of the random field (see Chiles and Delfiner, 1999, p. 89).

Value

RMfbm returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Chiles, J.-P. and P. Delfiner (1999) Geostatistics. Modeling Spatial Uncertainty. New York, Chichester: John Wiley & Sons.

Stein, M.L. (2002) Fast and exact simulation of fractional Brownian surfaces. J. Comput. Graph. Statist. 11, 587–599.

See Also

RMgenfbm, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMfbm(alpha=1)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]