RMfixcov {RandomFields}	R Documentation

Fixed Covariance Matrix

Description

RMfixcov is a user-defined covariance according to the given covariance matrix.

It extends to the space through a Voronoi tessellation.

Usage

RMfixcov(M, x, y=NULL, z=NULL, T=NULL, grid, var, proj, raw, norm)

Arguments

M
a numerical matrix defining the user-defined covariance for a random field; the matrix should be positive definite, symmetric and its dimension should be equal to the length of observation or simulation vector.

x,y,z,T,grid
optional. The usual arguments as in RFsimulate to define the locations where the covariates are given.

var,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

raw
logical. If FALSE then the data are interpolated. This approach is always save, but might be slow. If TRUE then the data may be accessed when covariance matrices are calculated. No rescaling or anisotropy definition is allowed in combination with the model. The use is dangerous, but fast.
Default: FALSE.

norm
optional model that gives the norm between locations

Details

The covariances passed are implemented for the given locations. Within any Voronoi cell (around a given location) the correlation is assumed to be one.

In particular, it is used in RFfit to define neighbour or network structure in the data.

Value

RMfixcov returns an object of class RMmodel.

Note

Starting with version 3.0.64, the former argument element is replaced by the general option set in RFoptions.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Ober, U., Ayroles, J.F., Stone, E.A., Richards, S., Zhu, D., Gibbs, R.A., Stricker, C., Gianola, D., Schlather, M., Mackay, T.F.C., Simianer, H. (2012): Using Whole Genome Sequence Data to Predict Quantitative Trait Phenotypes in Drosophila melanogaster. PLoS Genet 8(5): e1002685.

See Also

RMcovariate, RMmodel, RFsimulate, RFfit, RMuser.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## Example 1 showing that the covariance structure is correctly implemented
n <- 10
C <- matrix(runif(n^2), nc=n)
(C <- C %*% t(C))
RFcovmatrix(RMfixcov(C), 1:n)

## Example 2 showing that the covariance structure is interpolated
RFcovmatrix(RMfixcov(C, 1:n), c(2, 2.1, 2.5, 3))

## Example 3 showing the use in a separable space-time model
model <- RMfixcov(C, 1:n, proj="space") * RMexp(s=40, proj="time")
(z <- RFsimulate(model, x = seq(0,12, 0.5), T=1:100))
plot(z)

[Package RandomFields version 3.3.6 Index]