RMfixed {RandomFields}	R Documentation

Fixed Effect Model

Description

Expressions of the form X@RMfixed(beta) can be used within a formula of the type

response ~ fixed effects + random effects + error term

that specifies the Linear Mixed Model.

Important remark: RMfixed is NOT a function although the parentheses notation is used to specify the vector of coefficients.

The matrix X is the design matrix and β is a vector of coefficients.

Note that a fixed effect of the form X is interpreted as X@RMfixed(beta=NA) by default (and β is estimated provided that the formula is used in RFfit). Note that the 1 in an expression 1@RMfixed(beta) is interpreted as the identity matrix.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmodel, RFformula, RFsimulate.

Examples

 ## For examples see the help page of 'RFformula'. ##

[Package RandomFields version 3.3.6 Index]