RMfractdiff {RandomFields}	R Documentation

Fractionally Differenced Process Model

Description

RMfractdiff is a stationary isotropic covariance model. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given for integers r by

C(r) = (-1)^r Γ(1-a/2)^2 / (Γ(1-a/2+r) Γ(1-a/2-r))

and otherwise linearly interpolated. Here, -1 ≤ a < 1, Γ denotes the gamma function. It can only be used for one-dimensional random fields.

Usage

RMfractdiff(a, var, scale, Aniso, proj)

Arguments

a
-1 ≤ a < 1

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The model is only valid for dimension d = 1 . It stems from time series modelling where the grid locations are multiples of the scale parameter.

Value

RMfractdiff returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMfractdiff(0.5, scale=0.2)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]