RMfractgauss {RandomFields}	R Documentation

Fractal Gaussian Model Family

Description

RMfractgauss is a stationary isotropic covariance model. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(r) = 0.5 ((r+1)^(α)-2r^(α)+|r-1|^(α))

with 0 < α ≤ 2. It can only be used for one-dimensional random fields.

Usage

RMfractgauss(alpha,var, scale, Aniso, proj)

Arguments

alpha
0 < α ≤ 2

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The model is only valid for dimension d = 1 . It is the covariance function for the fractional Gaussian noise with self-affinity index (Hurst parameter) H=α /2 with 0 < α ≤ 2.

Value

RMfractgauss returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. and Schlather, M. (2004) Stochastic models which separate fractal dimension and Hurst effect. SIAM review 46, 269–282.

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMfractgauss(alpha=0.5, scale=0.2)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]