RMgauss {RandomFields}	R Documentation

Gaussian Covariance Model

Description

RMgauss is a stationary isotropic covariance model. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(r)=e^{-r^2}.

Usage

RMgauss(var, scale, Aniso, proj)

Arguments

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

This model is called Gaussian because of the functional similarity of the spectral density of a process with that covariance function to the Gaussian probability density function.

The Gaussian model has an infinitely differentiable covariance function. This smoothness is artificial. Furthermore, this often leads to singular matrices and therefore numerically instable procedures (cf. Stein, M. L. (1999), p. 29).

The Gaussian model is included in the symmetric stable class (see RMstable) for the choice alpha = 2.

Value

RMgauss returns an object of class RMmodel.

Note

The use of RMgauss is questionable from both a theoretical (analytical paths) and a practical point of view (e.g. speed of algorithms). Instead, RMgneiting should be used.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gelfand, A. E., Diggle, P., Fuentes, M. and Guttorp, P. (eds.) (2010) Handbook of Spatial Statistics. Boca Raton: Chapman & Hall/CRL.

Stein, M. L. (1999) Interpolation of Spatial Data. New York: Springer-Verlag

See Also

RMstable and RMmatern for generalizations;
RMmodel, RFsimulate, RFfit.

Do not mix up with RPgauss or RRgauss.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
model <- RMgauss(scale=0.4)
x <- seq(0, 10, 0.02)
plot(model)
lines(RMgauss(), col="red")
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]