RMgengneiting {RandomFields}	R Documentation

Gneiting-Wendland Covariance Models

Description

RMgengneiting is a stationary isotropic covariance model family whose elements are specified by the two parameters κ and μ with n being a non-negative integer and μ ≥ d/2 with d denoting the dimension of the random field (the models can be used for any dimension). A corresponding covariance function only depends on the distance r ≥ 0 between two points. For the case κ = 0 the Gneiting-Wendland model equals the Askey model RMaskey,

C(r) = (1-r)^β 1_{[0,1]}(r), β = μ + 1/2 = μ + 2κ + 1/2.

For κ = 1 the Gneiting model is given by

C(r) = (1+β r)(1-r)^β 1_{[0,1]}(r), β = μ + 2κ + 1/2.

If κ = 2

C(r) = (1 + β r + (β^2 - 1) r^2 / 3) (1 - r)^β 1_{[0,1]}(r), β = μ + 2κ + 1/2.

In the case κ = 3

C(r) = (1 + β r + (2 β^2 - 3 )r^2 / 5 + (β^2 - 4) β r^3 / 15)(1-r)^β 1_{[0,1]}(r), β = μ + 2κ + 1/2.

A special case of this model is RMgneiting.

Usage

RMgengneiting(kappa, mu, var, scale, Aniso, proj)

Arguments

kappa
0,...,3

; it chooses between the three different covariance models above.

mu
mu has to be greater than or equal to d / 2 where d is the dimension of the random field.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

This isotropic family of covariance functions is valid for any dimension of the random field.

A special case of this family is RMgneiting (with s = 1 there) for the choice κ = 3, μ = 3/2.

Value

RMgengneiting returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. (1999) Correlation functions for atmospherical data analysis. Q. J. Roy. Meteor. Soc Part A 125, 2449-2464.

Wendland, H. (2005) Scattered Data Approximation. Cambridge Monogr. Appl. Comput. Math.

See Also

RMaskey, RMbigneiting, RMgneiting, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMgengneiting(kappa=1, mu=1.5)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

## same models:
model2 <- RMgengneiting(kappa=3, mu=1.5, scale= 1 / 0.301187465825)
plot(RMgneiting(), model2=model2, type=c("p", "l"), pch=20)

[Package RandomFields version 3.3.6 Index]