RMgennsst {RandomFields}	R Documentation

Non-Separable Space-Time model

Description

RMgennsst is a univariate stationary spaceisotropic covariance model on R^d whose corresponding covariance is given by

C(h,u)= φ( sqrt(h^\top ψ^{-1}(u) h)) / √(det(psi))

Usage

RMgennsst(phi, psi, dim_u, var, scale, Aniso, proj)

Arguments

phi
is a normal mixture RMmodel, cf.
RFgetModelNames(monotone="normal mixture").

psi
is a d-variate variogram model RMmodel.

dim_u
the dimension of the component u

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

This model is used for space-time modelling where the spatial component is isotropic.

Value

RMgennsst returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Schlather, M. (2010) On some covariance models based on normal scale mixtures. Bernoulli, 16, 780-797.

See Also

RMnsst, RMmodel, RFsimulate, RFfit.

Examples

[Package RandomFields version 3.3.6 Index]