RMgneitingdiff {RandomFields}	R Documentation

Gneiting Covariance Model Used as Tapering Function

Description

RMgneitingdiff is a stationary isotropic covariance model which is only valid up to dimension 3. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(h) = C_0(h / t) W_ν(h / s)

where C_0 is Gneiting's model RMgneiting and W_ν is the Whittle model RMwhittle.

Usage

RMgneitingdiff(nu, taper.scale, scale, var, Aniso, proj)

Arguments

nu
see RMwhittle

taper.scale
is the parameter t in the above formula.

scale
is the parameter s in the above formula.

var,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The model allows to a certain degree the smooth modelling of the differentiability of a covariance function with compact support.

Value

RMgneitingdiff returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. (1999) Correlation functions for atmospherical data analysis. Q. J. Roy. Meteor. Soc Part A 125, 2449-2464.

See Also

RMbigneiting, RMgneiting, RMgengneiting, RMgauss, RMmodel, RMwhittle, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
model <- RMgneitingdiff(nu=2, taper.scale=1, scale=0.2)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]