RMhyperbolic {RandomFields}	R Documentation

Generalized Hyperbolic Covariance Model

Description

RMhyperbolic is a stationary isotropic covariance model called “generalized hyperbolic”. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(r) = δ^(-ν) (K_ν(ν δ))^{-1} (δ^2+r^2)^{ν/2} K_ν(ξ(δ^2+r^2)^{1/2})

where K_ν denotes the modified Bessel function of second kind.

Usage

RMhyperbolic(nu, lambda, delta, var, scale, Aniso, proj)

Arguments

nu, lambda, delta
numerical values; should either satisfy
δ ≥ 0, λ > 0 and ν > 0, or
δ > 0, λ > 0 and ν = 0, or
δ > 0, λ ≥ 0 and ν < 0.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

This class is over-parametrized, i.e. it can be reparametrized by replacing the three parameters λ, δ and scale by two other parameters. This means that the representation is not unique.

Each generalized hyperbolic covariance function is a normal scale mixture.

The model contains some other classes as special cases; for λ = 0 we get the Cauchy covariance function (see RMcauchy) with γ = -ν/2 and scale=δ; the choice δ = 0 yields a covariance model of type RMwhittle with smoothness parameter ν and scale parameter 1/λ.

Value

RMhyperbolic returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Shkarofsky, I.P. (1968) Generalized turbulence space-correlation and wave-number spectrum-function pairs. Can. J. Phys. 46, 2133-2153.

Barndorff-Nielsen, O. (1978) Hyperbolic distributions and distributions on hyperbolae. Scand. J. Statist. 5, 151-157.

Gneiting, T. (1997). Normal scale mixtures and dual probability densities. J. Stat. Comput. Simul. 59, 375-384.

See Also

RMcauchy, RMwhittle, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMhyperbolic(nu=1, lambda=2, delta=0.2)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]