RMintern {RandomFields}	R Documentation

Internal models

Description

Internal models or model names that may appear in feedbacks from 'RandomFields'. Those endings by ‘Intern’ should appear only in very rare cases.

Details

The following and many more internal models exist:

RF__Name__ : internal representation of certain functions RF__name__

RO# : model for transforming coordinates within the cartesian system

RO> : model for transforming earth coordinates to cartesian coordinates

ROmissing : for error messages only

RMmixed : internal representation of a mixed model

RMselect : will be obsolete in future

RMsetparam, RMptsGivenShape, RMstandardShape, RMstatShape : for max-stable processes and Poisson processes: models that combine shape functions with corresponding point processes

RP__name__Intern : internal representations of some processes

RPS, RPplusp, etc. : specific processes for RMS and RMplus etc. (For those covariance models that have specific simulation processes programmed.)

RMS : internal representation of the modifying arguments var, scale, Aniso, proj

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## in the following 'RPplus' appears as internal model
x <- seq(0, 10, 1) 
z <- RFsimulate(RPspecific(RMexp() + RMnugget()), x)
RFgetModelInfo(which="internal", level=0)

[Package RandomFields version 3.3.6 Index]