RMintexp {RandomFields}	R Documentation

Integral exponential operator

Description

RMintexp is a univariate stationary covariance model depending on a univariate variogram model phi. The corresponding covariance function only depends on the difference h between two points and is given by

C(h)=(1 - exp(-phi(h)))/phi(h)

Usage

RMintexp(phi, var, scale, Aniso, proj)

Arguments

phi
a variogram RMmodel

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Value

RMintexp returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Schlather, M. (2012) Construction of covariance functions and unconditional simulation of random fields. Lecture Notes in Statistics, Proceedings, 207, 25–54.

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMintexp(RMfbm(alpha=1.5, scale=0.2))
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]