RMintrinsic {RandomFields}	R Documentation

Intrinsic Embedding Covariance Model

Description

RMintrinsic is a univariate stationary isotropic covariance model which depends on a univariate stationary isotropic covariance model.

The corresponding covariance function C of the model only depends on the distance r ≥ 0 between two points and is given by

C(r)=a_0 + a_2 r^2 + φ(r), 0≤ r ≤ diameter

C(r)=b_0 (rawR D - r)^3/(r), diameter ≤ r ≤ rawR * diameter

C(r) = 0, rawR * diameter ≤ r

Usage

RMintrinsic(phi, diameter, rawR, var, scale, Aniso, proj)

Arguments

phi
an RMmodel; has to be stationary and isotropic

diameter
a numerical value; positive; should be the diameter of the domain on which simulation is done

rawR
a numerical value; greater or equal to 1

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The parameters a_0, a_2 and b_0 are chosen internally such that C becomes a smooth function. See formulas (3.8)-(3.10) in Gneiting et alii (2006). This model corresponds to the method Intrinsic Embedding. See also RPintrinsic.

NOTE: The algorithm that checks the given parameters knows only about some few necessary conditions. Hence it is not ensured that the Stein-model is a valid covariance function for any choice of φ and the parameters.

For certain models phi, i.e. stable, whittle, gencauchy, and the variogram model fractalB some sufficient conditions are known.

Value

RMintrinsic returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T., Sevecikova, H, Percival, D.B., Schlather M., Jiang Y. (2006) Fast and Exact Simulation of Large Gaussian Lattice Systems in $R^2$: Exploring the Limits. J. Comput. Graph. Stat. 15, 483–501.

Stein, M.L. (2002) Fast and exact simulation of fractional Brownian surfaces. J. Comput. Graph. Statist. 11, 587–599

See Also

RPintrinsic, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

x.max <- 10
model <- RMintrinsic(RMfbm(alpha=1), diameter=x.max)
x <- seq(0, x.max, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]