RMlgd {RandomFields}	R Documentation

Local-Global Distinguisher Family Covariance Model

Description

RMlgd is a stationary isotropic covariance model, which is valid only for dimensions d =1,2. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(r) =1 - β^(-1)(α + β)r^(α) 1_{[0,1]}(r) + α^(-1)(α + β)r^(-β) 1_{r>1}(r)

where β >0 and 0 < α ≤ (3-d)/2, with d denoting the dimension of the random field.

Usage

RMlgd(alpha, beta, var, scale, Aniso, proj)

Arguments

alpha
argument whose range depends on the dimension of the random field: 0< α ≤ (3-d)/2.

beta
positive number

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The model is only valid for dimension d=1,2.

This model admits simulating random fields where fractal dimension D of the Gaussian sample and Hurst coefficient H can be chosen independently (compare also RMgencauchy):

Here, the random field has fractal dimension

D = d+1 - α/2

and Hurst coefficient

H = 1-β/2

for 0< β ≤ 1.

Value

RMlgd returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. and Schlather, M. (2004) Stochastic models which separate fractal dimension and Hurst effect. SIAM review 46, 269–282.

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMlgd(alpha=0.7, beta=4, scale=0.5)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]