RMlsfbm {RandomFields}	R Documentation

Locally Positive Definite Function Given by the Fractal Brownian Motion

Description

RMlsfbm is a positive definite function on the unit ball in R^d centred at the origin,

C(r) = c - r^α

with 0 <= r = || x - y || <= 1.

Usage

RMlsfbm(alpha, const, var, scale, Aniso, proj)

Arguments

alpha
numeric in (0,2); refers to the fractal dimension of the process.

const
the constant c is given by the formula

c = 2^{-α} Γ(d / 2 + α/2) Γ(1 - α/2) / Γ(d / 2)

and should not be changed by the user in order to ensure positive definiteness.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Value

RMlsfbm returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Martini, J., Schlather, M., Simianer, H. (In preparation.)

See Also

RMbcw generalizes RMlsfbm in case that c is given, RMfbm, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMlsfbm(alpha=1, scale=10)
x <- seq(0, 10, 0.02)
plot(model, xlim=c(0,10))
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]