RMmastein {RandomFields}	R Documentation

Ma-Stein operator

Description

RMmastein is a univariate stationary covariance model depending on a variogram or covariance model on the real axis. The corresponding covariance function only depends on the difference h between two points and is given by

C(h, t)= [ Gamma(nu + phi(t))Gamma(nu + delta) ] / [Gamma(nu + phi(t) + delta) Gamma(nu) ] W_{nu + phi(t)}(|h - Vt|)

if φ is a variogram model. It is given by

C(h, t)= [ Gamma(nu + phi(0)-phi(t))Gamma(nu + delta) ] / [Gamma(nu + phi(0)-phi(t) + delta) Gamma(nu) ] W_{nu + phi(0)-phi(t)}(|h - Vt|)

if φ is a covariance model.

Here Γ is the Gamma function; W is the Whittle-Matern model (RMwhittle).

Usage

RMmastein(phi, nu, delta, var, scale, Aniso, proj)

Arguments

phi
an RMmodel on the real axis

nu
numerical value; positive; smoothness parameter of the Whittle-Matern model (for t=0)

delta
a numerical value; δ must be greater than or equal to half the dimension of h

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

See Stein (2005), formula (12).

Instead of the velocity parameter V in the original model description, a preceding anisotropy matrix is chosen appropriately:

matrix(c(A, -V, 0, 1), nr=2, by=TRUE)

A is a spatial transformation matrix. (I.e. (x,t) is multiplied from the left on the above matrix and the first elements of the obtained vector are interpreted as new spatial components and only these components are used to form the argument in the Whittle-Matern function.) The last component in the new coordinates is the time which is passed to phi. (Velocity is assumed to be zero in the new coordinates.)

Note, that for numerical reasons, ν+φ+d may not exceed the value 80.0. If exceeded the algorithm fails.

Value

RMmastein returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Ma, C. (2003) Spatio-temporal covariance functions generated by mixtures. Math. Geol., 34, 965-975.

Stein, M.L. (2005) Space-time covariance functions. JASA, 100, 310-321.

See Also

RMwhittle, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make plotthem all random again
model <- RMmastein(RMgauss(), nu=1, delta=10)
plot(RMexp(), model.mastein=model, dim=2)

x <- seq(0, 10, 0.1)
plot(RFsimulate(model, x=x, y=x))

[Package RandomFields version 3.3.6 Index]