RMwhittlematern {RandomFields}	R Documentation

Whittle-Matern Covariance Model

Description

RMmatern is a stationary isotropic covariance model belonging to the Matern family. The corresponding covariance function only depends on the distance r ≥ 0 between two points.

The Whittle model is given by

C(r)=W_{ν}(r)=2^{1- ν} Γ(ν)^{-1}r^{ν}K_{ν}(r)

where ν > 0 and K_ν is the modified Bessel function of second kind.

The Matern model is given by

C(r) = 2^{1- ν} Γ(ν)^{-1} (√{2ν} r)^ν K_ν(√{2ν} r)

The Handcock-Wallis parametrisation is given by

C(r) = 2^{1- ν} Γ(ν)^{-1} (2√{ν} r)^ν K_ν(2√{ν} r)

Usage

RMwhittle(nu, notinvnu, var, scale, Aniso, proj)

RMmatern(nu, notinvnu, var, scale, Aniso, proj)

RMhandcock(nu, notinvnu, var, scale, Aniso, proj)

Arguments

nu
a numerical value called “smoothness parameter”; should be greater than 0.

notinvnu
logical. If FALSE then in the definition of the models ν is replaced by 1/ν. This parametrization seems to be more natural. Default is, however, TRUE according with the definitions in literature.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The three models are alternative parametrizations of the same covariance function. The Matern model or the Handcock-Wallis parametrisation should be preferred as they seperate the effects of the scaling parameter and the shape parameter.

The Whittle-Matern model is the model of choice if the smoothness of a random field is to be parametrized: the sample paths of a Gaussian random field with this covariance structure are m times differentiable if and only if ν > m (see Gelfand et al., 2010, p. 24).

Furthermore, the fractal dimension (see also RFfractaldim) D of the Gaussian sample paths is determined by ν: We have

D = d + 1 - ν, 0 < ν < 1

and D = d for ν > 1 where d is the dimension of the random field (see Stein, 1999, p. 32).

If ν=0.5 the Matern model equals RMexp.

For ν tending to ∞ a rescaled Gaussian model RMgauss C(r) = -r^2 appears as limit of the above Handcock-Wallis parametrisation.

For generalizations see section ‘See Also’.

Value

The functions return an object of class RMmodel.

Note

The Whittle-Matern model is a normal scale mixture.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Covariance function

Chiles, J.-P. and Delfiner, P. (1999) Geostatistics. Modeling Spatial Uncertainty. New York: Wiley.

Gelfand, A. E., Diggle, P., Fuentes, M. and Guttorp, P. (eds.) (2010) Handbook of Spatial Statistics. Boca Raton: Chapman & Hall/CRL.

Guttorp, P. and Gneiting, T. (2006) Studies in the history of probability and statistics. XLIX. On the Matern correlation family. Biometrika 93, 989–995.

Handcock, M. S. and Wallis, J. R. (1994) An approach to statistical spatio-temporal modeling of meteorological fields. JASA 89, 368–378.

Stein, M. L. (1999) Interpolation of Spatial Data – Some Theory for Kriging. New York: Springer.

Tail correlation function (for 0 < ν ≤ 1/2)

Strokorb, K., Ballani, F., and Schlather, M. (2014) Tail correlation functions of max-stable processes: Construction principles, recovery and diversity of some mixing max-stable processes with identical TCF. Extremes, Submitted.

See Also

RMexp, RMgauss for special cases of the model (for ν=0.5 and ν=∞, respectively)

RMhyperbolic for a univariate generalization

RMbiwm for a multivariate generalization

RMnonstwm, RMstein for anisotropic (space-time) generalizations

RMmodel, RFsimulate, RFfit for general use.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

x <- seq(0, 1, len=100)
model <- RMwhittle(nu=1, Aniso=matrix(nc=2, c(1.5, 3, -3, 4)))
plot(model, dim=2, xlim=c(-1,1))
z <- RFsimulate(model=model, x, x)
plot(z)

[Package RandomFields version 3.3.6 Index]