RMmodel {RandomFields}	R Documentation

Covariance and Variogram Models in RandomFields (RM commands)

Description

Summary of implemented covariance and variogram models

Details

To generate a covariance or variogram model for use within RandomFields, calls of the form

RM_name_(..., var, scale, Aniso, proj)

can be used, where _name_ has to be replaced by a valid model name.

... can take model specific arguments.

var is the optional variance argument v,

scale the optional scale argument s,

Aniso an optional anisotropy matrix A or given by RMangle, and

proj is the optional projection.

With φ denoting the original model, the transformed model is C(h) = v * φ(A*h/s). See RMS for more details.

RM_name_ must be a function of class RMmodelgenerator. The return value of all functions RM_name_ is of class RMmodel.

The following models are available (cf. RFgetModelNames):

Basic stationary and isotropic models

RMcauchy 	 Cauchy family

RMexp 	 exponential model

RMgencauchy 	 generalized Cauchy family

RMgauss 	 Gaussian model

RMgneiting 	 differentiable model with compact support

RMmatern 	 Whittle-Matern model

RMnugget 	 nugget effect model

RMspheric 	 spherical model

RMstable 	 symmetric stable family or powered exponential model

RMwhittle 	 Whittle-Matern model, alternative parametrization

Variogram models (stationary increments/intrinsically stationary)

RMfbm 	 fractal Brownian motion

Basic Operations

RMmult, * 	 product of covariance models

RMplus, + 	 sum of covariance models or variograms

Others

RMtrend 	 trend

RMangle 	 defines a 2x2 anisotropy matrix by rotation and stretch arguments.

Author(s)

Alexander Malinowski; Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Chiles, J.-P. and Delfiner, P. (1999) Geostatistics. Modeling Spatial Uncertainty. New York: Wiley.

Schlather, M. (1999) An introduction to positive definite functions and to unconditional simulation of random fields. Technical report ST 99-10, Dept. of Maths and Statistics, Lancaster University.

Schlather, M. (2011) Construction of covariance functions and unconditional simulation of random fields. In Porcu, E., Montero, J.M. and Schlather, M., Space-Time Processes and Challenges Related to Environmental Problems. New York: Springer.

Yaglom, A.M. (1987) Correlation Theory of Stationary and Related Random Functions I, Basic Results. New York: Springer.

Wackernagel, H. (2003) Multivariate Geostatistics. Berlin: Springer, 3nd edition.

See Also

RM for an overview over more advanced classes of models
RC, RF, RP, RR, R., RFcov, RFformula, RMmodelsAdvanced, RMmodelsAuxiliary, trend modelling

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## an example of a simple model
model <- RMexp(var=1.6, scale=0.5) + RMnugget(var=0) #exponential + nugget
plot(model)

[Package RandomFields version 3.3.6 Index]