RMmodel-class {RandomFields}	R Documentation

Class RMmodel

Description

Class for RandomFields' representation of explicit covariance models

Usage

RFplotModel(x, y, dim=1,
           n.points=if (dim==1 || is.contour) 200 else 100,
           fct.type=NULL, MARGIN, fixed.MARGIN, maxchar=15, ...,
           plotmethod=if (dim==1) "matplot" else "contour")

## S4 method for signature 'RMmodel,missing'
plot(x, y, ...)
## S4 method for signature 'RMmodel'
points(x, ..., type="p")
## S4 method for signature 'RMmodel'
lines(x, ..., type="l")
## S4 method for signature 'RMmodel'
image(x, ..., dim=2)
## S4 method for signature 'RMmodel'
persp(x, ..., dim=2, zlab="")

Arguments

x
object of class RFsp or RFempVario or RFfit or RMmodel; in the latter case, x can be any sophisticated model but it must be either stationary or a variogram model.

y
ignored in most methods

MARGIN
vector of two; two integer values giving the coordinate dimensions w.r.t. whether the field or the covariance model is to be plotted; in all other directions, the first index is taken.

fixed.MARGIN
only for class(x)==CLASS_CLIST and if dim > 2; a vector of length dim-2 with distance values for the coordinates that are not displayed.

maxchar
integer. Maximum number of characters to print the model in the legend.

...
arguments to be passed to methods; mainly graphical arguments, or further models in case of class CLASS_CLIST, see Details.

dim
must equal 1 or 2; only for class(x)==CLASS_CLIST; the covariance function and the variogram are plotted as a function of R^\code{dim}.

n.points
integer; only for class(x)==CLASS_CLIST; the number of points at which the model is evaluated (in each dimension); defaults to 200.

fct.type
character; only for class(x)==CLASS_CLIST; must equal NULL, "Cov" or "Variogram"; controls whether the covariance (fct.type="Cov") or the variogram (fct.type="Variogram") is plotted; NULL implies automatic choice, where "Cov" is chosen whenever the model is stationary.

plotmethod
string or function. Internal.

type
character. See points.

zlab
character. See persp.

Value

If RFoptions()$split_screen=TRUE and RFoptions()$close_screen=TRUE then the plot functions return the screen numbers. Else NULL.

Creating Objects

Objects are created by calling a function of class RMmodelgenerator.

Slots

call:

language object; the function call by which the object was generated

name:

character string; nickname of the model, name of the function by which the object was generated

submodels:

list; contains submodels (if existent)

par.model:

list; contains model specific arguments

par.general:

list of 4; contains the four standard arguments var, scale, Aniso and proj that can be given for any model; if not specified by the user, the string "RFdefault" is inserted

Methods

+

signature(x = CLASS_CLIST): allows to sum up covariance models; internally calls RMplus.

-

signature(x = CLASS_CLIST): allows to substract covariance models; internally calls R.minus.

*

signature(x = CLASS_CLIST): allows to multiply covariance models; internally calls R.mult.

/

signature(x = CLASS_CLIST): allows to divide covariance models; internally calls R.div.

c

signature(x = CLASS_CLIST): concatenates covariance functions or variogram models.

plot

signature(x = CLASS_CLIST): gives a plot of the covariance function or of the variogram model; for more details see plot-method.

points

signature(x = CLASS_CLIST): adds a covariance plot to an existing plot; for more details see plot-method.

lines

signature(x = CLASS_CLIST): adds a covariance plot to an existing plot; for more details see plot-method.

str

signature(x = CLASS_CLIST): as the usual str-method for S4 objects but only those entries of the 'par.general'-slot are shown that contain values different from 'RFdefault'.

show

signature(x = CLASS_CLIST): returns the structure of x.

print

signature(x = CLASS_CLIST): identical with show-method, additional argument is max.level.

[

signature(x = CLASS_CLIST): enables accessing the slots via the "["-operator, e.g. x["par.general"].

[<-

signature(x = CLASS_CLIST): enables replacing the slots via the "["-operator.

signature(x = CLASS_CLIST, y = "missing")

Generates covariance function or variogram function plots in one or two dimensions.

Details

All the above arguments apply for all the S3 and S4 functions given here as they call RFplotModel immediately.

Author(s)

Alexander Malinowski, Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmodelgenerator RMmodel

Examples

# see RMmodel for introductory examples

# Compare:
model <- RMexp(scale=2) + RMnugget(var=3)
str(model)  ## S4 object as default in version 3 of RandomFields

model <- summary(model)
str(model)  ## list style as in version 2 of RandomFields
            ## see also 'spConform' in 'RFoptions' to make this style
            ## the default

[Package RandomFields version 3.3.6 Index]