RMmodelFit-class {RandomFields}	R Documentation

Class CLASS_FIT

Description

Extension of class RMmodel which additionally contains the likelihood of the data w.r.t. the covariance model represented by the CLASS_CLIST part, the estimated trend of the data if it is a constant trend, and the residuals of the data w.r.t. the model. Objects of this class only occur as slots in the output of "RFfit".

Creating Objects

Objects are only meant to be created by the function RFfit.

Slots

model,formel:

See RMmodel.

variab:

vector of estimated variables. Variables are used in the internal representation and can be a subset of the parameters.

param:

vector of estimated parameters

covariate:

to do

globalvariance:

to do

hessian:

to do

likelihood:

numeric; the likelihood of the data w.r.t. the covariance model

AIC:

the AIC value for the ml estimation

AICc:

the corrected AIC value for the ml estimation

BIC:

the BIC value for the ml estimation

residuals:

array or of class RFsp; residuals of the data w.r.t. the trend model

Extends

Class CLASS_CLIST, directly.

Methods

[

signature(x = CLASS_FIT): enables accessing the slots via the "["-operator, e.g. x["likelihood"]

[<-

signature(x = CLASS_FIT): enables replacing the slots via the "["-operator

show

signature(x = "RFfit"): returns the structure of x

print

signature(x = "RFfit"): identical with show-method

anova

performs a likelihood ratio test base on a chisq approximation

summary

gives a summary

Author(s)

Alexander Malinowski; Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmodel, RFfit.

Examples

# see RFfit

[Package RandomFields version 3.3.6 Index]