RMmodelgenerator-class {RandomFields}	R Documentation

Class RMmodelgenerator

Description

Class for all functions of this package with prefix RM, i.e. all functions that generate objects of class RMmodel; direct extension of class function.

Creating Objects

Objects should not be created by the user!

Slots

.Data:

function; the genuine function that generates an object of class RMmodel

type:

character string; specifies the category of RMmodel-function, see Details

domain:

character string; specifies whether the corresponding function(s) depend on 1 or 2 variables, see Details

isotropy:

character string; specifies the type of isotropy of the corresponding covariance model, see Details

operator:

logical; specifies whether the underlying covariance model is an operator, see Details

monotone:

character string; specifies the kind of monotonicity of the model

finiterange:

logical; specifies whether the underlying covariance model has finite range, see Details

simpleArguments:

logical. If TRUE than all the parameters are real valued (or integer valued).

maxdim:

numeric; the maximal dimension, in which the corresponding model is a valid covariance model, see Details

vdim:

numeric; dimension of the value of the random field at a single fixed location, equals 1 in most cases, see Details

Extends

Class function, directly.

Methods

show

signature(x = CLASS_CLIST): returns the structure of x

print

signature(x = CLASS_CLIST): identical with show-method

[

signature(x = CLASS_RM): enables accessing the slots via the "["-operator, e.g. x["maxdim"]

[<-

signature(x = CLASS_RM): enables replacing the slots via the "["-operator

Details

type:

can be one of the following strings:

'tail correlation function':

indicates that the function returns a tail correlation function (a subclass of the set of positive definite functions)

'positive definite':

indicates that the function returns a covariance function (positive definite function)

'negative definite':

indicates that the function returns a variogram model (negative definite function)

'process':

functions of that type determine the class of processes to be simulated

'method for Gauss processes':

methods to simulate Gaussian random fields

'method for Brown-Resnick processes':

methods to simulate Brown-Resnick fields

'point-shape function':

functions of that type determine the distribution of points in space

'distribution family':

e.g. (multivariate) uniform distribution, normal distribution, etc., defined in RandomFields. See RR for a complete list.

'shape function':

functions used in, e.g., M3 processes (RPsmith)

'trend':

RMtrend or a mixed model

'interface':

indicates internal models which are usually not visible for the users. These functions are the internal representations of RFsimulate, RFcov, etc. See RF for a complete list.

'undefined':

some models can take different types, depending on the parameter values and/or the submodels

'other type':

very very special internal functions, not belonging to any of the above types.

domain:

can be one of the following strings:

'single variable':

Function depending on a single variable

'kernel':

model refers to a kernel, e.g. a non-stationary covariance function

'framework dependent':

domain depends on the calling model

'mismatch':

this option is used only internally and should never appear

isotropy:

can be one of the following strings:

'isotropic':

indicates that the model is isotropic

'space-isotropic':

indicates that the spatial part of a spatio-temporal model is isotropic

'zero-space-isotropic':

this property refers to space-time models; the model is called zerospaceisotropic if it is isotropic as soon as the time-component is zero

'vector-isotropic':

multivariate vector model (flow fields) have a different notion of isotropy

'symmetric':

the most basic property of any covariance function or variogram model

'cartesian system', 'earth system', 'spherical system', 'cylinder system':

different coordinate systems

'non-dimension-reducing':

the property f(x) = f(-x)^\top does not hold

'parameter dependent':

indicates that the type of isotropy of the model depends on the parameters passed to the model; in particular parameters may be submodels if an operator model is considered

'<mismatch>':

this option is used only internally and should never appear

operator:

if TRUE, the model requires at least one submodel

monotone:

'mismatch in monotonicity':

used if a statement on the monotonocity does not make sense, e.g. for RRmodels

'submodel dependent monotonicity':

only for operators, e.g. RMS

'previous model dependent monotonicity':

internal; should not be used

'parameter dependent monotonicity':

some models change their properties according to the parameters

'not monotone':

none of the above categories; either the function is not monotone or properties are unknown

'monotone':

isotone or antitone

'Gneiting-Schaback class':

function belonging to Euclid's hat in Gneiting's 1999 paper

'normal mixture':

scale mixture of the Gaussian model

'completely monotone':

completely monotone function

'Bernstein':

Bernstein function

Note that

'not monotone' includes 'monotone' and 'Bernstein'

'monotone' includes 'Gneiting-Schaback class'

'Gneiting-Schaback class' includes 'normal mixture'

'normal mixture' includes 'completely monotone'

finiterange:

if TRUE, the covariance of the model has finite range

maxdim:

if a positive integer, maxdim gives the maximum dimension in which the model is a valid covariance model, can be Inf; maxdim=-1 means that the actual maxdim depends on the parameters; maxdim=-2 means that the actual maxdim depends on the submodel(s)

vdim:

if a positive integer, vdim gives the dimension of the random field, i.e. univariate, bi-variate, ...; vdim=-1 means that the actual vdim depends on the parameters; vdim=-2 means that the actual vdim depends on the submodel(s)

Author(s)

Alexander Malinowski, Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. (1999) Radial positive definite functions generated by Euclid's hat, J. Multivariate Anal., 69, 88-119.

See Also

RMmodel, RFgetModelNames

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
RFgetModelNames()

[Package RandomFields version 3.3.6 Index]