RMmodelsAdvanced {RandomFields}	R Documentation

Advanced features of the models

Description

Here, further models and advanced comments for RMmodel are given. See also RFgetModelNames.

Details

Further stationary and isotropic models

RMaskey 	 Askey model (generalized test or triangle model)

RMbcw 	 bridging model between RMcauchy and RMgenfbm

RMbessel 	 Bessel family

RMcircular 	 circular model

RMconstant 	 spatially constant model

RMcubic 	 cubic model (see Chiles and Delfiner)

RMdagum 	 Dagum model

RMdampedcos 	 exponentially damped cosine

RMqexp 	 variant of the exponential model

RMfractdiff 	 fractionally differenced process

RMfractgauss 	 fractional Gaussian noise

RMgengneiting 	 generalized Gneiting model

RMgneitingdiff 	 Gneiting model for tapering

RMhyperbolic 	 generalized hyperbolic model

RMlgd 	 Gneiting's local-global distinguisher

RMlsfbm 	 locally stationary fractal Brownian motion

RMpenta 	 penta model (see Chiles and Delfiner)

RMpower 	 Golubov's model

RMwave 	 cardinal sine

Variogram models (stationary increments/intrinsically stationary)

RMbcw 	 bridging model between RMcauchy and RMgenfbm

RMdewijsian 	 generalized version of the DeWijsian model

RMgenfbm 	 generalized fractal Brownian motion

RMflatpower 	 similar to fractal Brownian motion but always smooth at the origin

General composed models (operators)

Here, composed models are given that can be of any kind (stationary/non-stationary), depending on the submodel.

RMbernoulli 	 Correlation function of a binary field based on a Gaussian field

RMexponential 	 exponential of a covariance model

RMintexp 	 integrated exponential of a covariance model (INCLUDES ma2)

RMpower 	 powered variograms

RMqam 	 Porcu's quasi-arithmetic-mean model

RMS 	 details on the optional transformation arguments (var, scale, Aniso, proj)

Stationary and isotropic composed models (operators)

RMcutoff 	 Gneiting's modification towards finite range

RMintrinsic 	 Stein's modification towards finite range

RMnatsc 	 practical range

RMstein 	 Stein's modification towards finite range

RMtbm	 Turning bands operator

Stationary space-time models
See RMmodelsSpaceTime.

Non-stationary models
See RMmodelsNonstationary.

Negative definite models that are not variograms

RMsum 	 a non-stationary variogram model

Models related to max-stable random fields (tail correlation functions)
See RMmodelsTailCorrelation.

Other covariance models

RMcov 	 covariance structure given by a variogram

RMfixcov 	 User defined covariance structure

RMuser 	 User defined model

Trend models

Aniso 	 for space transformation (not really trend, but similar)

RMcovariate 	 spatial covariates

RMprod 	 to model variability of the variance

RMpolynome 	 easy modelling of polynomial trends

RMtrend 	 for explicit trend modelling

R.models 	 for implicit trend modelling

R.c 	 for multivariate trend modelling

Auxiliary models
See Auxiliary RMmodels.

Note

Note that, instead of the named arguments, a single argument k can be passed. This is possible if all the arguments are scalar. Then k must have a length equal to the number of arguments.

If an argument equals NULL the argument is not set (but must have a valid name).

Aniso can be given also by RMangle or any other RMmodel instead of a matrix

Note also that a completely different possibility exists to define a model, namely by a list. This format allows for easy flexible models and modifications (and some few more options, as well as some abbreviations to the model names, see PrintModelList()). Here, the argument var, scale, Aniso and proj must be passed by the model RMS. For instance,

model <- RMexp(scale=2, var=5)
is equivalent to
model <- list("RMS", scale=2, var=5, list("RMexp"))
The latter definition can be also obtained by

print(RMexp(scale=2, var=5))

model <- RMnsst(phi=RMgauss(var=7), psi=RMfbm(alpha=1.5), scale=2, var=5)
is equivalent to
model <- list("RMS", scale=2, var=5,
list("RMnsst", phi=list("RMS", var=7, list("RMgauss")),
psi=list("RMfbm", alpha=1.5)) ).

All models have secondary names that stem from RandomFields versions 2 and earlier and that can also be used as strings in the list notation. See RFgetModelNames(internal=FALSE) for the full list.

Author(s)

Alexander Malinowski; Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Chiles, J.-P. and Delfiner, P. (1999) Geostatistics. Modeling Spatial Uncertainty. New York: Wiley.

Schlather, M. (1999) An introduction to positive definite functions and to unconditional simulation of random fields. Technical report ST 99-10, Dept. of Maths and Statistics, Lancaster University.

Schlather, M. (2011) Construction of covariance functions and unconditional simulation of random fields. In Porcu, E., Montero, J.M. and Schlather, M., Space-Time Processes and Challenges Related to Environmental Problems. New York: Springer.

Schlather, M., Malinowski, A., Menck, P.J., Oesting, M. and Strokorb, K. (2015) Analysis, simulation and prediction of multivariate random fields with package RandomFields. Journal of Statistical Software, 63 (8), 1-25, url = ‘http://www.jstatsoft.org/v63/i08/’

‘multivariate’, the corresponding vignette.

Yaglom, A.M. (1987) Correlation Theory of Stationary and Related Random Functions I, Basic Results. New York: Springer.

Wackernagel, H. (2003) Multivariate Geostatistics. Berlin: Springer, 3nd edition.

See Also

RFformula, RM, RMmodels, RMmodelsAuxiliary.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## a non-stationary field with a sharp boundary
## of the differentiabilities
x <- seq(-0.6, 0.6, len=50)
model <- RMwhittle(nu=0.8 + 1.5 * R.is(R.p(new="isotropic"), "<=", 0.5))
z <- RFsimulate(model=model, x, x, n=4)
plot(z)

[Package RandomFields version 3.3.6 Index]