Others {RandomFields}	R Documentation

Auxiliary and other Models

Description

Here, auxiliary models are given that are not covariance functions or variograms, but which might be used within the definition of a model.

Implemented models

Distribution families See RR.

Evaluation operators See RF.

Random Fields / Random Processes See RP.

Mathematical functions See R.

Shape functions

Besides any of the covariance functions the following functions can be used as shape functions.

RMangle 	 defines an anisotropy matrix by angle and a diagonal matrix

RMball 	 Indicator of a ball of radius 1/2

RMm2r 	 spectral function belonging to a tail correlation function of the Gneiting class H_n

RMm3b 	 spectral function belonging to a tail correlation function of the Gneiting class H_n

RMmppplus 	 operator to define mixed moving maxima (M3) processes

RMmps 	 spectral functions belonging to a tail correlation function of the Gneiting class H_n

RMpolygon 	 Indicator of a typical Poisson polygon

RMrational 	 shape function used in the Bernoulli paper given in the references

RMrotat 	 shape function used in the Bernoulli paper given in the references

RMsign 	 random sign

RMtruncsupport 	 truncates the support of a shape in a Poisson based model

Special transformations

RMeaxxa 	 shape function used in the Bernoulli paper given in the references

RMetaxxa 	 shape function used in the Bernoulli paper given in the references

RMidmodel 	 model identity

RMid 	 identity but interpretation turns from a coordinate to a model value

RMtrafo 	 allows to model the identity within the set of coordinates

RMrotation 	 shape function used in the Bernoulli paper given in the references

Other models

RMuser 	 User defined covariance model

Author(s)

Alexander Malinowski; Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RM

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
RFgetModelNames()

[Package RandomFields version 3.3.6 Index]