RMmodelsMultivariate {RandomFields}	R Documentation

Multivariate models

Description

Here, multivariate and vector-valued covariance models are presented.

Details

Bivariate covariance models

RMbicauchy 	 a bivariate Cauchy model

RMbiwm 	 full bivariate Whittle-Matern model (stationary and isotropic)

RMbigneiting 	 bivariate Gneiting model (stationary and isotropic)

RMbistable 	 a bivariate stable model

Physically motivated, vector valued covariance and variogram models

RMcurlfree 	 curlfree (spatial) vector-valued field (stationary and anisotropic)

RMdivfree 	 divergence free (spatial) vector-valued field (stationary and anisotropic)

RMkolmogorov 	 Kolmogorov's model of turbulence

RMvector 	 vector-valued field (combining RMcurlfree and RMdivfree)

Multivariate covariance models

RMdelay 	 delay effect model (stationary)

RMderiv 	 field and its gradient

RMmatrix 	 linear model of coregionalization

RMparswm 	 multivariate Whittle-Matern model (stationary and isotropic)

Operators

RMcov 	 covariance structure given by a multivariate variogram

RMexponential 	 functional returning exp(C)

RMmatrix 	 linear model of coregionalization

RMmqam 	 multivariate quasi-arithmetic mean (stationary)

RMschur 	 element-wise product with a positive definite matrix

RMtbm 	 turning bands operator

Trend models

RMtrend 	 for explicit trend modelling

R.models 	 for implicit trend modelling

R.c 	 binding univariate trend models into a vector

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Chiles, J.-P. and Delfiner, P. (1999) Geostatistics. Modeling Spatial Uncertainty. New York: Wiley.

Schlather, M. (2011) Construction of covariance functions and unconditional simulation of random fields. In Porcu, E., Montero, J.M. and Schlather, M., Space-Time Processes and Challenges Related to Environmental Problems. New York: Springer.

Schlather, M., Malinowski, A., Menck, P.J., Oesting, M. and Strokorb, K. (2015) Analysis, simulation and prediction of multivariate random fields with package RandomFields. Journal of Statistical Software, 63 (8), 1-25, url = ‘http://www.jstatsoft.org/v63/i08/’

Wackernagel, H. (2003) Multivariate Geostatistics. Berlin: Springer, 3rd edition.

See Also

RFformula, RMmodels, RM, RMmodelsAdvanced

‘multivariate’, a vignette for multivariate geostatistics

Examples

 
RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

n <- 100
x <- runif(n=n, min=1, max=50)
y <- runif(n=n, min=1, max=50)

rho <- matrix(nc=2, c(1, -0.8, -0.8, 1))
model <- RMparswmX(nudiag=c(0.5, 0.5), rho=rho)

## generation of artifical data
dta <- RFsimulate(model = model, x=x, y=y, grid=FALSE)

## introducing some NAs ...
dta@data$variable1[1:10] <- NA
if (interactive()) dta@data$variable2[90:100] <- NA

plot(dta)

## co-kriging
x <- y <- seq(0, 50, 1)

k <- RFinterpolate(model, x=x, y=y, data= dta)
plot(k, dta)

## conditional simulation
z <- RFsimulate(model, x=x, y=y, data= dta) ## takes a while
plot(z, dta)

[Package RandomFields version 3.3.6 Index]