RMmodelsSpacetime {RandomFields}	R Documentation

Space-time Covariance Models

Description

Here, a collection of implemented space-time models is given.

Details

Stationary space-time models

Here, most of the models are composed models (operators). Note that in space-time modelling the argument proj may also take the values "space" for the projection on the space and "time" for the projection onto the time axis.

separable models 	 are easily constructed using +, *, and proj, see also the example below

RMave 	 space-time moving average model

RMcoxisham 	 Cox-Isham model

RMcurlfree 	 curlfree (spatial) field (stationary and anisotropic)

RMdivfree 	 divergence free (spatial) vector-valued field (stationary and anisotropic)

RMgennsst 	 generalization of Gneiting's non-separable space-time model

RMiaco 	 non-separable space-time model

RMmastein 	 Ma-Stein model

RMnsst 	 Gneiting's non-separable space-time model

RMstein 	 Stein's non-separable space-time model

RMstp 	 Single temporal process

RMtbm 	 Turning bands operator

Author(s)

Alexander Malinowski; Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Schlather, M. (2011) Construction of covariance functions and unconditional simulation of random fields. In Porcu, E., Montero, J.M. and Schlather, M., Space-Time Processes and Challenges Related to Environmental Problems. New York: Springer.

See Also

RFformula, RM, RMmodels, RMmodelsAdvanced.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## multiplicative separable model with exponential model in space
## and Gaussian in time
model <- RMexp(proj = "space") * RMgauss(proj = "time")
x <- T <- seq(0, 10, 0.1)
z <- RFsimulate(model, x=x, T=T)
plot(z)

## additive separable model with exponential model in space
## and Gaussian in time. The structure is getting rather simple,
## see the function stopifnot below
model <- RMexp(proj = "space") + RMgauss(proj = "time")
x <- T <- seq(0, 10, 0.1)
z <- RFsimulate(model, x=x, T=T)
stopifnot(sum(abs(apply(apply(z, 1, diff), 1, diff))) < 1e-14)
plot(z)

[Package RandomFields version 3.3.6 Index]