Spherical models {RandomFields}	R Documentation

Covariance models valid on a sphere

Description

This page summarizes the covariance models that can be used for spherical coordinates (and earth coordinates).

Details

The following models are available:

Completely monotone functions allowing for arbitrary scale

RMbcw 	 Model bridging stationary and intrinsically stationary processes for α ≤ 1 and β < 0

RMcubic 	 cubic model

RMdagum 	 Dagum model with β < γ and γ ≤ 1

RMexp 	 exponential model

RMgencauchy 	 generalized Cauchy family with α ≤ 1 (and arbitrary β> 0)

RMmatern 	 Whittle-Matern model with ν ≤ 1/2

RMstable 	 symmetric stable family or powered exponential model with α ≤ 1

RMwhittle 	 Whittle-Matern model, alternative parametrization with ν ≤ 1/2

Other isotropic models with arbitrary scale

RMconstant 	 spatially constant model

RMnugget 	 nugget effect model

Compactly supported covariance functions allowing for scales up to π (or 180 degrees)

RMaskey 	 Askey's model

RMcircular 	 circular model

RMgengneiting 	 Wendland-Gneiting model; differentiable models with compact support

RMgneiting 	 differentiable model with compact support

RMspheric 	 spherical model

Anisotropic models

None up to now.

Basic Operators

RMmult, * 	 product of covariance models

RMplus, + 	 sum of covariance models or variograms

See RMmodels for cartesian models.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

coordinate systems, RMmodels, RMtrafo.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

RFgetModelNames(isotropy=c("spherical isotropic"))

## an example of a simple model valid on a sphere
model <- RMexp(var=1.6, scale=0.5) + RMnugget(var=0) #exponential + nugget
plot(model)

## a simple simulation
l <- seq(0, 85, 1.2)
coord <- cbind(lon=l, lat=l)

z <- RFsimulate(RMwhittle(s=30, nu=0.45), coord, grid=TRUE) # takes 1 min
plot(z)

z <- RFsimulate(RMwhittle(s=500, nu=0.5), coord, grid=TRUE,
                new_coord_sys="orthographic", zenit=c(25, 25)) 
plot(z)

z <- RFsimulate(RMwhittle(s=500, nu=0.5), coord, grid=TRUE,
                new_coord_sys="gnomonic", zenit=c(25, 25)) 
plot(z)

## space-time modelling on the sphere
sigma <- 5 * sqrt((R.lat()-30)^2 + (R.lon()-20)^2)
model <- RMprod(sigma) * RMtrafo(RMexp(s=500, proj="space"), "cartesian") *
  RMspheric(proj="time") 
z <- RFsimulate(model, 0:10, 10:20, T=seq(0, 1, 0.1),
                coord_system="earth", new_coordunits="km")
plot(z, MARGIN.slices=3)

[Package RandomFields version 3.3.6 Index]