Tail Correlation Functions {RandomFields}	R Documentation

Covariance models valid for max-stable random fields

Description

This page summarizes the models that can be used for tail correlation functions.

Details

The following models are available:

Completely monotone functions allowing for arbitrary scale

RMbcw 	 Model bridging stationary and intrinsically stationary processes for alpha <= 1 and beta < 0

RMdagum 	 Dagum model with β < γ and γ ≤ 1

RMexp 	 exponential model

RMgencauchy 	 generalized Cauchy family with α ≤ 1 (and arbitrary β> 0)

RMmatern 	 Whittle-Matern model with ν ≤ 1/2

RMstable 	 symmetric stable family or powered exponential model with α ≤ 1

RMwhittle 	 Whittle-Matern model, alternative parametrization with ν ≤ 1/2

Other isotropic models with arbitrary scale

RMnugget 	 nugget effect model

Compactly supported covariance functions

RMaskey 	 Askey's model

RMcircular 	 circular model

RMconstant	 identically constant

RMcubic 	 cubic model

RMgengneiting 	 Wendland-Gneiting model; differentiable models with compact support

RMgneiting 	 differentiable model with compact support

RMspheric 	 spherical model

Anisotropic models

None up to now.

Basic Operators

RMmult, * 	 product of covariance models

RMplus, + 	 sum of covariance models or variograms

Operators related to process constructions

RMbernoulli 	 correlation of binary fields

RMbrownresnick	 tcf of a Brown-Resnick process

RMschlather	 tcf of an extremal Gaussian process / Schlather process

RMm2r	 M2 process with monotone shape function

RMm3b	 M3 process with balls of random radius

RMmps	 M3 process with hyperplane polygons

See RMmodels for cartesian models.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Strokorb, K., Ballani, F., and Schlather, M. (2015) Tail correlation functions of max-stable processes: Construction principles, recovery and diversity of some mixing max-stable processes with identical TCF. Extremes, 18, 241-271

See Also

coordinate systems, RM, RMmodels, RMtrafo.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
RFgetModelNames(type="tail")

## an example of a simple model
model <- RMexp(var=1.6, scale=0.5) + RMnugget(var=0) #exponential + nugget
plot(model)

[Package RandomFields version 3.3.6 Index]