RMmppplus {RandomFields}	R Documentation

Mixture of shape functions

Description

RMmppplus is a multivariate covariance model which depends on up to 10 submodels C_0, C_1, ..., C_9.

Used together with RPsmith, it allows for mixed moving maxima with a finite number of shape functions.

Usage

RMmppplus(C0, C1, C2, C3, C4, C5, C6, C7, C8, C9, p)

Arguments

C0
an RMmodel

C1,C2,C3,C4,C5,C6,C7,C8,C9
optional; each an RMmodel

p
vector of probabilities for the shape functions. The probabilities should add up to 1. The length of the vector equals the number of given submodels.

Value

RMmppplus returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMplus, RMmodel, RFsimulate, RFfit, RPsmith.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

[Package RandomFields version 3.3.6 Index]