RMmqam {RandomFields}	R Documentation

multivariate quasi-arithmetic mean

Description

RMmqam is a multivariate stationary covariance model depending on a submodel phi such that psi( . ) := phi(sqrt( . )) is completely monotone, and depending on further stationary covariance models C_i. The covariance is given by

C_{ij}(h) = φ(√(θ_i (φ^{-1}(C_i(h)))^2 + θ_j (φ^{-1}(C_j(h)))^2 ))

where φ is a completely monotone function, C_i are suitable covariance functions and θ_i≥0 such that ∑_i θ_i=1.

Usage

RMmqam(phi, C1, C2, C3, C4, C5, C6, C7, C8, C9, theta, var, scale, Aniso, proj)

Arguments

phi
a valid covariance RMmodel that is a normal scale mixture. See, for instance,
RFgetModelNames(monotone="normal mixture")

C1, C2, C3, C4, C5, C6, C7, C8, C9
optional further stationary RMmodels

theta
is a vector of values in [0,1], summing up to 1.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

Note that psi( . ) := phi(sqrt( . )) is completely monotone if and only if phi is a valid covariance function for all dimensions, e.g. RMstable, RMgauss, RMexponential.

Warning: RandomFields cannot check whether the combination of phi and C_i is valid.

Value

RMmqam returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Porcu, E., Mateu, J. & Christakos, G. (2009) Quasi-arithmetic means of covariance functions with potential applications to space-time data. Journal of Multivariate Analysis, 100, 1830–1844.

See Also

RMqam, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

RFoptions(modus_operandi="sloppy")
model <- RMmqam(phi=RMgauss(),RMgauss(),RMexp(),theta=c(0.4, 0.6), scale=0.5)
x <- seq(0, 10, 0.02)
plot(model)
z <- RFsimulate(model=model, x=x)
plot(z)

RFoptions(modus_operandi="normal")

[Package RandomFields version 3.3.6 Index]