RMmult {RandomFields}	R Documentation

Multiplication of Random Field Models

Description

RMmult is a multivariate covariance model which depends on up to 10 submodels C_0, C_1, ..., C_9. In general, realizations of the created RMmodel are pointwise products of independent realizations of the submodels.

In particular, if all submodels are given through a covariance function, the resulting model is defined through its covariance function, which is the product of the submodels' covariances.

Usage

RMmult(C0, C1, C2, C3, C4, C5, C6, C7, C8, C9, var, scale, Aniso, proj)

Arguments

C0
an RMmodel.

C1,C2,C3,C4,C5,C6,C7,C8,C9
optional; each an RMmodel.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above model remains unmodified.

Details

RMmodels can also be multiplied via the *-operator, e.g. C0 * C1.

The global arguments scale,Aniso,proj of RMmult are multiplied to the corresponding argument of the submodels (from the right side). E.g.,
RMmult(Aniso=A1, RMexp(Aniso=A2), RMspheric(Aniso=A3))
equals
RMexp(Aniso=A2 %*% A1) * RMspheric(Aniso=A3 %*% A1)

In case that all submodels are given through a covariance function, the global argument var of RMmult is multiplied to the product covariance of RMmult.

Value

RMmult returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMplus, RMmodel, RMprod, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

# separable, multiplicative model
model <- RMgauss(proj=1) * RMexp(proj=2, scale=5)
z <- RFsimulate(model=model, 0:10, 0:10, n=4)
plot(z)

[Package RandomFields version 3.3.6 Index]