RMmultiquad {RandomFields}	R Documentation

The Multiquadric Family Covariance Model on the Sphere

Description

RMmultiquad is an isotropic covariance model. The corresponding covariance function, the multiquadric family, only depends on the angle 0 ≤ θ ≤ π between two points on the sphere and is given by

ψ(θ) = (1 - δ)^{2*τ} / (1 + delta^2 - 2*δ*cos(θ))^{τ},

where 0 < δ < 1 and τ > 0.

Usage

RMmultiquad(delta, tau, var, scale, Aniso, proj)

Arguments

delta
a numerical value in (0,1)

tau
a numerical value greater than 0

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

Special cases (cf. Gneiting, T. (2013), p.1333) are known for fixed parameter τ=0.5 which leads to the covariance function called 'inverse multiquadric'

ψ(θ) = (1 - δ) / √( 1 + delta^2 - 2*δ*cos(θ) )

and for fixed parameter τ=1.5 which gives the covariance function called 'Poisson spline'

ψ(θ) = (1 - δ)^{3} / (1 + delta^2 - 2*δ*cos(θ))^{1.5}.

For a more general form, see RMchoquet.

Value

RMmultiquad returns an object of class RMmodel.

Author(s)

Christoph Berreth, Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. (2013) Strictly and non-strictly positive definite functions on spheres Bernoulli, 19(4), 1327-1349.

See Also

RMmodel, RFsimulate, RFfit, RMchoquet, spherical models

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

RFoptions(coord_system="sphere")
model <- RMmultiquad(delta=0.5, tau=1)
plot(model, dim=2)

## the following two pictures are the same

x <- seq(0, 0.12, 0.01)
z1 <- RFsimulate(model, x=x, y=x)
plot(z1)

x2 <- x * 180 / pi
z2 <- RFsimulate(model, x=x2, y=x2, coord_system="earth")
plot(z2)

stopifnot(all.equal(as.array(z1), as.array(z2)))

RFoptions(coord_system="auto")

[Package RandomFields version 3.3.6 Index]