RMnatsc {RandomFields}	R Documentation

Natural scale

Description

RMnatsc is a stationary isotropic covariance model that depends on a stationary isotropic covariance model phi. The covariance is given by

C(h) = phi(h / s),

where the argument s is chosen by RMnatsc such that the practical range or the mathematical range, if finite, is 1.

Usage

RMnatsc(phi, var, scale, Aniso, proj)

Arguments

phi
a stationary isotropic covariance RMmodel.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

For internal use only.

Value

RMnatsc returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMnatsc(RMexp())
x <- seq(0, 10, 0.02)
plot(RMexp(), model=model)
RFcov(model, 1)

[Package RandomFields version 3.3.6 Index]