RMnonstwm {RandomFields}	R Documentation

Non-stationary Whittle-Matern Covariance Model

Description

The non-stationary Whittle-Matern model C is given by

C(x, y)=Γ(μ) Γ(ν(x))^{-1/2} Γ(ν(y))^{-1/2} W_{μ} (|x-y|)

where μ = [ν(x) + ν(y)]/2, and ν must be a positive function.

W_{μ} is the covariance of the RMwhittle model or the RMmatern model.

Details

The non-stationary Whittle-Matern models are obtained by the respective stationary model, replacing the real-valued argument for nu by a non-negative function.

Note

It cannot be checked whether nu only takes positive values. So the responsibility is completely left to the user.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Stein, M. (2005) Nonstationary Spatial Covariance Functions. Tech. Rep., 2005

See Also

RMwhittle, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
x <- seq(-1.2, 1.2, len=50)
model <- RMwhittle(nu=RMgauss())

z <- RFsimulate(model=model, x, x, n=4)
plot(z)

[Package RandomFields version 3.3.6 Index]