RMnsst {RandomFields}	R Documentation

Non-Separable Space-Time model

Description

RMnsst is a univariate stationary spaceisotropic covariance model whose corresponding covariance is given by

C(h,u)= (ψ(u)+1)^{-δ/2} φ(h /√(ψ(u) +1))

Usage

RMnsst(phi, psi, delta, var, scale, Aniso, proj)

Arguments

phi
is a normal mixture RMmodel, cf.
RFgetModelNames(monotone="normal mixture")

psi
is a variogram RMmodel.

delta
a numerical value; must be greater than or equal to the spatial dimension of the field.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

This model is used for space-time modelling where the spatial component is isotropic.

Value

RMnsst returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T. (1997) Normal scale mixtures and dual probability densitites, J. Stat. Comput. Simul. 59, 375-384.

Gneiting, T. (2002) Nonseparable, stationary covariance functions for space-time data, JASA 97, 590-600.

Gneiting, T. and Schlather, M. (2001) Space-time covariance models. In El-Shaarawi, A.H. and Piegorsch, W.W.: The Encyclopedia of Environmetrics. Chichester: Wiley.

Schlather, M. (2010) On some covariance models based on normal scale mixtures. Bernoulli, 16, 780-797.

See Also

RMgennsst, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMnsst(phi=RMgauss(), psi=RMfbm(alpha=1), delta=2)
x <- seq(0, 10, 0.25)
plot(model, dim=2)
plot(RFsimulate(model, x=x, y=x))

[Package RandomFields version 3.3.6 Index]