RMnugget {RandomFields}	R Documentation

Nugget Effect Covariance Model

Description

RMnugget is a multivariate stationary isotropic covariance model called “nugget effect”. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given for i,j in 1,...,vdim by

C_{ij}(r) = δ_{ij} 1_{0}(r),

where δ_{ij}=1 if i=j and δ_{ij}=0 otherwise.

Usage

RMnugget(tol, vdim, var, Aniso, proj)

Arguments

tol
Only for advanced users. See RPnugget.

vdim
Must be set only for multivariate models (advanced).

var
optional argument; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Aniso,proj
(zonal modelling and repeated measurements(advanced)); see RPnugget for details.

Details

The nugget effect belongs to Gaussian white noise and is used for modeling measurement errors or to model spatial ‘nuggets’.

Value

RMnugget returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmodel, RFsimulate, RFfit, RPnugget (advanced users).

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

x <- y <- 1:2
xy <- as.matrix(expand.grid(x, y)) ## we get 4 locations

## Standard use of the nugget effect
model <- RMnugget(var = 100)
RFcovmatrix(model, x=xy)
as.vector(RFsimulate(model, x=x, y=x, tol=1e-10))

[Package RandomFields version 3.3.6 Index]