RMflatpower {RandomFields}	R Documentation

Variogram Model Similar to Fractal Brownian Motion

Description

RMflatpower is an intrinsically stationary isotropic variogram model. The corresponding centered semi-variogram only depends on the distance r ≥ 0 between two points and is given by

γ(r) = r^2 / ( 1 + r^2)^α

where 0 < α ≤ 1.

For related models see RMgenfbm.

Usage

RMflatpower(alpha, var, scale, Aniso, proj)

Arguments

alpha
numeric in (0,1]; refers to the fractal dimension of the process

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above variogram remains unmodified.

Details

The model is always smooth at the origin.

The parameter α only gives the tail behaviour and satisfies 0 < α ≤ 1.

The variogram is unbounded and belongs to a non-stationary process with stationary increments.

Value

RMflatpower returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Oesting, M., Schlather, M., and Friederichs, P. (2014) Conditional Modelling of Extreme Wind Gusts by Bivariate Brown-Resnick Processes arxiv 1312.4584.

See Also

RMgenfbm, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMflatpower(alpha=0.5)
x <- seq(0, 10, 0.1)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]