RMparswm {RandomFields}	R Documentation

Parsimonious Multivariate Whittle Matern Model

Description

RMparswm is a multivariate stationary isotropic covariance model whose corresponding covariance function only depends on the distance r ≥ 0 between two points and is given for i,j = 1,2 by

C_{ij}(r)=c_{ij} W_{ν_{ij}}(r).

Here W_ν is the covariance of the RMwhittle model.

RMparswmX ist defined as

ρ_{ij} C_{ij}(r)

where ρ_{ij} is any covariance matrix.

Usage

RMparswm(nudiag, var, scale, Aniso, proj)
RMparswmX(nudiag, rho, var, scale, Aniso, proj)

Arguments

nudiag
a vector of arbitrary length of positive values; the vector (ν_{11},ν_{22},...). The offdiagonal elements ν_{ij} are calculated as 0.5 (ν_{ii} + ν_{jj}).

rho
any positive definite m x m matrix; here, m equals length(nudiag). For the calculation of c_{ij} see Details.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

In the equation above we have

c_{ij} = ρ_{ ij} √ G_{ij}

and

G_{ij} = Γ(ν_{11} + d/2) Γ(ν_{22} + d/2) Γ(ν_{12}) / (Γ(ν_{11}) Γ(ν_{22}) Γ(ν_{12}+d/2))^2)

where Γ is the Gamma function and d is the dimension of the space.

Note that the definition of RMparswmX is RMschur(M=rho, RMparswm(nudiag, var, scale, Aniso, proj)).

Value

RMparswm returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T., Kleiber, W., Schlather, M. (2010) Matern covariance functions for multivariate random fields JASA

See Also

RMbiwm, RMwhittle, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

rho <- matrix(nc=3, c(1, 0.5, 0.2, 0.5, 1, 0.6, 0.2, 0.6, 1))
model <- RMparswmX(nudiag=c(1.3, 0.7, 2), rho=rho)
plot(model)
x.seq <- y.seq <- seq(-10, 10, 0.1)
z <- RFsimulate(model = model, x=x.seq, y=y.seq)
plot(z)

[Package RandomFields version 3.3.6 Index]