RMpenta {RandomFields}	R Documentation

Penta Covariance Model

Description

RMpenta is a stationary isotropic covariance model, which is only valid for dimensions d ≤ 3. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(r)=(1 - 22/3 r^{2} + 33 r^{4} - 77/2 r^{5} + 33/2 r^{7} - 11/2 r^{9} + 5/6 r^{11}) 1_{[0,1]}(r).

Usage

RMpenta(var, scale, Aniso, proj)

Arguments

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The model is only valid for dimensions d ≤ 3.

It has a 4 times differentiable covariance function with compact support (cf. Chiles, J.-P. and Delfiner, P. (1999), p. 84).

Value

RMpenta returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Chiles, J.-P. and Delfiner, P. (1999) Geostatistics. Modeling Spatial Uncertainty. New York: Wiley.

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMpenta()
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]