RMplus {RandomFields}	R Documentation

Addition of Random Field Models

Description

RMplus is an additive covariance model which depends on up to 10 submodels C_0, C_1, ..., C_9. In general, realizations of the created RMmodel are pointwise sums of independent realizations of the submodels.

In particular, if all submodels are given through a covariance function, the resulting model is defined through its covariance function, which is the sum of the submodels' covariances. Analogously, if all submodels are given through a variogram.

Usage

RMplus(C0, C1, C2, C3, C4, C5, C6, C7, C8, C9, var, scale, Aniso, proj)

Arguments

C0
an RMmodel.

C1,C2,C3,C4,C5,C6,C7,C8,C9
optional; each an RMmodel.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above model remains unmodified.

Details

RMmodels can also be summed up via the +-operator, e.g. C0 + C1.

The global arguments var,scale,Aniso,proj of RMplus are multiplied to the corresponding arguments of the submodels (from the right side).

Value

RMplus returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmult, RMmodel, RMsum, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMplus(RMgauss(), RMnugget(var=0.1))
model2<- RMgauss() + RMnugget(var=0.1)
plot(model, "model.+"=model2, type=c("p", "l"), pch=20, xlim=c(0,3)) # the same

[Package RandomFields version 3.3.6 Index]