RMpolygon {RandomFields}	R Documentation

RMpolygon

Description

RMpolygon refers to the indicator function of a typical Poisson polygon, used for instance in the (mixed) Storm process.

Usage

RMpolygon(lambda)

Arguments

lambda
intensity of the hyperplane process creating the random shape function.

The default value is 1.

Author(s)

Felix Ballani, http://www.mathe.tu-freiberg.de/sto/mitarbeiter/felix-ballani

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Poisson polygons / Poisson hyperplane tessellation

Lantuejoul, C. (2002) Geostatistical Simulation: Models and Algorithms. Springer.

Poisson storm process

Lantuejoul, C., Bacro, J.N., Bel L. (2011) Storm processes and stochastic geometry. Extremes, 14(4), 413-428.

Mixed Poisson storm process

Strokorb, K., Ballani, F., and Schlather, M. (2014) Tail correlation functions of max-stable processes: Construction principles, recovery and diversity of some mixing max-stable processes with identical TCF. Extremes, Submitted.

See Also

RMball, RMspheric, RFsimulate, RMmodel.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

[Package RandomFields version 3.3.6 Index]