RMpolynome {RandomFields}	R Documentation

Creating polynomial models

Description

Polynomials, mainly used in trend models, can be created easily with this function.

Usage

RMpolynome(degree, dim, value=NA, coordnames = c("x", "y", "z", "T"),
           proj=1:4)

Arguments

degree
degree of the polynome

dim
number of variables in the polynome

value
values of the coefficients. See Details.

coordnames
the names of the variables

proj
the projection to certain dimensions

Details

If the length of value is smaller than the number of monomials, the remaining terms are filled with NAs. If the length is larger, the vector is cut.

Value

RMpolynome returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMtrend, RFfit.

Examples

 ## For examples see the help page of 'RFformula' ##

RMpolynome(1, 1)
RMpolynome(1, 2)
RMpolynome(2, 1)
RMpolynome(2, 2)
RMpolynome(3, 3)

[Package RandomFields version 3.3.6 Index]