RMpower {RandomFields}	R Documentation

Power operator for Variograms and Covariance functions

Description

RMpower yields a variogram or covariance model from a given variogram or covariance model. The variogram gamma of the model is given by

gamma=phi^alpha

if phi is a variogram model. The covariance C of the model is given by

C(h) = phi(0)-(phi(0)-phi(h))^alpha

if phi is a covariance model.

Usage

RMpower(phi, alpha, var, scale, Aniso, proj)

Arguments

phi
a valid RMmodel; either a variogram model or a covariance model

alpha
a numerical value in the interval [0,1]

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

If γ is a variogram, then γ^α is a valid variogram for α in the interval [0,1].

Value

RMpower returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Schlather, M. (2012) Construction of covariance functions and unconditional simulation of random fields. In Porcu, E., Montero, J. M., Schlather, M. Advances and Challenges in Space-time Modelling of Natural Events, Springer, New York.

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMpower(RMgauss(), alpha=0.5)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]