RMqam {RandomFields}	R Documentation

Quasi-arithmetic mean

Description

RMqam is a univariate stationary covariance model depending on a submodel phi such that psi( . ) := phi(sqrt( . )) is completely monotone, and depending on further stationary covariance models C_i. The covariance is given by

C(h) = phi(sqrt(sum_i theta_i (phi^{-1}(C_i(h)))^2))

Usage

RMqam(phi, C1, C2, C3, C4, C5, C6, C7, C8, C9, theta, var, scale, Aniso, proj)

Arguments

phi
a valid covariance RMmodel that is a normal scale mixture. See, for instance,
RFgetModelNames(monotone="normal mixture").

C1, C2, C3, C4, C5, C6, C7, C8, C9
optional further univariate stationary RMmodels

theta
a vector with positive entries

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

Note that psi( . ) := phi(sqrt( . )) is completely monotone if and only if phi is a valid covariance function for all dimensions, e.g. RMstable, RMgauss, RMexponential.

Warning: RandomFields cannot check whether the combination of phi and C_i is valid.

Value

RMqam returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Porcu, E., Mateu, J. & Christakos, G. (2007) Quasi-arithmetic means of covariance functions with potential applications to space-time data. Submitted to Journal of Multivariate Analysis.

See Also

RMmqam, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMqam(phi=RMgauss(), RMexp(), RMgauss(),
               theta=c(0.3, 0.7), scale=0.5)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]