RMscale {RandomFields}	R Documentation

Scale model for arbitrary areas of scales

Description

Let s_x the scaling at location x and p a bijective penalizing function for (different) scales. Then covariance function is given by

C(x,y) = φ(\|x-y\| + |p(s_x) - p(s_y)|)

Usage

RMscale(phi, scaling, penalty, var, scale, Aniso, proj)

Arguments

phi
isotropic submodel

scaling
model that gives the non-stationary scaling s_x

penalty
bijective function p applied to the scaling

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Value

RMscale returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Bonat, W.H. , Ribeiro, P. Jr. and Schlather, M. (2019) Modelling non-stationarity in scale. In preparation.

See Also

RMSadvanced, RMblend, RMbubble

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

x <- seq(0,1, 0.01)
scale <- RMcovariate(x=c(0,1), y=c(1,0),#2 areas separated by the 1st bisector
                     grid=FALSE, data=c(1, 3))

model <- RMscale(RMexp(), scaling = scale, penalty=RMid() / 2)
plot(z <- RFsimulate(model, x, x))

[Package RandomFields version 3.3.6 Index]