RMschlather {RandomFields}	R Documentation

Covariance Model for binary field based on Gaussian field

Description

RMschlather gives the tail correlation function of the extremal Gaussian process, i.e.

C(h) = 1 - √{ (1-φ(h)/φ(0)) / 2 }

where φ is the covariance of a stationary Gaussian field.

Usage

RMschlather(phi, var, scale, Aniso, proj)

Arguments

phi
covariance function of class RMmodel.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

This model yields the tail correlation function of the field that is returned by RPschlather.

Value

RMschlather returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RPschlather, RMmodel, RFsimulate.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## This example considers an extremal Gaussian random field
## with Gneiting's correlation function.

## first consider the covariance model and its corresponding tail
## correlation function
model <- RMgneiting()
plot(model, model.tail.corr.fct=RMschlather(model),  xlim=c(0, 5))

## the extremal Gaussian field with the above underlying
## correlation function that has the above tail correlation function
x <- seq(0, 10, 0.1)
z <- RFsimulate(RPschlather(model), x)
plot(z)

## Note that in RFsimulate R-P-schlather was called, not R-M-schlather.
## The following lines give a Gaussian random field with correlation
## function equal to the above tail correlation function.
z <- RFsimulate(RMschlather(model), x)
plot(z)

[Package RandomFields version 3.3.6 Index]