RMschur {RandomFields}	R Documentation

Schur product

Description

The covariance function is

C(x)= M * φ(x)

where ‘*’ denotes the Schur product, i.e. elementwise multiplication.

Usage

RMschur(phi, M,  diag, rhored, var, scale, Aniso, proj)

Arguments

phi
covariance function of class RMmodel

M
constant n x n covariance matrix of the same size as multivariate model phi

diag,rhored
alternative way of passing M: diag is a vector of variances, rhored is a vector containing the correlations of the lower triangle of the M.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Value

RMschur returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

?

See Also

RMmodel, RMmatrix.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
model <- RMschur(M=matrix(c(2, 1, 1, 1), ncol=2), RMparswm(nudiag=c(0.5, 2)))
plot(model)
x <- seq(0, 10, 0.02)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]