RMsign {RandomFields}	R Documentation

Random sign

Description

RMsign defines a random sign. It can be used as part of the model definition of a Poisson field.

Usage

RMsign(phi, p)

Arguments

phi
shape function of class RMmodel

p
probability of keeping the sign

Details

RMsign changes the sign of the shape function phi with probability 1-p and keeps it otherwise.

Value

RMsign returns an object of class RMmodel.

Note

Random univariate or multivariate objects usually start with RR, not with RM. This is an exception here, as it operates on shape functions.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmodel, RR.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RPpoisson(RMsign(RMtent(), p=0.8))
x <- seq(0, 10, 0.02)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]