RMstein {RandomFields}	R Documentation

Stein's non-separable space-time model

Description

RMstein is a univariate stationary covariance model whose corresponding covariance function only depends on the difference h between two points and is given by

C(h, t) = W_ν(y) - < h, z > t W_{ν-1}(y) / [ (ν-1) (2ν + d + 1) ]

Here, W_ν is the covariance of the RMwhittle model with smoothness parameter ν; y = ||(h,t)|| is the norm of the vector (h,t), d is the dimension of the space on which the random field is considered.

Usage

RMstein(nu, z, var, scale, Aniso, proj)

Arguments

nu
numerical value; greater than 1; smoothness parameter of the RMwhittle model

z
a vector; the norm of z must be less or equal to 1.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

See Stein (2005).

Value

RMstein returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Stein, M.L. (2005) Space-time covariance functions. J. Amer. Statist. Assoc. 100, 310-321. Equation (8).

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMstein(nu=1.5, z=0.9)
x <- seq(0, 10, 0.05)
plot(RFsimulate(model, x=x, y=x))

[Package RandomFields version 3.3.6 Index]