RMstp {RandomFields}	R Documentation

Single temporal process

Description

RMstp is a univariate covariance model which depends on a normal mixture submodel phi. The covariance is given by

C(x,y) = |S_x|^{1/4} |S_y|^{1/4} |A|^{-1/2} φ(Q(x,y)^{1/2})

where

Q(x,y) = c^2 - m^2 + h^t (S_x + 2(m + c)M) A^{-1} (S_y + 2 (m-c)M)h,

c = -z^t h + ξ_2(x) - ξ_2(y),

A = S_x + S_y + 4 M h h^t M

m = h^t M h

h = x - y

Usage

RMstp(xi, phi, S, z, M, var, scale, Aniso, proj) 

Arguments

xi
arbitrary univariate function on R^d

phi
an RMmodel that is a normal mixture model, cf.
RFgetModelNames(monotone="normal mixture")

S
functions that returns strictly positive definite d x d

z
arbitrary vector, z \in R^d

M
an arbitrary, symmetric d x d matrix

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

See Schlather (2008) formula (13). The model allows for mimicking cyclonic behaviour.

Value

RMstp returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Paciorek C.J., and Schervish, M.J. (2006) Spatial modelling using a new class of nonstationary covariance functions, Environmetrics 17, 483-506.

Schlather, M. (2010) Some covariance models based on normal scale mixtures. Bernoulli, 16, 780-797.

See Also

RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMstp(xi = RMrotat(phi= -2 * pi, speed=1),
               phi = RMwhittle(nu = 1),
               M=matrix(nc=3, rep(0, 9)),
               S=RMetaxxa(E=rep(1, 3), alpha = -2 * pi,
                          A=t(matrix(nc=3, c(2, 0, 0, 1, 1 , 0, 0, 0, 0))))
              )
x <- seq(0, 10, 0.7)
plot(RFsimulate(model, x=x, y=x, z=x))

[Package RandomFields version 3.3.6 Index]