Strokorb's Functions {RandomFields}	R Documentation

Tail correlation function of the Brown-Resnick process

Description

The models define various shape functions for max-stable processes for a given tail correlation function.

Usage

RMm2r(phi)
RMm3b(phi)
RMmps(phi)

Arguments

phi
a model for a tail correlation function belonging to the Gneiting class H_d

Details

RMm2r used with RPsmith defines a monotone shape function that corresponds to a tail correlation function belonging to Gneiting's class H_d. Currently, the function is implemented for dimensions 1 and 3. Called as such it returns the corresponding monotone function.

RMm3b used with RPsmith defines balls with random radius that corresponds to a tail correlation function belonging to Gneiting's class H_d. Currently, the function is implemented for dimensions 1 and 3. (Note that in Strokorb et al. (2014) the density function for twice the radius is considered.) Called as such it returns the corresponding density function for the radius of the balls.

RMmps used with RPsmith defines random hyperplane polygons that correspond to a tail correlaton function belonging to Gneiting's class H_d. It currently only allows for RMbrownresnick(RMfbm(alpha=1)) and dimension 2. Called as such it returns the tcf defined by the submodel – this definition may change in future.

Value

object of class RMmodel

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Strokorb, K. (2013) Properties of the Extremal Coefficient Functions. Univ. Goettingen. PhD thesis.

Strokorb, K., Ballani, F. and Schlather, M. (2014) In Preparation.

See Also

RFsimulate, RMmodel.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMbrownresnick(RMfbm(alpha=1.5, s=0.2))
plot(RMm2r(model))

x <- seq(0, 10, 0.005)
z <- RFsimulate(RPsmith(RMm2r(model), xi=0), x)
plot(z, type="p", pch=20)

[Package RandomFields version 3.3.6 Index]