RMsum {RandomFields}	R Documentation

Plain scalar product

Description

RMsum is given by

C(x,y) = φ(x) + φ(y).

It is a negative definite function although not a variogram.

Usage

RMsum(phi, var, scale, Aniso, proj)

Arguments

phi
any function of class RMmodel

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Value

RMsum returns an object of class RMmodel.

Note

Do not mix up this model with RMplus.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmodel, RMplus, RMprod.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

[Package RandomFields version 3.3.6 Index]