RMtbm {RandomFields}	R Documentation

Turning Bands Method

Description

RMtbm is a univariate or multivaraiate stationary isotropic covariance model in dimension reduceddim which depends on a univariate or multivariate stationary isotropic covariance phi in a bigger dimension fulldim. For formulas for the covariance function see details.

Usage

RMtbm(phi, fulldim, reduceddim, layers, var, scale, Aniso, proj)

Arguments

phi, fulldim, reduceddim, layers
See RPtbm.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The turning bands method stems from the 1:1 correspondence between the isotropic covariance functions of different dimensions. See Gneiting (1999) and Strokorb and Schlather (2014).

The standard case is reduceddim=1 and fulldim=3. If only one of the arguments is given, then the difference of the two arguments equals 2.

For d == n + 2, where n=reduceddim and d==fulldim the original dimension, we have

C(r) = phi(r) + r phi'(r) / n

which for n=1 reduces to the standard TBM operator

C(r) = d/dr [ r phi(r) ]

For d == 2 && n == 1 we have

C(r) = d/dr int_0^r [ r phi(r) ] / [ sqrt{r^2 - u^2} ] d u

‘Turning layers’ is a generalization of the turning bands method, see Schlather (2011).

Value

RMtbm returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Turning bands

Gneiting, T. (1999) On the derivatives of radial positive definite function. J. Math. Anal. Appl, 236, 86-99

Matheron, G. (1973). The intrinsic random functions and their applications. Adv . Appl. Probab., 5, 439-468.

Strokorb, K., Ballani, F., and Schlather, M. (2014) Tail correlation functions of max-stable processes: Construction principles, recovery and diversity of some mixing max-stable processes with identical TCF. Extremes, Submitted.

Turning layers

Schlather, M. (2011) Construction of covariance functions and unconditional simulation of random fields. In Porcu, E., Montero, J.M. and Schlather, M., Space-Time Processes and Challenges Related to Environmental Problems. New York: Springer.

See Also

RPtbm, RFsimulate.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

x <- seq(0, 10, 0.02)
model <- RMspheric()
plot(model, model.on.the.line=RMtbm(RMspheric()), xlim=c(-1.5, 1.5))

z <- RFsimulate(RPtbm(model), x, x)
plot(z)

[Package RandomFields version 3.3.6 Index]