RMtruncsupport {RandomFields}	R Documentation

Truncating the Support of a Shape Function

Description

RMtruncsupport may be used to truncate the support of a shape function when Poisson fields or M3 processes are created.

Usage

RMtruncsupport(phi, radius)

Arguments

phi
function of class RMmodel

radius
truncation at radius

Value

RMtruncsupport returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Schlather, M. (2002) Models for stationary max-stable random fields. Extremes 5, 33-44.

See Also

RMmodel, RMmatrix, RPpoisson.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again 

model <- RMgauss()
model1 <- RMtruncsupport(model, radius=1)
plot(model)
lines(model1, col="red")

## For a real application of 'RMtruncsupport' see example 2 of 'RPpoisson'.

[Package RandomFields version 3.3.6 Index]