RMvector {RandomFields}	R Documentation

Vector Covariance Model

Description

RMvector is a multivariate covariance model which depends on a univariate covariance model that is stationary in the first Dspace coordinates h and where the covariance function phi(h,t) is twice differentiable in the first component h.

The corresponding matrix-valued covariance function C of the model only depends on the difference h between two points in the first component. It is given by

C(h,t)=( -0.5 * (a + 1) Δ + a \nabla \nabla^T ) C_0(h, t)

where the operator is applied to the first component h only.

Usage

RMvector(phi, a, Dspace, var, scale, Aniso, proj)

Arguments

phi
an RMmodel; has two components h (2- or 3-dimensional and stationary) and t (arbitrary dimension).

a
a numerical value; should be in the interval [-1,1].

Dspace
an integer; either 2 or 3; the first Dspace coordinates give the first component h.

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

C_0 is either a spatio-temporal model (then t is the time component) or it is an isotropic model. Then, the first Dspace coordinates are considered as h coordinates and the remaining ones as t coordinates. By default, Dspace equals the dimension of the field (and there is no t component). If a=-1 then the field is curl free; if a=1 then the field is divergence free.

Value

RMvector returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Scheuerer, M. and Schlather, M. (2012) Covariance Models for Divergence-Free and Curl-Free Random Vector Fields. Stochastic Models 28:3.

See Also

RMcurlfree, RMdivfree, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMvector(RMgauss(), scale=0.3)
x <- seq(0, 10, 0.4)
plot(RFsimulate(model, x=x, y=x, z=0), select.variables=list(1:2))

[Package RandomFields version 3.3.6 Index]