RMwave {RandomFields}	R Documentation

Wave Covariance Model / Cardinal Sine

Description

RMwave is a stationary isotropic covariance model, which is valid only for dimensions d ≤ 3. The corresponding covariance function only depends on the distance r ≥ 0 between two points and is given by

C(r)=sin(r)/r 1_{r>0} + 1_{r=0} .

It is a special case of RMbessel.

Usage

RMwave(var, scale, Aniso, proj)
RMcardinalsine(var, scale, Aniso, proj)

Arguments

var,scale,Aniso,proj
optional arguments; same meaning for any RMmodel. If not passed, the above covariance function remains unmodified.

Details

The model is only valid for dimensions d ≤ 3. It is a special case of RMbessel for ν = 0.5.

This covariance models a hole effect (cf. Chiles, J.-P. and Delfiner, P. (1999), p. 92).

Value

RMwave returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Chiles, J.-P. and Delfiner, P. (1999) Geostatistics. Modeling Spatial Uncertainty. New York: Wiley.

See Also

RMbessel, RMmodel, RFsimulate, RFfit.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMwave(scale=0.1)
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x))

[Package RandomFields version 3.3.6 Index]