RPprocess {RandomFields}	R Documentation

Models for classes of random fields (RP commands)

Description

Here, all classes of random fields are described that can be simulated.

Implemented processes

Gaussian Random Fields 	 see Gaussian

Max-stable Random Fields 	 see Maxstable

Other Random Fields 	 Binary field

chi2 field

composed Poisson (shot noise, random coin)

t field

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RC, RR, RM, RF, R.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
x <- seq(0, 10, 0.1)
model <- RMexp()

## a Gaussian field with exponential covariance function
z <- RFsimulate(model, x)
plot(z)

## a binary field obtained as a thresholded Gaussian field
b <- RFsimulate(RPbernoulli(model), x)
plot(b)

sum( abs((z@data$variabl1 >=0 ) - b@data$variable1)) == 0 ## TRUE,
## i.e. RPbernoulli is indeed a thresholded Gaussian process

[Package RandomFields version 3.3.6 Index]