BrownResnick {RandomFields}	R Documentation

Brown-Resnick process

Description

RPbrownresnick defines a Brown-Resnick process.

Usage

  RPbrownresnick(phi, tcf, xi, mu, s)

Arguments

phi
specifies the covariance model or variogram, see RMmodel and RMmodelsAdvanced.

tcf
the extremal correlation function; either phi or tcf must be given.

xi, mu, s
the extreme value index, the location parameter and the scale parameter, respectively, of the generalized extreme value distribution. See Details.

Details

The argument xi is always a number, i.e. ξ is constant in space. In contrast, μ and s might be constant numerical values or (in future!) be given by an RMmodel, in particular by an RMtrend model.
For xi=0, the default values of mu and s are 0 and 1, respectively. For xi\not=0, the default values of mu and s are 1 and |ξ|, respectively, so that it defaults to the standard Frechet case if ξ > 0.

The functions RPbrorig, RPbrshifted and RPbrmixed perform the simulation of a Brown-Resnick process, which is defined by

Z(x) = max_{i=1, 2, ...} X_i * exp(W_i(x) - gamma^2),

where the X_i are the points of a Poisson point process on the positive real half-axis with intensity 1/x^2 dx, W_i ~ Y are iid centered Gaussian processes with stationary increments and variogram gamma given by phi.

For simulation, internally, one of the methods RPbrorig, RPbrshifted and RPbrmixed is chosen automatically.

Note

Advanced options are maxpoints and max_gauss, see RFoptions.

Further advanced options related to the simulation methods RPbrorig, RPbrshifted and RPbrmixed can be found in the paragraph ‘Specific method options for Brown-Resnick Fields’ in RFoptions.

Author(s)

Marco Oesting, oesting@mathematik.uni-siegen.de, https://www.uni-siegen.de/fb6/src/scheffler/mitarbeiter/oesting; Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Brown, B.M. and Resnick, S.I. (1977). Extreme values of independent stochastic processes. J. Appl. Probab. 14, 732-739.

Buishand, T., de Haan , L. and Zhou, C. (2008). On spatial extremes: With application to a rainfall problem. Ann. Appl. Stat. 2, 624-642.

Kabluchko, Z., Schlather, M. and de Haan, L (2009) Stationary max-stable random fields associated to negative definite functions Ann. Probab. 37, 2042-2065.

Oesting, M., Kabluchko, Z. and Schlather M. (2012) Simulation of Brown-Resnick Processes, Extremes, 15, 89-107.

See Also

RPbrorig, RPbrshifted, RPbrmixed, RMmodel, RPgauss, maxstable, maxstableAdvanced.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## for some more sophisticated models see 'maxstableAdvanced'

[Package RandomFields version 3.3.6 Index]