Coins {RandomFields}	R Documentation

Random coin method

Description

The random coin method (or dilution method) is a simulation method for stationary Gaussian random fields. It is based on the following procedure: For a stationary Poisson point process on R^d consider the random field

Y(y) = ∑_{x\in X} f(y-x)

for a function f. The covariance of Y is proportional to the convolution

C(h) = \int f(x)f(x+h) dx

If the intensity of the Poisson point process increases, the random field Y approaches a Gaussian random field with covariance function C.

Usage

RPcoins(phi, shape, boxcox, intensity, method) 

RPaverage(phi, shape, boxcox, intensity, method) 

Arguments

phi
object of class RMmodel; specifies the covariance function of the Poisson process; either phi or shape must be given.

shape
object of class RMmodel; specifies the function which is attached to the Poisson points; note that this is not the covariance function of the simulated random field.

boxcox
the one or two parameters of the box cox transformation. If not given, the globally defined parameters are used. See RFboxcox for details.

intensity
positive number, intensity of the underlying Poisson point process.

method
integer. Default is the value 0 which addresses the current standard procedure. There might be further methods implemented mainly for internal purposes.

Value

RPcoins returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Lantuejoul, C. (2002) Geostatistical Simulation: Models and Algorithms. Springer.

See Also

Gaussian, RP, RPhyperplane, RPspectral, RPtbm.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

[Package RandomFields version 3.3.6 Index]