RPgauss {RandomFields}	R Documentation

Simulation of Gaussian Random Fields

Description

This function is used to specify a Gaussian random field that is to be simulated or estimated. Returns an object of class RMmodel.

Usage

RPgauss(phi, boxcox, stationary_only)

Arguments

phi
the RMmodel.

boxcox
the one or two parameters of the box cox transformation. If not given, the globally defined parameters are used. See RFboxcox for details.

stationary_only
Logical or NA. Used for the automatic choice of methods.

TRUE: The simulation of non-stationary random fields is refused. In particular, the intrinsic embedding method is excluded and the simulation of Brownian motion is rejected.

FALSE: Intrinsic embedding is always allowed; actually, it's the first one considered in the automatic selection algorithm.

NA: The simulation of the Brownian motion is allowed, but intrinsic embedding is not used for translation invariant (“stationary”) covariance models.

Default: NA.

Value

The function returns an object of class RMmodel.

Note

In most cases, RPgauss need not be given explicitly as Gaussian random fields are assumed as default.

RPgauss may not find the fastest method neither the most precise one. It just finds any method among the available methods. (However, it guesses what is a good choice.) See RFgetMethodNames for further information. Note that some of the methods do not work for all covariance or variogram models, see RFgetModelNames(intern=FALSE).

By default, all Gaussian random fields have zero mean. Simulating with trend can be done by including RMtrend in the model.

RPgauss allows to simulate different classes of random fields, controlled by the wrapping model:

If the submodel is a pure covariance or variogram model, i.e. of class RMmodel, a corresponding centered Gaussian field is simulated. Not only stationary fields but also non-stationary and anisotropic models can be used, e.g. zonal anisotropy, geometrical anisotropy, separable models, non-separable space-time models, multiplicative or nested models; see RMmodel for a list of all available models.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RP, Gaussian, RMmodel, RFoptions, RPbrownresnick, RPchi2, RPopitz, RPt, RPschlather.

Do not mix up with RMgauss or RRgauss.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMexp()
x <- seq(0, 10, 0.02)
plot(model)
plot(RFsimulate(model, x=x, seed=0))
plot(RFsimulate(RPgauss(model), x=x, seed=0), col=2) ## the same

[Package RandomFields version 3.3.6 Index]