Hyperplane {RandomFields}	R Documentation

Hyperplane method

Description

The Hyperplane method is a simulation method for stationary, isotropic random fields with exponential covariance function. It is based on a tessellation of the space by hyperplanes. Each cell takes a spatially constant value of an i.i.d. random variable. The superposition of several such random fields yields approximatively a Gaussian random field.

Usage

RPhyperplane(phi, boxcox, superpos, maxlines, mar_distr, mar_param ,additive)

Arguments

phi
object of class RMmodel; specifies the covariance function to be simulated; only exponential covariance functions and scale mixtures of it are allowed.

boxcox
the one or two parameters of the box cox transformation. If not given, the globally defined parameters are used. See RFboxcox for details.

superpos
integer. number of superposed hyperplane tessellations. Default: 300.

maxlines
integer. Maximum number of allowed lines.

Default: 1000.

mar_distr
integer. code for the marginal distribution used in the simulation:

0

uniform distribution

1

Frechet distribution with form parameter mar_param

2

Bernoulli distribution (Binomial with n=1) with argument mar_param

This argument should not be changed yet.

Default: 0.

mar_param
Argument used for the marginal distribution. The argument should not be changed yet.

Default: NA.

additive
logical. If TRUE independent realizations are added, else the maximum is taken.

Default: TRUE.

Value

RPhyperplane returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Lantuejoul, C. (2002) Geostatistical Simulation: Models and Algorithms. Springer.

See Also

Gaussian, RP.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
model <- RPhyperplane(RMexp(s=2), superpos=1)
x <- seq(0, 3, 0.04)
z <- RFsimulate(x=x, x, model=model, n=1)
plot(z)

[Package RandomFields version 3.3.6 Index]