Max-stable random fields {RandomFields}	R Documentation

Simulation of Max-Stable Random Fields

Description

Here, a list of models and methods for simulating max-stable random fields is given.

See also maxstableAdvanced for more advanced examples.

Implemented models and methods

Models

RPbrownresnick 	 Brown-Resnick process using an automatic choice of the 3 RPbr* methods below

RPopitz 	 extremal t process

RPschlather 	 extremal Gaussian process

RPsmith 	 M3 processes

Methods

RPbrmixed 	 simulation of Brown-Resnick processes using M3 representation

RPbrorig 	 simulation of Brown-Resnick processes using the original definition

RPbrshifted 	 simulation of Brown-Resnick processes using a random shift

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Kabluchko, Z., Schlather, M. & de Haan, L (2009) Stationary max-stable random fields associated to negative definite functions Ann. Probab. 37, 2042-2065.

Schlather, M. (2002) Models for stationary max-stable random fields. Extremes 5, 33-44.

Smith, R.L. (1990) Max-stable processes and spatial extremes Unpublished Manuscript.

See Also

RP, RMmodel, RPgauss, RPbernoulli, maxstableAdvanced.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

### currently not programmed

## Not run: \dontshow{
## to do : seq(0, 10, 0.02) oben ist furchtbar langsam. Warum?
}
## End(Not run)

## Not run: \dontshow{
model <- RMball()
x <- seq(0, 10, 5) # nice for   x <- seq(0, 10, 0.02)
z <- RFsimulate(RPsmith(model, xi=0), x, n=1000, every=1000)
plot(z)
hist(unlist(z@data), 150, freq=FALSE) #not correct; to do; sqrt(2) wrong
curve(exp(-x) * exp(-exp(-x)), from=-3, to=8, add=TRUE, col=3)
}
## End(Not run)

model <- RMgauss()
x <- seq(0, 10, 0.05)
z <- RFsimulate(RPschlather(model, xi=0), x, n=1000)
plot(z)
hist(unlist(z@data), 50, freq=FALSE)
curve(exp(-x) * exp(-exp(-x)), from=-3, to=8, add=TRUE)

## for some more sophisticated models see maxstableAdvanced

[Package RandomFields version 3.3.6 Index]