Extremal t {RandomFields}	R Documentation

Extremal t process

Description

RPopitz defines an extremal t process.

Usage

RPopitz(phi, xi, mu, s, alpha)

Arguments

phi
an RMmodel; covariance model for a standardized Gaussian random field, or the field itself.

xi,mu,s
the extreme value index, the location parameter and the scale parameter, respectively, of the generalized extreme value distribution. See Details.

alpha
originally referred to the α-Frechet marginal distribution, see the original literature for details.

Details

The argument xi is always a number, i.e. ξ is constant in space. In contrast, μ and s might be constant numerical values or (in future!) be given by an RMmodel, in particular by an RMtrend model.
For xi=0, the default values of mu and s are 0 and 1, respectively. For xi\not=0, the default values of mu and s are 1 and |ξ|, respectively, so that it defaults to the standard Frechet case if ξ > 0.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Davison, A.C., Padoan, S., Ribatet, M. (2012). Statistical modelling of spatial extremes. Stat. Science 27, 161-186.

Opitz, T. (2012) A spectral construction of the extremal t process. arxiv 1207.2296.

See Also

RMmodel, RPgauss, maxstable, maxstableAdvanced.

Examples

## sorry, does not work savely yet

[Package RandomFields version 3.3.6 Index]