Sequential {RandomFields}	R Documentation

Methods relying on square roots of the covariance matrix

Description

Sequential method relying on square roots of the covariance matrix

Usage

RPsequential(phi, boxcox, back_steps, initial)

Arguments

phi
object of class RMmodel; specifies the covariance model to be simulated.

boxcox
the one or two parameters of the box cox transformation. If not given, the globally defined parameters are used. See RFboxcox for details.

back_steps
Number of previous instances on which the algorithm should condition. If less than one then the number of previous instances equals max / (number of spatial points).

Default: 10.

initial
First, N=(number of spatial points) * back_steps number of points are simulated. Then, sequentially, all spatial points for the next time instance are simulated at once, based on the previous back_steps instances. The distribution of the first N points is the correct distribution, but differs, in general, from the distribution of the sequentially simulated variables. We prefer here to have the same distribution all over (although only approximatively the correct one), hence do some initial sequential steps first. If initial is non-negative, then initial first steps are performed. If initial is negative, then back_steps - initial initial steps are performed. The latter ensures that none of the very first N variables are returned.

Default: -10.

Details

RPsequential is programmed for spatio-temporal models where the field is modelled sequentially in the time direction conditioned on the previous k instances. For k=5 the method has its limits for about 1000 spatial points. It is an approximative method. The larger k the better. It also works for certain grids where the last dimension should contain the highest number of grid points.

Value

RPsequential returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Schlather, M. (1999) An introduction to positive definite functions and to unconditional simulation of random fields. Technical report ST 99-10, Dept. of Maths and Statistics, Lancaster University.

See Also

Gaussian, RP, RPdirect.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
model <- RMgauss(var=10, s=10) + RMnugget(var=0.01)
plot(model, xlim=c(-25, 25))

z <- RFsimulate(model=RPsequential(model), 0:10, 0:10, n=4)
plot(z)

[Package RandomFields version 3.3.6 Index]