Smith {RandomFields}	R Documentation

(Mixed) Moving Maxima

Description

RPsmith defines a moving maximum process or a mixed moving maximum process with finite number of shape functions.

Usage

RPsmith(shape, tcf, xi, mu, s)

Arguments

shape
an RMmodel giving the spectral function

tcf
an RMmodel specifying the extremal correlation function; either shape or tcf must be given. If tcf is given a shape function is tried to be constructed via the RMm2r construction of deterministic, monotone functions.

xi,mu,s
the extreme value index, the location parameter and the scale parameter, respectively, of the generalized extreme value distribution. See Details.

Details

The argument xi is always a number, i.e. ξ is constant in space. In contrast, μ and s might be constant numerical values or (in future!) be given by an RMmodel, in particular by an RMtrend model.
For xi=0, the default values of mu and s are 0 and 1, respectively. For xi\not=0, the default values of mu and s are 1 and |ξ|, respectively, so that it defaults to the standard Frechet case if ξ > 0.

It simulates max-stable processes Z that are referred to as “Smith model”.

Z(x) = max_{i=1, 2, ...} X_i * Y_i(x - W_i),

where (W_i, X_i) are the points of a Poisson point process on R^d x (0, ∞) with intensity dw * c/x^2 dx and Y_i ~ Y are iid measurable random functions with E[int max(0, Y(x)) dx ] < ∞. The constant c is chosen such that Z has standard Frechet margins.

Note

IMPORTANT: For consistency reasons with the geostatistical definitions in this package the scale argument differs froms the original definition of the Smith model! See the example below.

RPsmith depends on RRrectangular and its arguments.

Advanced options are maxpoints and max_gauss, see RFoptions.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Haan, L. (1984) A spectral representation for max-stable processes. Ann. Probab., 12, 1194-1204.

Smith, R.L. (1990) Max-stable processes and spatial extremes Unpublished Manuscript.

See Also

Advanced RMmodels, Auxiliary RMmodels, RMmodel, RPbernoulli, RPgauss, maxstable, maxstableAdvanced.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RMball()
x <- seq(0, 1000, 0.2)
z <- RFsimulate(RPsmith(model, xi=0), x)
plot(z)
hist(z@data$variable1, 50, freq=FALSE)
curve(exp(-x) * exp(-exp(-x)), from=-3, to=8, add=TRUE)

## 2-dim
x <- seq(0, 10, 0.1) 
z <- RFsimulate(RPsmith(model, xi=0), x, x)
plot(z)

## original Smith model
x <- seq(0, 10, 0.05)
model <- RMgauss(scale = sqrt(2)) # !! cf. definition of RMgauss
z <- RFsimulate(RPsmith(model, xi=0), x, x)
plot(z)

## for some more sophisticated models see 'maxstableAdvanced'

[Package RandomFields version 3.3.6 Index]