Specific {RandomFields}	R Documentation

Methods that are specific to certain covariance models

Description

This model determines that the (Gaussian) random field should be modelled by a particular method that is specific to the given covariance model.

Usage

RPspecific(phi, boxcox)

Arguments

phi
object of class RMmodel; specifies the covariance model to be simulated.

boxcox
the one or two parameters of the box cox transformation. If not given, the globally defined parameters are used. See RFboxcox for details.

Details

RPspecific is used for specific algorithms or specific features for simulating certain covariance functions.

RMplus is able to simulate separately the fields given by its summands. This is necessary, e.g., when a trend model RMtrend is involved.

RMmult for Gaussian random fields only. RMmult simulates the random fields of all the components and multiplies them. This is repeated several times and averaged.

RMS Then, for instance, sqrt(var) is multiplied onto the (Gaussian) random field after the field has been simulated. Hence, when var is random, then for each realization of the Gaussian field (for n>1 in RFsimulate) a new realization of var is used.

Further, new coordinates are created where the old coordinates have been divided by the scale and/or multiplied with the Aniso matrix or a projection has been performed.

RPspecific(RMS()) is called internally when the user wants to simulate Anisotropic fields with isotropic methods, e.g. RPtbm.

RMmppplus

RMtrend

Note that RPspecific applies only to the first model or operator in the argument phi.

Value

RPspecific returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Schlather, M. (1999) An introduction to positive definite functions and to unconditional simulation of random fields. Technical report ST 99-10, Dept. of Maths and Statistics, Lancaster University.

See Also

Gaussian, RP.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## example for implicit use
model <- RMgauss(var=10, s=10) + RMnugget(var=0.1)
plot(model)
plot(RFsimulate(model=model, 0:10, 0:10, n=4))
## The following function shows the internal structure of the model.
## In particular, it can be seen that RPspecific is applied to RMplus.
RFgetModelInfo(level=0, which="internal")

## example for explicit use: every simulation has a different variance
model <- RPspecific(RMS(var=unif(min=0, max=100), RMgauss()))
x <- seq(0,50,0.02)
plot(RFsimulate(model, x=x, n=4), ylim=c(-15,15))

[Package RandomFields version 3.3.6 Index]