RPt {RandomFields}	R Documentation

Simulation of T Random Fields

Description

RPt defines a t field.

Usage

RPt(phi, boxcox, nu)

Arguments

phi
the RMmodel. If a model for the distribution is not specified, RPgauss is used as default and a covariance model is expected.

boxcox
the one or two parameters of the box cox transformation. If not given, the globally defined parameters are used. See RFboxcox for details.

nu
non-negative number. Degree of freedom.

Value

The function returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Related to the extremal t process

T. Opitz (2012) A spectral construction of the extremal t process. arxiv 1207.2296.

See Also

Auxiliary RMmodels, RP, RPgauss.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

model <- RPt(RMexp(), nu=2)
x <- seq(0, 10, 0.1)
z <- RFsimulate(model, x, x, n=4)
plot(z)

[Package RandomFields version 3.3.6 Index]