RRspheric {RandomFields}	R Documentation

Random scaling used with balls

Description

This model delivers the distribution of the radius of a ball obtained by the intersection of a balldim-dimensional ball with diameter R by a spacedim-dimensional hyperplane that has uniform distance from the center.

Usage

RRspheric(spacedim, balldim, R) 

Arguments

spacedim
dimension of the hyperplane; defaults to 1.

balldim
the dimension of the ball

R
radius. Default: 1.

Value

RRspheric returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmodel, RMball.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

hist(RFrdistr(RRspheric(balldim=2), n=1000), 50)

[Package RandomFields version 3.3.6 Index]