RRunif {RandomFields}	R Documentation

Uniform Distribution in Higher Dimensions

Description

The model refers to the d-dimensional uniform distribution on a rectangular window.

Usage

RRunif(min, max, normed) 

Arguments

min,max
lower and upper corner of a rectangular window

normed
logical with default value TRUE.

Advanced. If FALSE then the indicator function for the window is not normed to get a probability distribution. Nonetheless, random drawing from the distribution still works.

Details

In the one-dimensional case it has the same effect as RRdistr(unif(min=min, max=max, log=log)).

Value

RRunif returns an object of class RMmodel.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RMmodel, RRdistr, RRgauss, RRspheric.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
## uniform distribution on [0,2] x [-2, -1]
RFrdistr(RRunif(c(0, -2), c(2, -1)), n=5, dim=2)
RFpdistr(RRunif(c(0, -2), c(2, -1)), q=c(1, -1.5), dim=2)
RFddistr(RRunif(c(0, -2), c(2, -1)), x=c(1, -1.5), dim=2)

[Package RandomFields version 3.3.6 Index]