RandomFields {RandomFields}	R Documentation

Simulation and Analysis of Random Fields

Description

The package RandomFields allows for simulating various kinds of random fields, including anisotropic processes. Furthermore, algorithms for conditional simulation and simulation of max-stable random fields are provided.

Additionally, the package includes tools for analysing spatial data: Hurst parameter, fractal dimension, empirical variogram, interactive fitting of parameters, LSQ and MLE estimation of parameters. Basic kriging procedures are also provided.

Starting with version 2.0, it also allows for the simulation of random fields that are non-stationary or multivariate or sophisticated space-time fields. fitvario allows for multivariate models and mixed effect models.

There are some changings in the definitions and in the output, see help("changings")

Details

The following random fields and related functionalities are provided by the package.

stationary and isotropic Gaussian random fields

CondSimu : conditional simulation

CovarianceFct, sophisticated models: covariance functions and variogram models

EmpiricalVariogram : empirical variogram

GaussRF : simulation of Gaussian random fields; nice examples to get familiar with the simulation features of the package;

Kriging : simple and ordinary kriging

fitvario : variogram/covariance function fit by least squares, maximum likelihood and cross validation techniques

PrintMethodList : list of implemented simulation methods

ShowModels : interactive, graphical choice of models – currently not available.

Use ShowModels in version 1.3.x

soil : Soil physical and chemical data; the example gives a simple geostatistical analysis using features of the package

stationary (and isotropic) max-stable random fields

CovarianceFct : covariance models for extremal Gaussian random fields

MaxStableRF : simulation of max-stable random fields

Note: Simulation algorithm for Brown-Resnick processes will come up soon.

stationary and isotropic Poisson random fields (not implemented yet)

Data and example code:

soil : soil physical data

papers : code used in the papers published by the author

weather : UWME weather data

Functions used in diverse simulation methods:

DeleteRegister : deleting internal registers

RFparameters : control parameters (advanced settings)

Functions of general purpose:

fractal.dim : estimation of the fractal dimension

hurst : estimation of the Hurst parameter

Print : nice print function

regression : interactive regression plot

Locator, Readline : locator and readline, respectively, with storage and replay functionality

sleep.milli : sets the process into sleeping status

In the beta version, the following functionalities are currently not available:

ShowModels

numerical evaluation of the covariance function in tbm2

Harvard Rue's Markov fields

Acknowledgement

Many thanks to

R Core Team making available the algoithm for fft (fft.c) by Richard Singleton and advicing

Ben Pfaff, 12167 Airport Rd, DeWitt MI 48820, USA making available an algorithm for AVL trees (avltr*)

Peter Menck implemented the multivariate circulant embedding for version 2.0.

Yindeng Jiang jiangyindeng@gmail.com implemented the circulant embedding methods ‘cutoff’ and ‘intrinsic’ in 2004 for the versions 1.2.

Martin Maechler, Paulo Ribeiro, and Tilmann Gneiting were proof-reading parts of the code and the help text for the versions 1.0.

Financial support

V1.0 has been financially supported by the German Federal Ministry of Research and Technology (BMFT) grant PT BEO 51-0339476C during 2000-03.

V1.0 has been financially supported by the EU TMR network ERB-FMRX-CT96-0095 on “Computational and statistical methods for the analysis of spatial data” in 1999.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de;

References

Singleton, R.C. (1979). In Programs for Digital Signal Processing Ed.: Digital Signal Processing Committee and IEEE Acoustics, Speech, and Signal Processing Committe (1979) IEEE press.

[Package RandomFields version 2.0.71 Index]