Readline {RandomFields}	R Documentation

Read a Line

Description

Readline reads a line from the terminal or from storage

Usage

   Readline(prompt="", info=NULL)

Arguments

prompt
the string printed when prompting the user for input. Should usually end with a space " "

info
arbitrary object; tracing information that is useful in case the user likes to edit the stored sequence of inputs; the value of info is considered only if useraction has been called by action="start.register" or action="continue.register"

Details

The behaviour of Readline depends on the the value of action set by useraction, see there for more information.

The prompt string will be truncated to a maximum allowed length, normally 256 chars (but can be changed in the source code), since the function is based on readline.

Value

A character vector of length one.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de

See Also

getactions, Locator, readline useraction, userinput,

Examples

## see useraction

[Package RandomFields version 2.0.66 Index]