DeleteRegister {RandomFields}	R Documentation

Deleting Intermediate Results

Description

DeleteRegister deletes an internal register of the random field simulation

DeleteAllRegisters deletes all internal registers

Usage

DeleteRegister(nr=0)
DeleteAllRegisters()

Arguments

nr
number of the register to be deleted; in 0:9

Details

DeleteRegister and DeleteAllRegister should be used if

the simulations are finished and there is not enough memory available to continue the R session, and RFparameters()$Storing has been TRUE.

one of the parameters RFparameters()$Storing or RFparameters()$PracticalRange ((or RFparameters()$TBM.method, once this parameter is available)) has been changed by the user, and if these options should become global.

See GaussRF and RFparameters for further details.

Value

NULL.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

GaussRF, RandomFields and RFparameters.

Examples

 DeleteAllRegisters() ## no effect visible for the user

[Package RandomFields version 2.0.71 Index]