SimulateRF {RandomFields}	R Documentation

Simulation of Random Fields

Description

DoSimulateRF performs an already initialised simulation.

InitSimulateRF internal function; use InitGaussRF and InitMaxStableRF, instead.

Usage

DoSimulateRF(n=1, register=0, paired=FALSE, trend=NULL)

InitSimulateRF(x, y=NULL, z=NULL, T=NULL, grid=!missing(gridtriple),
               model, param, trend, method=NULL, register=0, gridtriple,
               distribution=NA)

Arguments

x
matrix of coordinates, or vector of x coordinates

y
vector of y coordinates

z
vector of z coordinates

T
time instances

grid
logical; determines whether the vectors x, y, and z should be interpreted as a grid definition, see Details.

model
string; covariance or variogram model, see CovarianceFct, or type PrintModelList() to get all options

param
vector or list. param=c(mean, variance, nugget, scale, ...), param=list(c(variance, scale, ...), ..., c(variance,scale,...)), param=matrix(...), or param=list(list(variance, anisotropy, kappa),..., list(variance, anisotropy, kappa)); the parameters must be given in this order; further parameters are to be added in case of a parametrised class of models, see CovarianceFct

method
NULL or string; Method used for simulating, see RFMethods, or type PrintMethodList() to get all options

register
0:9; place where intermediate calculations are stored; the numbers are aliases for 10 internal registers

gridtriple
logical; if gridtriple=FALSE ascending sequences for the parameters x, y, and z are expected; if gridtriple=TRUE triples of form c(start,end,step) expected; this parameter is used only if grid=TRUE

distribution
marginal distribution:
'Gauss', 'Poisson', or 'MaxStable'

n
number of realisations to generate; if paired=TRUE then n must be even.

paired
logical. paired may be TRUE only for the simulation of Gaussian random fields. If TRUE then every second simulation is obtained by only changing the signs of the standard Gaussian random variables, the simulation is based on (“antithetic pairs”).

trend
only used for universal and intrinsic kriging. In case of universal kriging trend is a non-negative integer (monomials up to order k as trend functions), a list of functions or a formula (the summands are the trend functions); you have the choice of using either x, y, z or X1, X2, X3,... as spatial variables; in case of intrinsic kriging trend should be a nonnegative integer which is the order of the underlying model.

Value

InitSimulateRF returns 0 if no error has occurred during the initialisation process, and a positive value if failed.

DoSimulateRF returns NULL if an error has occurred; otherwise the returned object depends on the parameters n and grid:
n=1:
* grid=FALSE. A vector of simulated values is returned (independent of the dimension of the random field)
* grid=TRUE. An array of the dimension of the random field is returned.

n>1:
* grid=FALSE. A matrix is returned. The columns contain the realisations.
* grid=TRUE. An array of dimension d+1, where d is the dimension of the random field, is returned. The last dimension contains the realisations.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de

See Also

GaussRF, MaxStableRF, RandomFields

[Package RandomFields version 2.0.66 Index]