CheckXT {RandomFields}	R Documentation

Internal functions – do not use them directly

Description

CheckXT checks whether the coordinates of the data and related parameters are specified correctly and transforms the coordinates into a standard format

PrepareModel checks whether the parameters of the covariance model and related parameters are specified correctly and transforms the parameters into a standard format

plotWithCircles displays data values of marked point processes by circles

GetDistributionNames returns the names of the currently available marginal distributions of the random fields

GetrfParameters returns some internal constants

Usage

CheckXT(x, y, z, T, grid, gridtriple)
PrepareModel(model, param, trend, method=NULL, nugget.remove=TRUE)
plotWithCircles(data, factor=1.0, xlim=range(data[,1])+c(-maxr,maxr),
                ylim=range(data[,2])+c(-maxr,maxr),col=1, fill=0, ...)
GetDistributionNames()
GetrfParameters(initcov)

Arguments

x
x coordinates

y
y coordinates

z
z coordinates

T
time instances

grid
logical; simulation on a grid; see GaussRF

gridtriple
logical; grid is given in abbreviated form; see GaussRF

model
name of the covariance model; see GaussRF

param
parameter for the covariance model; see GaussRF

trend
mean or trend of the random field

method
simulation method

nugget.remove
if nugget is 0, the nugget part be removed in the returned model?

data
matrix of 3 columns; first two columns give the coordinates, the third the data

factor
enlargement factor for data

xlim
see plot

ylim
see plot

col
border colour of circles

fill
filling colour of circles

...
further graphical parameters in plotWithCircles and any variables to be printed in Print

initcov
boolean, whether to call the C function InitModelList

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de

See Also

CovarianceFct

Examples

model <- list(model="whittle", param=c(1,2,3,4,5))
Print(PrepareModel(model=model, me="ci"))

model <- list("+",
              list("$", var=2, s=4, list(model="whi", kappa=5)),
              list("$", var=3, s=0, list(model="whi", kappa=1)) )
Print(PrepareModel(model=model, me="ci"))

[Package RandomFields version 2.0.66 Index]