Changings {RandomFields}	R Documentation

Changings

Description

Here, only the important inconsistencies between Version 1.3 to 2.0 of RandomFields are listed.

Documentation of the extensions will be enhanced later.

Details

output of fitvario
$vario is defunctioned. Use, for instance, $ml$model instead of $vario$ml.

The options for mle.methods in fitvario are currently restricted to ‘ml’. This options ‘reml’ and ‘rml1’ have not been tested yet.

the list definition of sophisticated covariance models has become obsolete. Furthermore, the parameter kappa for the model parameters is not accepted anymore. Instead k should be used.

See help("sophisticated") for the new definition of sophisticated covariance models

Currently, ShowModels does not work, neither Havard Rue's Markov fields. Both should be available in summer 2011 again.

Currently, the numerical evaluation of the covariance function in tbm2 is not implemented. This should be available in summer 2011 again.

Please let the author know if further inconsistencies exist.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de;

[Package RandomFields version 2.0.71 Index]