conventional2RFspDataFrame {RandomFields}	R Documentation

Coercion to class 'RFsp' objects

Description

Generate an object of class RFsp from conventional objects.

Usage

conventional2RFspDataFrame(data, coords=NULL, gridTopology=NULL, n=1,
                           vdim=1, T=NULL, vdim_close_together)

Arguments

data
array; of dimension c(vdim, space-time-dim, n); contains the values of the random field

coords
matrix of coordinates

gridTopology
3-row-matrix or of class GridTopology; specifies the grid vectors; either coords or gridTopology must be NULL

n
number of iid copies of the random field, default is 1

vdim
number of dimensions of the values of the random field, default is 1

T
time component if any. The length of the temporal grid is needed by as.array if the spatial locations are randomly scattered.

vdim_close_together
logical. Currently, only vdim_close_together=FALSE is coded. In this case the dimensions of the data follow the order “locations, multivariate, repeated”. Otherwise “multivariate, locations, repeated”.

Value

Object of class RFspatialGridDataFrame, RFspatialPointsDataFrame, RFgridDataFrame or RFpointsDataFrame.

Author(s)

Alexander Malinowski, Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
x <- 1:20
z <- RFsimulate(RMexp(), x, spConform=FALSE)
z2 <- conventional2RFspDataFrame(z, coord=x)
Print(z, z2)

[Package RandomFields version 3.3.6 Index]