fitgauss {RandomFields}	R Documentation

Details on fitting Gaussian random fields, including Box-Cox transformation

Description

Here, some details of RFfit are given concerning the fitting of models for Gaussian random fields.

This documentation is far from being complete.

Maximum likelihood

The application of the usual maximum likelihood method and reporting the result is the default.

Least squares

The weighted least squares methods minimize

∑_{i} w_i (\hat γ(h_i) - γ(h_i))^2

over all parametrized models of γ. Here, i runs over all N bins of the binned variogram \hat γ and h_i is the centre of bin i.

The following variants of the least squares methods, passed as sub.methods in RFfit are implemented:

'self'

w_i = (γ(h_i))^{-2}

'plain'

w_i = 1 for all i.

'sqrt.nr'

w_i^2 equals the number of points n_i in bin i.

'sd.inv'

1 / w_i equals the standard deviation of the variogram cloud within bin i.

'internal'

Three subvariants are implemented:

'internal1'

w_i^2 = (N-i+1) n_i

'internal2'

w_i = N-i+1

'internal3'

w_i^2 = N-i+1

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RFfit, RFfit-class.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again
## See 'RFfit'.

[Package RandomFields version 3.3.6 Index]