getactions {RandomFields}	R Documentation

Get input behaviour

Description

The functions return values stored by useraction, Locator and Readline

Usage

  getactions()
  getactionlist()

Value

getactionlist returns a list of stored values created by Locator and Readline

getactions returns a list of all parameters that influence the behaviour of Locator and Readline.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de

See Also

Locator, Readline, useraction,

Examples

## see useraction

[Package RandomFields version 2.0.71 Index]