host {RandomFields}	R Documentation

System calls

Description

The functions hostname and pid return the host name and the PID, respectively.

Usage

hostname()

pid()

Details

If R runs on a unix platform the host name and the PID are returned, otherwise the empty string and naught, respectively.

Value

hostname
returns a string

pid
returns an unsigned integer

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de

[Package RandomFields version 2.0.66 Index]