Obsolete Functions Version 2 {RandomFields}	R Documentation

Obsolete functions Version 2

Description

This part gives the obsolete functions of RandomFields Version 2 that are not maintained anymore.

Usage

Covariance(x, y = NULL, model, param = NULL, dim = NULL,
 Distances, fctcall = c("Cov", "Variogram", "CovMatrix"))
CovarianceFct(x, y = NULL, model, param = NULL, dim =NULL,
 Distances, fctcall = c("Cov", "Variogram", "CovMatrix"))
CovMatrix(x, y = NULL, model, param = NULL, dim = NULL, Distances)
DeleteAllRegisters()
DeleteRegister(nr=0)
DoSimulateRF(n = 1, register = 0, paired=FALSE, trend=NULL) 
InitSimulateRF(x, y = NULL, z = NULL, T=NULL, grid=!missing(gridtriple),
 model, param, trend, method = NULL, register = 0,
 gridtriple, distribution=NA)
InitGaussRF(x, y = NULL, z = NULL, T=NULL, grid=!missing(gridtriple),
 model, param, trend=NULL, method = NULL, register = 0, gridtriple) 
GaussRF(x, y = NULL, z = NULL, T=NULL, grid=!missing(gridtriple), model,
 param, trend=NULL, method = NULL, n = 1, register = 0, gridtriple,
 paired=FALSE, PrintLevel=1, Storing=TRUE, ...) 
Variogram(x, model, param = NULL, dim = NULL, Distances)
InitMaxStableRF(x, y = NULL, z = NULL, grid=NULL, model, param, maxstable,
 method = NULL, register = 0, gridtriple = FALSE)
MaxStableRF(x, y = NULL, z = NULL, grid=NULL, model, param, maxstable,
 method = NULL, n = 1, register = 0, gridtriple = FALSE, ...)
EmpiricalVariogram(x, y = NULL, z = NULL, T=NULL, data, grid=NULL, bin,
 gridtriple = FALSE, phi, theta, deltaT)
Kriging(krige.method, x, y=NULL, z=NULL, T=NULL, grid=NULL, gridtriple=FALSE,
 model, param, given, data, trend=NULL,pch=".", return.variance=FALSE,
 allowdistanceZero = FALSE, cholesky=FALSE) 
CondSimu(krige.method, x, y=NULL, z=NULL, T=NULL, grid=NULL, gridtriple=FALSE,
 model, param, method=NULL, given, data, trend=NULL, n=1, register=0, 
 err.model=NULL, err.param=NULL, err.method=NULL, err.register=1, 
 tol=1E-5, pch=".", paired=FALSE, na.rm=FALSE) 
RFparameters(...)
hurst(x, y = NULL, z = NULL, data, gridtriple = FALSE, sort=TRUE,
 block.sequ = unique(round(exp(seq(log(min(3000, dim[1] / 5)),
 log(dim[1]), len=min(100, dim[1]))))),
 fft.m = c(1, min(1000, (fft.len - 1) / 10)),
 fft.max.length = Inf, 
 method=c("dfa", "fft", "var"), mode=c("plot", "interactive"),
 pch=16, cex=0.2, cex.main=0.85,
 PrintLevel=RFoptions()$basic$printlevel,height=3.5, ...)
fractal.dim(x, y = NULL, z = NULL, data, grid=TRUE, gridtriple = FALSE,
 bin, vario.n=5, sort=TRUE, fft.m = c(65, 86), fft.max.length=Inf,
 fft.max.regr=150000, fft.shift = 50, method=c("variogram", "fft"),
 mode=c("plot", "interactive"), pch=16, cex=0.2, cex.main=0.85,
 PrintLevel = RFoptions()$basic$printlevel, height=3.5, ...)
fitvario(x, y=NULL, z=NULL, T=NULL, data, model, param, lower=NULL,
 upper=NULL, sill=NA, grid=!missing(gridtriple), gridtriple=FALSE, ...)

Arguments

x, y, model, param, dim, Distances, fctcall, n, register, paired, trend, z, T, grid, method, gridtriple, distribution, PrintLevel, Storing, ..., maxstable, data, bin, phi, theta, deltaT, krige.method, pch, return.variance, allowdistanceZero, cholesky, given, err.model, err.param, err.method, err.register, tol, na.rm, sort, block.sequ, fft.m, fft.max.length, mode, cex, cex.main, height, vario.n, fft.max.regr, fft.shift, lower, upper, sill, nr
As the functions are obsolete, all these arguments are not documented anymore.

Value

See ‘version2’ for details on these obsolete commands.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

The functions that should be used instead are, for instance, RFcov, RFcovmatrix, RFvariogram, RFsimulate, RFinterpolate, RFvariogram, RFfit, RFoptions, RFhurst, RFfractaldim

See ‘version2’ for details on the obsolete commands.

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

## no examples given, as command is obsolete

[Package RandomFields version 3.3.6 Index]