parameter.range {RandomFields}	R Documentation

Range of model specific parameters

Description

parameter.range returns the range of the parameters specific to a parametrised variogram model or a covariance function

— NOT PROGRAMMED YET –

Usage

parameter.range(model, param, dim=1)

Arguments

model
the variogram model or covariance function, see CovarianceFct

param
see CovarianceFct

dim
the dimension of the random field

Value

If the model is not valid or not specified for the given dimension, NaN is returned.

If the model does not have additional parameters, NULL is returned. Otherwise a list is returned.

If the parameter space is simple, a list of the following elements is returned

theoretical
The theoretical ranges of the parameters. Note that open and closed intervals are not distinguished here.

practical
The ranges of the parameters usually not exceeded in practical applications.

Both elements are matrices with two rows and the number of columns being the number of parameters.

If the parameter space is complex, the parameter space is devided into rectangular areas and theoretical and practical are lists where each element is a matrix as above.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de

See Also

CovarianceFct

Examples

  parameter.range("exponential", param=NA, dim=1) # NULL
  try(parameter.range("power", param=NA, dim=1))
  try(parameter.range("power", param=c(NA, NA, NA, NA, 2), dim=2))
  Print(parameter.range("gengneiting", param=NA, dim=2))
  Print(parameter.range("nsst", param=NA,  dim=1))
  Print(parameter.range("hyper",param, dim=1))

[Package RandomFields version 2.0.66 Index]