regression {RandomFields}	R Documentation

Regression plot

Description

Interactive regression plot

Usage

regression(x, y, main, scr, value=function(regr) regr$coeff[2], 
           variable="var", mode=c("nographics", "plot", "interactive"),
           regr.select=c("coefficients"), averaging=FALSE, cex.main=0.85,
           col.passive = "grey", col.active = "green",
           col.main.passive = "black", col.main.active = "red",
           col.points.chosen = "lightblue", col.abline = "yellow",
           col.abline.chosen = "darkblue", col.smooth = "red", ...)

Arguments

x
x-coordinates of the regression plot

y
y-coordinates

main
main title

scr
split.screen must have been called beforehand; scr gives the screen number

value
function; the characteristic of interest is a function of the regression output.

variable
name of the characteristic of interest, plotted in the graphs

mode
A scalar character of one of the following values 'silent', 'plot', or 'interactive':

'silent':

no output except the return of the result

'plot':

the regression line is plotted

'interactive':

the regression domain can be chosen interactively

Any mode includes all the functionalities of the modes that precede in the above enumeration. In the interactive mode, the regression domain is chosen by two mouse clicks with the left mouse; a right mouse click leaves the plot.

regr.select
vector of characters that select elements from the lm return list; these elements are returned by regression.

averaging
logical. If TRUE and y is matrix then the values of y are averaged row-wise beforehand

cex.main
font size of the title

col.passive
colour of the data points if plot is not active

col.active
colour of the data points if plot is active

col.main.passive
colour of the title if plot is not active

col.main.active
colour of the title if plot is active

col.points.chosen
colour of the selected points

col.abline
colour of the regression line for all points

col.abline.chosen
colour of the regression line for the selected points

col.smooth
colour for the smoothed data

...
further graphical parameters

Value

The function returns a list with the following components

regr
part of the return list of lm

val
the calculated value of the variable of interest

regr.u
NULL or part of the return list of lm for x.u and y.u

val.u
NULL or the calculated value of the variable of interest for the restricted domain

x.u
NULL or the restricted x-coordinates given by the user in the interactive plot

y.u
NULL or y-coordinates according to x.u

sm
smoothed curve through the (x,y) points

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de http://ms.math.uni-mannheim.de

[Package RandomFields version 2.0.66 Index]