PrintModelList {RandomFields}	R Documentation

Information about the implemented covariance models

Description

PrintModelList prints the list of currently implemented models including the corresponding simulation methods.

Usage

PrintModelList(operators=FALSE, internal=FALSE, newstyle=TRUE)

Arguments

operators
logical. Flag whether operators should also be considered.

internal
logical. Flag whether internal models should also be considered. In case of PrintModelList and internal=2, variants of internal models are also printed.

newstyle
logical. If FALSE then only the old style model names (Version 2 and earlier) are shown. These names can still be used in the list definition of models, see RMmodelsAdvanced. If TRUE then the standard names will also be shown.

Details

See RMmodel for a description of the models and their use.

Value

PrintModelList prints a table of the currently implemented covariance functions and the matching methods. PrintModelList returns NULL.

Note

From version 3.0 on, the command PrintModelList() is replaced by the call RFgetModelNames(internal=FALSE).

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

See Also

RFgetModelNames

Examples

RFoptions(seed=0) ## *ANY* simulation will have the random seed 0; set
##                   RFoptions(seed=NA) to make them all random again

PrintModelList()

[Package RandomFields version 3.3.6 Index]