papers {RandomFields}	R Documentation

Papers involving RandomFields and co-authored by M. Schlather

Description

Here, an overview is given over the papers co-authored by M. Schlather that involve RandomFields.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

References

Gneiting, T., Kleiber, W. and Schlather, M. (2010) Matern cross-covariance functions for multivariate random fields J. Amer. Statist. Assoc. 105, 1167-1177.

See GKS11 for the code.

Gneiting, T., Sevcikova, H., Percival, D.B., Schlather, M., Jiang, Y. (2006) Fast and Exact Simulation of Large Gaussian Lattice Systems in R2: Exploring the Limits. J. Comput. Graph. Stat., 15, 483-501.

See GSPSJ06 for the code.

Scheuerer, M. and Schlather, M. (2012) Covariance Models for Random Vector Fields. Stochastic Models, 82, 433-451.

See SS12 for the code.

Schlather, M. (2002) Models for stationary max-stable random fields. Extremes 5, 33-44.

See S02 for the code.

Schlather, M. (2010) On some covariance models based on normal scale mixtures. Bernoulli, 16, 780-797.

See S10 for the code.

Schlather, M., Malinowski, A., Menck, P.J., Oesting, M. and Strokorb, K. (2015) Analysis, simulation and prediction of multivariate random fields with package RandomFields. Journal of Statistical Software, 63 (8), 1-25, url = ‘http://www.jstatsoft.org/v63/i08/’,

See ‘multivariate_jss’ for the vignette.

Strokorb, K., Ballani, F. and Schlather, M. (2014) In Preparation.

See SBS14 for the code.

See Also

weather, GSPSJ06, SS12, S02, S10, jss14.

Examples

# For examples, see the specific papers.

[Package RandomFields version 3.3.6 Index]