weather {RandomFields}	R Documentation

Pressure and temperature forecast errors over the Pacific Northwest

Description

Meteorological dataset, which consists of differences between forecasts and observations (forecasts minus observations) of temperature and pressure at 157 locations in the North American Pacific Northwest.

Usage

data(weather)

Format

The data frame weather contains the following columns:

pressure

in units of Pascal

temperature

in units of degree Celcius

lon

longitudinal coordinates of the locations

lat

latitude coordinates of the locations

Furthermore, some results obtained from the data analysis in jss14 are delivered that are pars.model, pars, whole.model, whole.

Finally, the variable information contains packing information (the date and the version of RandomFields).

Details

The forecasts are from the GFS member of the University of Washington regional numerical weather prediction ensemble (UWME; Grimit and Mass 2002; Eckel and Mass 2005); they were valid on December 18, 2003 at 4 pm local time, at a forecast horizon of 48 hours.

Author(s)

Martin Schlather, schlather@math.uni-mannheim.de, http://ms.math.uni-mannheim.de

Source

The data were obtained from Cliff Mass and Jeff Baars from the University of Washington Department of Atmospheric Sciences.

References

Eckel, A. F. and Mass, C. F. (2005) Aspects of effective mesoscale, short-range ensemble forecasting Wea. Forecasting 20, 328-350.

Gneiting, T., Kleiber, W. and Schlather, M. (2010) Matern cross-covariance functions for multivariate random fields J. Amer. Statist. Assoc. 105, 1167-1177.

Grimit, E. P. and Mass, C. F. (2002) Initial results of a mesoscale short-range forecasting system over the Pacific Northwest Wea. Forecasting 17, 192-205.

See Also

A reanalysis has been performed in Section 5 of the jss14 paper.

Examples

## See 'jss14'.

[Package RandomFields version 3.3.6 Index]